==> script/01_Wrangling-Senegal-Census-data.Rmd <==
---
title: "Wrangling Senegal Census data with tidyverse tools"
author: "IBRAHIM KASSOUM Habibou & HEMA Aboubacar"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, error=FALSE)
```

# Preambule: 

```{r}
# Clear the current workspace by removing all objects.
# This is useful to start with a clean slate and avoid potential conflicts or unexpected behavior.
rm(list=ls())
```

## Importing library
```{r Package needed, results = "hide"}


## Importing library
### List of required packages
required_packages <- c("tidyverse","janitor" ,"readr","dplyr","haven","sf",
                       "flextable","sp", "factoextra", "FactoMineR","gtsummary", "sjPlot")

# Check if packages are installed
missing_packages <- setdiff(required_packages, installed.packages()[,"Package"])

### Install missing packages
if (length(missing_packages) > 0) {
  install.packages(missing_packages)
}

### Load all packages
lapply(required_packages, library, character.only = TRUE)

```


## Importing the datasets an checking for duplicated row

### Importing the shapefile data

```{r, results='hide'}
# Read shapefile data for 2002 and 2013
sp_rgph_2002 <- sf::read_sf(paste0(here::here(),"/data/EAs_2013_2002_SHP/EAs_2002.shp"))
# Mutating the 'sp_rgph_2002' dataset to create a new 'id_dr' column based on conditions
sp_rgph_2002 <- sp_rgph_2002 %>% 
  mutate(id_dr = ifelse(NUMDEPT=="3", 
                      paste0(NUM_REG, NUMDEPT, NUM_ARRDT, "18", NUM_CA, IDDR, "O"), 
                      paste0("0", NUM_REG, NUMDEPT, NUM_ARRDT, "8", NUM_CA, IDDR, "O")))
sp_rgph_2013 <- sf::read_sf(paste0(here::here(),"/data/EAs_2013_2002_SHP/EAs_2013.shp"))

# Display the first 10 rows of the 2002 and 2013 shapefiles
head(sp_rgph_2002, 10L)
head(sp_rgph_2013, 10L)
```

### Importing the census datasets 

**For the 2002 census data**

```{r, results='hide'}
# Read and check for duplicated rows in 2002 datasets

# Read and check for duplicated rows in the "deces_2002" dataset
deces_2002 <- read_sav(paste0(here::here(),"/data/1. RGPH 2002/Requete PDBA2k2 Deces.sav"))
deces_2002 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "deces_complement" dataset
deces_2002_complement <- read_sav(paste0(here::here(),"/data/1. RGPH 2002/PDBA_rgph2002_Deces Complement Pikine.sav"))
deces_2002_complement %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "individus_2002" dataset
individus_2002 <- haven::read_sav(paste0(here::here(),"/data/1. RGPH 2002/Requete PDBA2k2 Individus.v2.sav"))
individus_2002 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "menage_2002" dataset
menage_2002 <- haven::read_sav(paste0(here::here(),"/data/1. RGPH 2002/Requete PDBA2k2 menage.sav"))
menage_2002 %>% janitor::get_dupes()
menage_2002 %>% janitor::get_dupes(ID_MENAGE)

# Read and check for duplicated rows in the "menage_2002_complement" dataset
menage_2002_complement <- read_sav(paste0(here::here(),"/data/1. RGPH 2002/PDBA_rgph2002_menage_complement.sav"))
menage_2002_complement %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "individus_2002_complement_pikine" dataset
individus_2002_complement_pikine <- read_sav(paste0(here::here(),"/data/1. RGPH 2002/PDBA_rgph2002_Individus Complement Pikine.sav"))
individus_2002_complement_pikine %>% janitor::get_dupes() 
```



```{r}

# Joining the 2002 dataset with the "deces_2002_complement" for deceased individuals
deces_2002 <- deces_2002_complement %>% 
  # Creating unique identifiers for each record in the dataset
  mutate(ID_MENAGE = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE, QA08, QA09, QA10),
         DR = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE)) %>%
  # Combining the dataset with the existing "deces_2002" dataset, filling missing values
  plyr::rbind.fill(deces_2002)

# Joining with the 2002 dataset for individual data
individus_2002 <- individus_2002_complement_pikine %>% 
  # Creating unique identifiers for each record in the dataset
  mutate(ID_MENAGE = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE, QA08, QA09, QA10),
         DR = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE)) %>%
  # Combining the dataset with the existing "individus_2002" dataset
  plyr::rbind.fill(individus_2002)

# Joining with the 2002 dataset for household data
menage_2002 <- menage_2002_complement %>% 
  # Creating unique identifiers for each record in the dataset
  mutate(ID_MENAGE = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE, QA08, QA09, QA10),
         DR = paste0(QA01, QA02, QA03, QA04, QA05, QA06, "0", QA07, QA072_CLE)) %>%
  # Combining the dataset with the existing "menage_2002" dataset
  plyr::rbind.fill(menage_2002)
```



**For the 2013 census dataset**

```{r}

# Read and check for duplicated rows in 2013 datasets

# Read and check for duplicated rows in the "deces_2013" dataset
deces_2013 <- read_sav(paste0(here::here(),"/data/4. RGPH 2013/Requete PDBA2k13 Deces_Guediawaye.sav"))
deces_2013 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "individus_2013" dataset
individus_2013 <- haven::read_sav(paste0(here::here(),"/data/4. RGPH 2013/Requete PDBA2k13 Indidivus GUEDIAWAYE.sav"))
individus_2013 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "menage_2013" dataset
menage_2013 <- haven::read_sav(paste0(here::here(),"/data/4. RGPH 2013/Requete PDBA2k13 menage.sav"))
menage_2013 %>% janitor::get_dupes() 
menage_2013 %>% janitor::get_dupes(ID_MENAGE)
```



# Preparation of data on the two censuses (Wrangling)

```{r}
rm(list = c("deces_2002_complement","individus_2002_complement_pikine","menage_2002_complement","missing_packages","required_packages"))
```


```{r}
# Creating a variable "RGPH" to identify each survey 

# For 2002
## Add a new column "RGPH" with the value "2002" to the "individus_2002" dataset
individus_2002 <- 
  individus_2002 %>% 
  mutate(RGPH = "2002") %>% 
  mutate(id_dr = DR)

# For 2013
## Add a new column "RGPH" with the value "2013" to the "individus_2013" dataset
individus_2013 <- 
  individus_2013 %>% 
  mutate(RGPH = "2013") %>% 
  mutate(id_dr = IDDR)

## For RGPH 2013
individus_2002 <- individus_2002 %>%
 select(id_dr,ID_MENAGE, RGPH, everything())# 

## For RGPH 2002
individus_2013 <- individus_2013 %>%
 select(id_dr,ID_MENAGE, RGPH, everything())# 
```

# Study Variables

## Socio-demographic variables

### Sex/gender

```{r echo=TRUE}
#For the RGPH 2002 dataset
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
  group_by(ID_MENAGE) %>% 
  summarise(
    homme = sum(QB04 == 1, na.rm = TRUE),
    femme = sum(QB04 == 2, na.rm = TRUE),
    hh_size = homme + femme
  ) %>% 
  ungroup() %>% 
  mutate(
    homme = as.numeric(homme) %>%
           structure(label = "Nombre d'homme dans le ménage"),
    femme = as.numeric(femme) %>%
           structure(label = "Nombre de femmes dans le ménage"),
    hh_size = as.numeric(hh_size) %>%
           structure(label = "Taille du ménage")),
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE"))

#1 = Male, 2 = Female
#For the RGPH 2013 dataset

#1 = Male, 2 = Female
menage_2013 <- menage_2013 %>% 
  left_join( 
    individus_2013 %>% 
  group_by(ID_MENAGE) %>% 
  summarise(
    homme = sum(B06 == 1, na.rm = TRUE),
    femme = sum(B06 == 2, na.rm = TRUE),
    hh_size = homme + femme
  ) %>% 
  ungroup() %>% 
  mutate(
    homme = as.numeric(homme) %>%
           structure(label = "Nombre d'homme dans le ménage"),
    femme = as.numeric(femme) %>%
           structure(label = "Nombre de femmes dans le ménage"),
    hh_size = as.numeric(hh_size) %>%
           structure(label = "Taille du ménage")),
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE"))

gender_var = c("homme","femme","hh_size")
```


### Age

```{r echo=TRUE, eval=TRUE, include=TRUE}
#For babies under one year old, write “0” years. For people aged 98 and over, the instruction was to write 98 for age.



# For the RGPH 2002 dataset

## Look if it range for individual data is from 0 to 98

individus_2002 %>% 
  summarise(n_individus=n() %>%
           structure(label = "Nombre des individus"),
           # 273939
           n_nan = sum(is.na(QB06))  %>%
           structure(label = "Nombre de valeurs manquantes"),
           # 0
            min_age=min(QB06, na.rm = TRUE) %>%
           structure(label = "Minimun âge des individus"),
           # 0
            mean_age = mean(QB06, na.rm = TRUE) %>%
           structure(label = "Moyenne âge des individus"),
           # 23.26
            max_age=max(QB06, na.rm = TRUE) %>%
           structure(label = "Max âge des individus")
           # 98
            )
## Count the number of nbr_babies (aged 0), nbr_under_14 (number of children between 0 - 14), nbr_young_people (number of individuals aged between 15-29), nbr_adult (number of individuals aged between 30-44), nbr_elder_adult (number of individuals aged between 45-59), nbr_senior (number of individuals aged more than 60)

# Left join the 'menage_2002' dataframe with a summarized version of 'individus_2002'
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing various counts based on age (QB06 column)
      summarise(
        nbr_babies = sum(QB06 == 0, na.rm = TRUE),  # Count of babies aged 0
        nbr_under_14 = sum(QB06 <= 14, na.rm = TRUE),  # Count of individuals under 14
        nbr_young_people = sum(QB06 > 14 & QB06 <= 29, na.rm = TRUE),  # Count of individuals aged 15-29
        nbr_adult = sum(QB06 > 29 & QB06 <= 44, na.rm = TRUE),  # Count of individuals aged 30-44
        nbr_elder_adult = sum(QB06 > 44 & QB06 <= 59, na.rm = TRUE),  # Count of individuals aged 45-59
        nbr_senior = sum(QB06 > 59, na.rm = TRUE)  # Count of individuals aged 60 and above
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        nbr_babies = as.numeric(nbr_babies) %>%
          structure(label = "Nombre d'enfant agée de moins d'un an"),
        nbr_under_14 = as.numeric(nbr_under_14) %>%
          structure(label = "Nombre d'enfant agée de moins de 14 ans"),
        nbr_young_people = as.numeric(nbr_young_people) %>%
          structure(label = "Nombre d'individus agée entre 15 et 29 ans"),
        nbr_adult = as.numeric(nbr_adult) %>%
          structure(label = "Nombre d'individus agées entre 30 et 44 ans"),
        nbr_elder_adult = as.numeric(nbr_elder_adult) %>%
          structure(label = "Nombre d'individus agées entre 45 et 59 ans"),
        nbr_senior = as.numeric(nbr_senior) %>%
          structure(label = "Nombre d'individus agée de plus de 60 ans")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )


#For the RGPH 2013 dataset

## Look if it range for individual data is from 0 to 98
individus_2013 %>% 
  summarise(n_individus=n() %>%
           structure(label = "Nombre des individus"),
           # 315908
           n_nan = sum(is.na(B08))  %>%
           structure(label = "Nombre de valeurs manquantes"),
           # 0
            min_age=min(B08, na.rm = TRUE) %>%
           structure(label = "Minimun âge des individus"),
           # 0
            mean_age = mean(B08, na.rm = TRUE) %>%
           structure(label = "Moyenne âge des individus"), #25.5258
            max_age=max(B08, na.rm = TRUE) %>%
           structure(label = "Maximun âge des individus") #  114
            )



# Left join the 'menage_2013' dataframe with a summarized version of 'individus_2013'
menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing various counts based on age (B08 column)
      summarise(
        nbr_babies = sum(B08 == 0, na.rm = TRUE),  # Count of babies aged 0
        nbr_under_14 = sum(B08 <= 14, na.rm = TRUE),  # Count of individuals under 14
        nbr_young_people = sum(B08 > 14 & B08 <= 29, na.rm = TRUE),  # Count of individuals aged 15-29
        nbr_adult = sum(B08 > 29 & B08 <= 44, na.rm = TRUE),  # Count of individuals aged 30-44
        nbr_elder_adult = sum(B08 > 44 & B08 <= 59, na.rm = TRUE),  # Count of individuals aged 45-59
        nbr_senior = sum(B08 > 59, na.rm = TRUE)  # Count of individuals aged 60 and above
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        nbr_babies = as.numeric(nbr_babies) %>%
          structure(label = "Nombre d'enfant agée de moins d'un an"),
        nbr_under_14 = as.numeric(nbr_under_14) %>%
          structure(label = "Nombre d'enfant agée de moins de 14 ans"),
        nbr_young_people = as.numeric(nbr_young_people) %>%
          structure(label = "Nombre d'individus agée entre 15 et 29 ans"),
        nbr_adult = as.numeric(nbr_adult) %>%
          structure(label = "Nombre d'individus agées entre 30 et 44 ans"),
        nbr_elder_adult = as.numeric(nbr_elder_adult) %>%
          structure(label = "Nombre d'individus agées entre 45 et 59 ans"),
        nbr_senior = as.numeric(nbr_senior) %>%
          structure(label = "Nombre d'individus agée de plus de 60 ans")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )


age_var = c("nbr_babies","nbr_under_14","nbr_young_people","nbr_adult", "nbr_elder_adult", "nbr_senior")
```


### Marital Status

```{r}
#For the RGPH 2002 dataset

     # 0         Monogame
     # 1     Polygame à 2
     # 2     Polygame à 3
     # 3 Polygame à 4 et+
     # 4      Célibataire
     # 5     Veuf / Veuve
     # 6          Divorcé
     # 7            Autre
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
  group_by(ID_MENAGE) %>% 
  summarise(
    Monogame = sum(QB20 == 0, na.rm = TRUE),
    Polygame = sum(QB20 == 1 | QB20 == 2 | QB20 == 3 , na.rm = TRUE),
    Celibataire = sum(QB20 == 4, na.rm = TRUE),
    Veuf = sum(QB20 == 5, na.rm = TRUE),
    Divorce = sum(QB20 == 6, na.rm = TRUE)
  ) %>% 
  ungroup() %>% 
  mutate(
    Monogame = as.numeric(Monogame) %>%
           structure(label = "Nombre de monogame dans le ménage"),
    Polygame = as.numeric(Polygame) %>%
           structure(label = "Nombre de polygame dans le ménage"),
    Celibataire = as.numeric(Celibataire) %>%
           structure(label = "Nombre de celibataire dans le ménage"),
    Veuf = as.numeric(Veuf) %>%
           structure(label = "Nombre de veuf/veuve dans le ménage"),
    Divorce = as.numeric(Divorce) %>%
           structure(label = "Nombre de divorcé dans le ménage")
    ),
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE"))
#For the RGPH 2013 dataset
    #  0                   Monogame
    #  1           Poly/1ère épouse
    #  2 Poly/2ié  épouse/2 épouses
    #  3  Poly/3iè épouse/3 épouses
    #  4 Poly/4iè épouse /4 épouses
    #  5 Poly/5iè épouse /5 épouses
    #  6                Célibataire
    #  7                 Veuf/Veuve
    #  8                 Divorcé(e)
    #  9   Union libre(concubinage)
    # 10                  Séparé(e)
menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        Monogame = sum(B41 == 0, na.rm = TRUE),
        Polygame = sum(B41 == 1 | B41 == 2 | B41 == 3 | B41 == 4 | B41 == 5 , na.rm = TRUE),
        Celibataire = sum(B41 == 6 | B41 == 9 , na.rm = TRUE),
        Veuf = sum(B41 == 7, na.rm = TRUE),
        Divorce = sum(B41 == 8 | B41 == 10 , na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        Monogame = as.numeric(Monogame) %>%
          structure(label = "Nombre de monogame dans le ménage"),
        Polygame = as.numeric(Polygame) %>%
          structure(label = "Nombre de polygame dans le ménage"),
        Celibataire = as.numeric(Celibataire) %>%
          structure(label = "Nombre de celibataire dans le ménage"),
        Veuf = as.numeric(Veuf) %>%
          structure(label = "Nombre de veuf/veuve dans le ménage"),
        Divorce = as.numeric(Divorce) %>%
          structure(label = "Nombre de divorce dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))
mat_status_var = c("Monogame","Polygame","Celibataire","Veuf","Divorce")
```


### Fertility

```{r}
#For the RGPH 2002 dataset
ferti_2002 = c("QB21M","QB21F","QB22M","QB22F","QB23M","QB23F")

menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
  
  mutate(
    fertility = rowSums(select(., .dots=all_of(ferti_2002)),na.rm = T)
  ) %>% 
    group_by(ID_MENAGE) %>% 
  
  summarise(
    fertility = sum(fertility,na.rm = TRUE) %>%
           structure(label = "Fertilité dans le ménage")
    ) , 
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE"))
#For the RGPH 2013 dataset

ferti_2013 = c("B43","B44","B45","B46","B47","B48")
menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
  
  mutate(
    fertility = rowSums(select(., .dots=all_of(ferti_2013)),na.rm = T)
  ) %>% 
    group_by(ID_MENAGE) %>% 
  
  summarise(
    fertility = sum(fertility,na.rm = TRUE) %>%
           structure(label = "Fertilité dans le ménage")
    ) , 
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE"))
fertility_var = c("fertility")
```


### Educational level

```{r}
#For the RGPH 2002 dataset
#QB15 | educ
     # 0      Aucun
     # 1   Primaire
     # 2      Moyen
     # 3 Secondaire
     # 4  Supérieur
     # 9   manquant
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        Aucun = sum(educ == 0, na.rm = TRUE),
        Primaire = sum(educ == 1, na.rm = TRUE),
        Moyen = sum(educ == 2, na.rm = TRUE),
        Secondaire = sum(educ == 3, na.rm = TRUE),
        Superieur = sum(educ == 4, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        Aucun = as.numeric(Aucun) %>%
          structure(label = "Nombre de Aucun dans le ménage"),
        Primaire = as.numeric(Primaire) %>%
          structure(label = "Nombre de Primaire dans le ménage"),
        Moyen = as.numeric(Moyen) %>%
          structure(label = "Nombre de Moyen dans le ménage"),
        Secondaire = as.numeric(Secondaire) %>%
          structure(label = "Nombre de Secondaire dans le ménage"),
        Superieur = as.numeric(Superieur) %>%
          structure(label = "Nombre de Supérieur dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))
#For the RGPH 2013 dataset

individus_2013 <- individus_2013 %>% 
  
  mutate(educ = case_when(
    
    B32 %in% seq(1,3,1)  ~ 1, # If B32 is in pre-school levels, set it to "prescolaire" for c("Petite section","Moyenne section", "Grnde section")
    
    B32 %in% seq(4,9,1) ~ 1, # If B32 corresponds to elementary levels, set it to "elementaire" for c("Cours d'initiation (CI)", "Cours préparatoire (CP)", "Cours élémentaire 1ère année (CE1)", "Cours élémentaire 2ème année (CE2)", "Cours moyen 1ère année (CM1)", "Cours moyen 2ème année (CM2)"
    
    B32 %in% seq(10,13,1) ~ 2, # If B32 corresponds to middle school levels, set it to "moyen" for c("Sixième (6ième)", "Cinquième (5ième)", "Quatrième (4ième)", "Troisième (3ième)")
    
    B32 %in% seq(14,16,1) ~ 3, # If B32 corresponds to high school levels, set it to "secondaire" for c("Seconde", "Première", "Terminale")
    
    B32 %in% seq(17,24,1) ~ 4, # If B32 corresponds to higher education levels, set it to "superieur" c("1ère année", "2ème année", "Licence/3ième année", "Maitrise/4ième année", "DEA/5ième année", "6ième année", "7ième année", "8ième année et +")
    TRUE ~ as.numeric(B32) %>% 
  structure(label = Hmisc::label(individus_2013$B32)))) # Apply labels to the B32 column

menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        Aucun = sum(educ == 0, na.rm = TRUE),
        Primaire = sum(educ == 1, na.rm = TRUE),
        Moyen = sum(educ == 2, na.rm = TRUE),
        Secondaire = sum(educ == 3, na.rm = TRUE),
        Superieur = sum(educ == 4, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        Aucun = as.numeric(Aucun) %>%
          structure(label = "Nombre de Aucun dans le ménage"),
        Primaire = as.numeric(Primaire) %>%
          structure(label = "Nombre de Primaire dans le ménage"),
        Moyen = as.numeric(Moyen) %>%
          structure(label = "Nombre de Moyen dans le ménage"),
        Secondaire = as.numeric(Secondaire) %>%
          structure(label = "Nombre de Secondaire dans le ménage"),
        Superieur = as.numeric(Superieur) %>%
          structure(label = "Nombre de Supérieur dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))

educ_var = c("Aucun","Primaire","Moyen","Secondaire","Superieur")
```


### Literacy level

```{r}
#For the RGPH 2002 dataset
#QB16A,QB16B,QB16C,QB16D,QB16E,QB16F,QB16G,QB16H,QB16I,QB16J
#1-None, 2-French, 3-Arabic, 4-Wolof, 5-Pulaar, 6-Sereer, 7-Mandingo, 8-Diola, 9 -Soninke, 10-Other Language.
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        Literacy_None = sum(QB16A == 1, na.rm = TRUE),
        Literacy_French = sum(QB16B == 1, na.rm = TRUE),
        Literacy_Arabic = sum(QB16C == 1, na.rm = TRUE),
        Literacy_Wolof = sum(QB16D == 1, na.rm = TRUE),
        Literacy_Pulaar = sum(QB16E == 1, na.rm = TRUE),
        Literacy_Sereer = sum(QB16F == 1, na.rm = TRUE),
        Literacy_Mandingo = sum(QB16G == 1, na.rm = TRUE),
        Literacy_Diola = sum(QB16H == 1, na.rm = TRUE),
        Literacy_Soninke = sum(QB16I == 1, na.rm = TRUE),
        Literacy_Other = sum(QB16J == 1, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        Literacy_None = as.numeric(Literacy_None) %>%
          structure(label = "Nombre de Aucun dans le ménage"),
        Literacy_French = as.numeric(Literacy_French) %>%
          structure(label = "Nombre de French dans le ménage"),
        Literacy_Arabic = as.numeric(Literacy_Arabic) %>%
          structure(label = "Nombre de Arabic dans le ménage"),
        Literacy_Wolof = as.numeric(Literacy_Wolof) %>%
          structure(label = "Nombre de Wolof dans le ménage"),
        Literacy_Pulaar = as.numeric(Literacy_Pulaar) %>%
          structure(label = "Nombre de Pulaar dans le ménage"),
        Literacy_Sereer = as.numeric(Literacy_Sereer) %>%
          structure(label = "Nombre de Sereer dans le ménage"),
        Literacy_Mandingo = as.numeric(Literacy_Mandingo) %>%
          structure(label = "Nombre de Mandingo dans le ménage"),
        Literacy_Diola = as.numeric(Literacy_Diola) %>%
          structure(label = "Nombre de Diola dans le ménage"),
        Literacy_Soninke = as.numeric(Literacy_Soninke) %>%
          structure(label = "Nombre de Soninke dans le ménage"),
        Literacy_Other = as.numeric(Literacy_Other) %>%
          structure(label = "Nombre de Other dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))
#For the RGPH 2013 dataset
#B34_FR,B34_AR,B34_WO,B34_PU,B34_SE,B34_MA,B34_DI,B34_SO
menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        Literacy_French = sum(B34_FR == 1, na.rm = TRUE),
        Literacy_Arabic = sum(B34_AR == 1, na.rm = TRUE),
        Literacy_Wolof = sum(B34_WO == 1, na.rm = TRUE),
        Literacy_Pulaar = sum(B34_PU == 1, na.rm = TRUE),
        Literacy_Sereer = sum(B34_SE == 1, na.rm = TRUE),
        Literacy_Mandingo = sum(B34_MA == 1, na.rm = TRUE),
        Literacy_Diola = sum(B34_DI == 1, na.rm = TRUE),
        Literacy_Soninke = sum(B34_SO == 1, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        Literacy_French = as.numeric(Literacy_French) %>%
          structure(label = "Nombre de French dans le ménage"),
        Literacy_Arabic = as.numeric(Literacy_Arabic) %>%
          structure(label = "Nombre de Arabic dans le ménage"),
        Literacy_Wolof = as.numeric(Literacy_Wolof) %>%
          structure(label = "Nombre de Wolof dans le ménage"),
        Literacy_Pulaar = as.numeric(Literacy_Pulaar) %>%
          structure(label = "Nombre de Pulaar dans le ménage"),
        Literacy_Sereer = as.numeric(Literacy_Sereer) %>%
          structure(label = "Nombre de Sereer dans le ménage"),
        Literacy_Mandingo = as.numeric(Literacy_Mandingo) %>%
          structure(label = "Nombre de Mandingo dans le ménage"),
        Literacy_Diola = as.numeric(Literacy_Diola) %>%
          structure(label = "Nombre de Diola dans le ménage"),
        Literacy_Soninke = as.numeric(Literacy_Soninke) %>%
          structure(label = "Nombre de Soninke dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))
literacy_var = c("Literacy_French","Literacy_Arabic","Literacy_Wolof","Literacy_Pulaar","Literacy_Sereer","Literacy_Mandingo","Literacy_Diola","Literacy_Soninke")
```

### Migration variable

```{r}
#For the RGPH 2002 dataset
#the place of residence 5 years ago was also asked for residents present, absent or for visitors.

     # 1 Resident present
     # 2  Resident absent
     # 3         Visiteur
menage_2002 <- menage_2002 %>% 
  left_join(
    individus_2002 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        resident_pres = sum(QB08 == 1, na.rm = TRUE),
        resident_abs = sum(QB08 == 2, na.rm = TRUE),
        resident_vis = sum(QB08 == 3, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        resident_pres = as.numeric(resident_pres) %>%
          structure(label = "Nombre de Résident Présent dans le ménage"),
        resident_abs = as.numeric(resident_abs) %>%
          structure(label = "Nombre de Résident Absent dans le ménage"),
        resident_vis = as.numeric(resident_vis) %>%
          structure(label = "Nombre de Visiteur dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))
#For the RGPH 2013 dataset

     # 1 Résident Présent
     # 2  Résident Absent
     # 3         Visiteur
menage_2013 <- menage_2013 %>% 
  left_join(
    individus_2013 %>% 
      group_by(ID_MENAGE) %>% 
      summarise(
        resident_pres = sum(B11 == 1, na.rm = TRUE),
        resident_abs = sum(B11 == 2, na.rm = TRUE),
        resident_vis = sum(B11 == 3, na.rm = TRUE)
      ) %>% 
      ungroup() %>% 
      mutate(
        resident_pres = as.numeric(resident_pres) %>%
          structure(label = "Nombre de Résident Présent dans le ménage"),
        resident_abs = as.numeric(resident_abs) %>%
          structure(label = "Nombre de Résident Absent dans le ménage"),
        resident_vis = as.numeric(resident_vis) %>%
          structure(label = "Nombre de Visiteur dans le ménage")
      ),
    
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE"))

mig_var = c("resident_pres","resident_abs","resident_vis")
```

### Maternal death variable

```{r echo=TRUE, eval=TRUE, include=TRUE}

# For the RGPH 2002 dataset, 'deces'
menage_2002 <- menage_2002 %>% 
  # Left join 'menage_2002' with a summarized version of 'deces_2002' based on household ID
  left_join(
    deces_2002 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of maternal deaths (C06 column) in the household
      summarise(
        deces_mat_oui = sum(C06 == 1, na.rm = TRUE),  # Count of maternal deaths: Yes
        deces_mat_non = sum(C06 == 2, na.rm = TRUE),  # Count of maternal deaths: No
        deces_mat_nsp = sum(C06 == 3, na.rm = TRUE)   # Count of maternal deaths: NSP (Not Specified)
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        deces_mat_oui = as.numeric(deces_mat_oui) %>%
          structure(label = "Nombre de décès maternels dans le ménage: Oui."),
        deces_mat_non = as.numeric(deces_mat_non) %>%
          structure(label = "Nombre de décès maternels dans le ménage: Non."),
        deces_mat_nsp = as.numeric(deces_mat_nsp) %>%
          structure(label = "Nombre de décès maternels dans le ménage: NSP.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

#For the RGPH 2013 dataset
#No variable

deces_mat_var = c("deces_mat_oui","deces_mat_non","deces_mat_nsp")
```

## Socio-economic variables


### Employed variable

```{r echo=TRUE, eval=TRUE, include=TRUE}
#For the RGPH 2002 dataset
     # 1      Employeur
     # 2   Indépendant
     # 3       Salarié
     # 4      Stagiaire
     # 5 Aide familaile
     # 6       Apprenti
     # 7          Autre

menage_2002 <- menage_2002 %>% 
  # Left join 'menage_2013' with a summarized version of 'individus_2013' based on household ID
  left_join(
    individus_2002 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their primary occupation (QB19 column) in the household
      summarise(
        occup_independant = sum(QB19 == 2, na.rm = TRUE),  # Count of individuals who are independent workers
        occup_employeur = sum(QB19 == 1, na.rm = TRUE),    # Count of individuals who are employers/bosses
        occup_salarie = sum(QB19 == 3 | QB19 == 4, na.rm = TRUE),  # Count of individuals who are employees
        occup_apprenti = sum(QB19 == 6, na.rm = TRUE),     # Count of individuals who are apprentices
        occup_aideFamilial = sum(QB19 == 5, na.rm = TRUE),  # Count of individuals who are family aides
        occup_autre = sum(QB19 == 7, na.rm = TRUE)          # Count of individuals with other primary occupations
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        occup_independant = as.numeric(occup_independant) %>%
          structure(label = "Nombre d'individus qui sont des travailleurs indépendants dans leur occupation principale dans le ménage."),
        occup_employeur = as.numeric(occup_employeur) %>%
          structure(label = "Nombre d'individus qui sont des employeurs/patrons dans leur occupation principale dans le ménage."),
        occup_salarie = as.numeric(occup_salarie) %>%
          structure(label = "Nombre d'individus qui sont des salariés/stagiaires dans leur occupation principale dans le ménage."),
        occup_apprenti = as.numeric(occup_apprenti) %>%
          structure(label = "Nombre d'individus qui sont des apprentis dans leur occupation principale dans le ménage."),
        occup_aideFamilial = as.numeric(occup_aideFamilial) %>%
          structure(label = "Nombre d'individus qui sont des aides familiaux dans leur occupation principale dans le ménage."),
        occup_autre = as.numeric(occup_autre) %>%
          structure(label = "Nombre d'individus qui ont d'autres activités comme occupation principale dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

#For the RGPH 2013 dataset
     # 1     Travailleur indépendant
     # 2             Employeur/Patron
     # 3  Salarié/Employé permanent
     # 4 Salarié/Employé temporaire
     # 5                     Apprenti
     # 6                Aide familial
     # 7                       Autres

# For the 2013 household dataset, 'menage_2013'
menage_2013 <- menage_2013 %>% 
  # Left join 'menage_2013' with a summarized version of 'individus_2013' based on household ID
  left_join(
    individus_2013 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their primary occupation (B38 column) in the household
      summarise(
        occup_independant = sum(B38 == 1, na.rm = TRUE),  # Count of individuals who are independent workers
        occup_employeur = sum(B38 == 2, na.rm = TRUE),    # Count of individuals who are employers/bosses
        occup_salarie = sum(B38 == 3 | B38 == 4, na.rm = TRUE),  # Count of individuals who are employees
        occup_apprenti = sum(B38 == 5, na.rm = TRUE),     # Count of individuals who are apprentices
        occup_aideFamilial = sum(B38 == 6, na.rm = TRUE),  # Count of individuals who are family aides
        occup_autre = sum(B38 == 7, na.rm = TRUE)          # Count of individuals with other primary occupations
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        occup_independant = as.numeric(occup_independant) %>%
          structure(label = "Nombre d'individus qui sont des travailleurs indépendants dans leur occupation principale dans le ménage."),
        occup_employeur = as.numeric(occup_employeur) %>%
          structure(label = "Nombre d'individus qui sont des employeurs/patrons dans leur occupation principale dans le ménage."),
        occup_salarie = as.numeric(occup_salarie) %>%
          structure(label = "Nombre d'individus qui sont des salariés/stagiaires dans leur occupation principale dans le ménage."),
        occup_apprenti = as.numeric(occup_apprenti) %>%
          structure(label = "Nombre d'individus qui sont des apprentis dans leur occupation principale dans le ménage."),
        occup_aideFamilial = as.numeric(occup_aideFamilial) %>%
          structure(label = "Nombre d'individus qui sont des aides familiaux dans leur occupation principale dans le ménage."),
        occup_autre = as.numeric(occup_autre) %>%
          structure(label = "Nombre d'individus qui ont d'autres activités comme occupation principale dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

occup_var=c("occup_independant","occup_employeur","occup_salarie","occup_apprenti","occup_aideFamilial","occup_autre")     
```

### Occupational status variable

```{r echo=TRUE, eval=TRUE, include=TRUE}
#For the RGPH 2002 dataset
##QB17
     # 1          Occupé
     # 2          Chomeur
     # 3   Eleve_etudiant
     # 4   Femme au foyer
     # 5        Retraité
     # 6       Vieillesse
     # 7 Accident/maladie
     # 8         Mendiant
     # 9            Autre

# For the 2013 household dataset, 'menage_2013'
menage_2002 <- menage_2002 %>% 
  # Left join 'menage_2013' with a summarized version of 'individus_2013' based on household ID
  left_join(
    individus_2002 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their main activity status (B36 column) in the household
      summarise(
        sit_occup_occup = sum(QB17 == 1, na.rm = TRUE),  # Count of individuals who are occupied in their primary activity
        sit_occup_chomeur = sum(QB17 == 2, na.rm = TRUE),  # Count of individuals who are unemployed or seeking their first job
        sit_occup_etudiant = sum(QB17 == 3, na.rm = TRUE),  # Count of individuals who are students/pupils
        sit_occup_foyer = sum(QB17 == 4, na.rm = TRUE),  # Count of individuals who are occupied at home
        sit_occup_retraite = sum(QB17 == 5 | QB17 == 6, na.rm = TRUE),  # Count of individuals who are retired or elderly

        sit_occup_rentier = 0,  # Count of individuals who are retirees
        
        sit_occup_autre = sum(QB17 == 7 | QB17 == 8 | QB17 == 9, na.rm = TRUE)  # Count of individuals with other main activity status
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        sit_occup_occup = as.numeric(sit_occup_occup) %>%
          structure(label = "Nombre d'individus qui sont occupés dans leur occupation principale dans le ménage."),
        sit_occup_chomeur = as.numeric(sit_occup_chomeur) %>%
          structure(label = "Nombre d'individus qui sont au chômage ou à la recherche d'un premier emploi dans leur occupation principale dans le ménage."),
        sit_occup_foyer = as.numeric(sit_occup_foyer) %>%
          structure(label = "Nombre d'individus qui sont occupés au foyer dans leur occupation principale dans le ménage."),
        sit_occup_etudiant = as.numeric(sit_occup_etudiant) %>%
          structure(label = "Nombre d'individus qui sont étudiants/élèves dans leur occupation principale dans le ménage."),
        sit_occup_rentier = as.numeric(sit_occup_rentier) %>%
          structure(label = "Nombre d'individus qui sont rentiers dans leur occupation principale dans le ménage."),
        sit_occup_retraite = as.numeric(sit_occup_retraite) %>%
          structure(label = "Nombre d'individus qui sont retraités ou personnes du 3ème âge dans leur occupation principale dans le ménage."),
        sit_occup_autre = as.numeric(sit_occup_autre) %>%
          structure(label = "Nombre d'individus qui ont d'autres situations dans leur occupation principale dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )


#For the RGPH 2013 dataset
##B36
     # 1                              Occupé
     # 2            Chômeur ayant travaillé
     # 3   A la recherche d'un premier emploi
     # 4                     Occupé au foyer
     # 5                      Etudiant/elève
     # 6                              Rentier
     # 7 Retraités et personne du 3ème âge
     # 8                      Autres inactifs

# For the 2013 household dataset, 'menage_2013'
menage_2013 <- menage_2013 %>% 
  # Left join 'menage_2013' with a summarized version of 'individus_2013' based on household ID
  left_join(
    individus_2013 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their main activity status (B36 column) in the household
      summarise(
        sit_occup_occup = sum(B36 == 1, na.rm = TRUE),  # Count of individuals who are occupied in their primary activity
        sit_occup_chomeur = sum(B36 == 2 | B36 == 3, na.rm = TRUE),  # Count of individuals who are unemployed or seeking their first job
        sit_occup_foyer = sum(B36 == 4, na.rm = TRUE),  # Count of individuals who are occupied at home
        sit_occup_etudiant = sum(B36 == 5, na.rm = TRUE),  # Count of individuals who are students/pupils
        sit_occup_rentier = sum(B36 == 6, na.rm = TRUE),  # Count of individuals who are retirees
        sit_occup_retraite = sum(B36 == 7, na.rm = TRUE),  # Count of individuals who are retired or elderly
        sit_occup_autre = sum(B36 == 8, na.rm = TRUE)  # Count of individuals with other main activity status
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        sit_occup_occup = as.numeric(sit_occup_occup) %>%
          structure(label = "Nombre d'individus qui sont occupés dans leur occupation principale dans le ménage."),
        sit_occup_chomeur = as.numeric(sit_occup_chomeur) %>%
          structure(label = "Nombre d'individus qui sont au chômage ou à la recherche d'un premier emploi dans leur occupation principale dans le ménage."),
        sit_occup_foyer = as.numeric(sit_occup_foyer) %>%
          structure(label = "Nombre d'individus qui sont occupés au foyer dans leur occupation principale dans le ménage."),
        sit_occup_etudiant = as.numeric(sit_occup_etudiant) %>%
          structure(label = "Nombre d'individus qui sont étudiants/élèves dans leur occupation principale dans le ménage."),
        sit_occup_rentier = as.numeric(sit_occup_rentier) %>%
          structure(label = "Nombre d'individus qui sont rentiers dans leur occupation principale dans le ménage."),
        sit_occup_retraite = as.numeric(sit_occup_retraite) %>%
          structure(label = "Nombre d'individus qui sont retraités ou personnes du 3ème âge dans leur occupation principale dans le ménage."),
        sit_occup_autre = as.numeric(sit_occup_autre) %>%
          structure(label = "Nombre d'individus qui ont d'autres situations dans leur occupation principale dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

sit_occup_var=c("sit_occup_occup","sit_occup_chomeur","sit_occup_foyer","sit_occup_etudiant","sit_occup_rentier","sit_occup_retraite","sit_occup_autre")  
```

### Income variable

Two identical questions asked for the two censuses are: (1) during the last 12 months, has your household skipped a meal because of lack of income? (2) During the last 12 months, has it happened that a sick member of your household could not receive medical care because of lack of income? The answers to these questions were either 1. Yes or 2. No

```{r}
#For the RGPH 2002 dataset, base menage
#E16: Repas sauté
     # 1   Oui
     # 2   Non
     # 9    nd
#E17:Soins médicaux
     # 1            Oui
     # 2            Non
     # 3 Non Applicable
     # 9             nd
menage_2002 <- menage_2002 %>% 
  dplyr::mutate(I_repas_saute = ifelse(E16==1,1,0) %>% structure(label=Hmisc::label(individus_2002$E16)),
                I_soins_medicaux = ifelse(E17==1,1,0) %>% structure(label=Hmisc::label(individus_2002$E17)))

#For the RGPH 2013 dataset
menage_2013 <- menage_2013 %>% 
  dplyr::mutate(I_repas_saute = ifelse(E18==1,1,0) %>% structure(label=Hmisc::label(individus_2002$E18)),
                I_soins_medicaux = ifelse(E20==1,1,0) %>% structure(label=Hmisc::label(individus_2002$E20)))
income_var = c("I_repas_saute","I_soins_medicaux")
```

## Household variables

### Capital goods variable

```{r}
#For the RGPH 2002 dataset
#OK
# 1. Radio, 2. Television, 3. Video/VCD/DVD, 4. Refrigerator/freezer, 5. Telephone, 6. Improved fireplace, 7. Air conditioner, 8. Sewing machine, 9. Water heater, 10. Stove

#For the RGPH 2013 dataset
```

### Access to basic equipment

```{r}
#For the RGPH 2002 dataset
#OK
# 1. Main mode of household waste disposal, 2. Main mode of waste water disposal, 3. Main mode of lighting, 4. Main fuel for cooking
#For the RGPH 2013 dataset
```

### Household size

```{r}
#For the RGPH 2002 dataset
#OK
#For the RGPH 2013 dataset
```

### Head of household

```{r}
## In the following we merge the "menage_2002" with the individual dataset
menage_2002 <- menage_2002 %>% 
  left_join(
    
    individus_2002 %>% 
      
      # Filtering to get only individuals marked as "Chef de menage" in QB03
      filter(QB03== 1) %>% #"Chef de menage"
      
      # Creating the "niveau_instruction_cm" variable from QB15 column
      mutate(niveau_instruction_cm = educ %>% structure(label=paste0(Hmisc::label(individus_2002$educ), " CM")),
      
             # Creating the "sexe_cm" variable from QB04 column
          sexe_cm = QB04 %>% structure(label= paste0(Hmisc::label(individus_2002$QB04), " Sexe CM")),
          
          # Creating the "age_cm" variable from QB06 column
          age_cm = QB06 %>% structure(label=paste0(Hmisc::label(individus_2002$QB06), "Age CM"))) %>% 
      # Selecting specific columns from the filtered dataset
      select(ID_MENAGE, niveau_instruction_cm, sexe_cm, age_cm) ,
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE")# = "ID_MENAGE", "id_dr" = "DR"
             
    
  )


## In the following we merge the "menage_2013" with the individual dataset
menage_2013 <- menage_2013 %>% 
  left_join(
    
    individus_2013 %>% 
      
      # Filtering to get only individuals marked as "Chef de menage" in B04
      filter(B04==1) %>% #"Chef de ménage"
      
      # Renaming the B32 column to "niveau_instruction_cm"
      mutate(niveau_instruction_cm = educ %>% structure(label=paste0(Hmisc::label(individus_2013$educ),"du CM")), 
      
      # Creating the "sexe_cm" variable from B06 column
          sexe_cm = B06 %>% structure(label= paste0(Hmisc::label(individus_2013$B06), "du CM")),
          
          # Creating the "age_cm" variable from QB08 column
          age_cm = B08 %>% structure(label=paste0(Hmisc::label(individus_2013$B08), " du CM")))%>% 
      # Selecting specific columns from the filtered dataset
      select(ID_MENAGE, niveau_instruction_cm, sexe_cm, age_cm),
      
      # Merging by matching columns in both datasets
      by = c("ID_MENAGE")# = "ID_MENAGE", "id_dr" = "IDDR"
    
  )




cm_var=c("niveau_instruction_cm","sexe_cm","age_cm")
```

```{r}
# FOr CM
#For the RGPH 2002 dataset
#QB15 | educ
     # 0      Aucun
     # 1   Primaire
     # 2      Moyen
     # 3 Secondaire
     # 4  Supérieur
     # 9   manquant
menage_2002 <- menage_2002 %>% 
      mutate(
        Aucun_cm =ifelse(niveau_instruction_cm == 0, 1,0),
        Primaire_cm = ifelse(niveau_instruction_cm == 1, 1,0),
        Moyen_cm = ifelse(niveau_instruction_cm == 2, 1,0),
        Secondaire_cm = ifelse(niveau_instruction_cm == 3, 1,0),
        Superieur_cm = ifelse(niveau_instruction_cm == 4, 1,0)
      ) %>% 
      
  mutate(
        Aucun_cm = as.numeric(Aucun_cm)%>%
          structure(label = "CM a aucun niveau de d'instruction") ,
        Primaire_cm = as.numeric(Primaire_cm) %>%
          structure(label = "CM a un niveau d'instruction Primaire"),
        Moyen_cm = as.numeric(Moyen_cm) %>%
          structure(label = "CM a un niveau d'instruction Moyen"),
        Secondaire_cm = as.numeric(Secondaire_cm) %>%
          structure(label = "CM a un niveau d'instruction de Secondaire"),
        Superieur_cm = as.numeric(Superieur_cm) %>%
          structure(label = "CM a un niveau d'instruction Supérieur"))

#For the RGPH 2013 dataset

menage_2013 <- menage_2013 %>% 
  
  mutate(
        Aucun_cm =ifelse(niveau_instruction_cm == 0, 1,0),
        Primaire_cm = ifelse(niveau_instruction_cm == 1, 1,0),
        Moyen_cm = ifelse(niveau_instruction_cm == 2, 1,0),
        Secondaire_cm = ifelse(niveau_instruction_cm == 3, 1,0),
        Superieur_cm = ifelse(niveau_instruction_cm == 4, 1,0)
      ) %>% 
      
  mutate(
        Aucun_cm = as.numeric(Aucun_cm)%>%
          structure(label = "CM a aucun niveau de d'instruction") ,
        Primaire_cm = as.numeric(Primaire_cm) %>%
          structure(label = "CM a un niveau d'instruction Primaire"),
        Moyen_cm = as.numeric(Moyen_cm) %>%
          structure(label = "CM a un niveau d'instruction Moyen"),
        Secondaire_cm = as.numeric(Secondaire_cm) %>%
          structure(label = "CM a un niveau d'instruction de Secondaire"),
        Superieur_cm = as.numeric(Superieur_cm) %>%
          structure(label = "CM a un niveau d'instruction Supérieur"))
    

educ_var = c("Aucun_cm","Primaire_cm","Moyen_cm","Secondaire_cm","Superieur_cm")
```



### Household composition

```{r echo=TRUE, eval=TRUE, include=TRUE}
#For the RGPH 2002 dataset
#QB03 
     # 1       Chef de menage
     # 2               Epouse
     # 3               Enfant
     # 4               Parent
     # 5         Grand-parent
     # 6         Frere, soeur
     # 7    Petit-fils/-fille
     # 8       Autres parents
     # 9 Sans lien de parente

# For the 2002 household dataset, 'menage_2002'
menage_2002 <- menage_2002 %>% 
  # Left join 'menage_2002' with a summarized version of 'individus_2002' based on household ID
  left_join(
    individus_2002 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their relationship to the head of the household (QB03 column)
      summarise(
        liencm_cm = sum(QB03 == 1, na.rm = TRUE),  # Count of individuals who are the head of the household (Chef de Ménage - CM)
        liencm_epous = sum(QB03 == 2, na.rm = TRUE),  # Count of individuals who are spouses of the head
        liencm_enf = sum(QB03 == 3, na.rm = TRUE),   # Count of individuals who are children, nephew or children in law of the head
        liencm_parent = sum(QB03 == 4 | QB03 == 5| QB03 == 8, na.rm = TRUE),  # Count of individuals who are parents/grandparents or other relatives of the head
        liencm_frere = sum(QB03 == 6, na.rm = TRUE),  # Count of individuals who are siblings of the head
        liencm_petit_fils = sum(QB03 == 7, na.rm = TRUE),  # Count of individuals who are grandchildren of the head
        liencm_sans_lien = sum(QB03 == 9, na.rm = TRUE)   # Count of individuals without any familial relationship/maid to the head
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        liencm_cm = as.numeric(liencm_cm) %>%
          structure(label = "Nombre de Chefs de Ménage résidant dans le ménage."),
        liencm_epous = as.numeric(liencm_epous) %>%
          structure(label = "Nombre d'épouses des Chefs de Ménage résidant dans le ménage."),
        liencm_enf = as.numeric(liencm_enf) %>%
          structure(label = "Nombre d'enfants des Chefs de Ménage résidant dans le ménage."),
        liencm_parent = as.numeric(liencm_parent) %>%
          structure(label = "Nombre de parents/grands-parents ou autres parents des Chefs de Ménage résidant dans le ménage."),
        liencm_frere = as.numeric(liencm_frere) %>%
          structure(label = "Nombre de frères et sœurs des Chefs de Ménage résidant dans le ménage."),
        liencm_petit_fils = as.numeric(liencm_petit_fils) %>%
          structure(label = "Nombre de petits-fils/petites-filles des Chefs de Ménage résidant dans le ménage."), 
        liencm_sans_lien = as.numeric(liencm_sans_lien) %>%
          structure(label = "Nombre d'individus sans lien de parenté avec les Chefs de Ménage vivant dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

#For the RGPH 2013 dataset
#B04
    #  0 Sans lien de parenté avec le CM
    #  1                  Chef de ménage
    #  2         Epouse (ou époux) du CM
    #  3              Fils ou fille du CM
    #  4       Frère ou S�\u0093ur du CM
    #  5             Père ou Mère du CM
    #  6             Petit(e)s fils/fille
    #  7                     Neveu/Nièce
    #  8                      Oncle/Tante
    #  9            Belle fille/Beau-fils
    # 10             Autres parents du CM
    # 11                       Domestique

# For the 2013 household dataset, 'menage_2013'
menage_2013 <- menage_2013 %>% 
  # Left join 'menage_2013' with a summarized version of 'individus_2013' based on household ID
  left_join(
    individus_2013 %>% 
      # Grouping by household ID
      group_by(ID_MENAGE) %>% 
      # Summarizing counts of individuals based on their relationship to the head of the household (QB03 column)
      summarise(
        liencm_cm = sum(B04 == 1, na.rm = TRUE),  # Count of individuals who are the head of the household (Chef de Ménage - CM)
        liencm_epous = sum(B04 == 2, na.rm = TRUE),  # Count of individuals who are spouses of the head
        liencm_enf = sum(B04 == 3 | B04 == 7| B04 == 9 , na.rm = TRUE),  # Count of individuals who are children, nephew or children in law of the head
        liencm_parent = sum(B04 == 5 | B04 == 8| B04 == 10, na.rm = TRUE),  # Count of individuals who are parents/grandparents or other relatives of the head
        liencm_frere = sum(B04 == 4, na.rm = TRUE),  # Count of individuals who are siblings of the head
        liencm_petit_fils = sum(B04 == 6, na.rm = TRUE),  # Count of individuals who are grandchildren of the head
        liencm_sans_lien = sum(B04 == 0 | B04 == 11, na.rm = TRUE)  # Count of individuals without any familial relationship/maid to the head
      ) %>% 
      ungroup() %>% 
      # Adding labels to the counts for better readability
      mutate(
        liencm_cm = as.numeric(liencm_cm) %>%
          structure(label = "Nombre de Chefs de Ménage résidant dans le ménage."),
        liencm_epous = as.numeric(liencm_epous) %>%
          structure(label = "Nombre d'épouses des Chefs de Ménage résidant dans le ménage."),
        liencm_enf = as.numeric(liencm_enf) %>%
          structure(label = "Nombre d'enfants des Chefs de Ménage résidant dans le ménage."),
        liencm_parent = as.numeric(liencm_parent) %>%
          structure(label = "Nombre de parents/grands-parents ou autres parents des Chefs de Ménage résidant dans le ménage."),
        liencm_frere = as.numeric(liencm_frere) %>%
          structure(label = "Nombre de frères et sœurs des Chefs de Ménage résidant dans le ménage."),
        liencm_petit_fils = as.numeric(liencm_petit_fils) %>%
          structure(label = "Nombre de petits-fils/petites-filles des Chefs de Ménage résidant dans le ménage."), 
        liencm_sans_lien = as.numeric(liencm_sans_lien) %>%
          structure(label = "Nombre d'individus sans lien de parenté avec les Chefs de Ménage vivant dans le ménage.")
      ),
    # Merging by matching columns in both datasets
    by = c("ID_MENAGE")
  )

liencm_var=c("liencm_cm","liencm_epous","liencm_enf","liencm_parent","liencm_frere","liencm_petit_fils","liencm_sans_lien")
```



```{r}
# Encadré 1: Indice de peuplement
# L'indice de peuplement donne une certaine idée de la promiscuité qui règne dans un ménage. Pour un ménage donné, l'indice de peuplement est le rapport entre le nombre de membres et le nombre de pièces à usage d'habitation ou, plus simplement, le nombre de personnes par pièces habitées. Ainsi, un ménage est dit : 
# •sous-peuplé si son indice est inférieur à 1 (c'est-à-dire quand il y a plus de pièces que de
# personnes)
# •surpeuplé s'il est égal ou supérieur à 3 (c'est-à-dire quand il y a 3 personnes ou plus par pièce)
# •normalement peuplé si son indice de peuplement est compris entre 1 et 2 (c'est-à-dire quand
# le nombre de personnes par pièce est 1 ou 2).



# Quelques exemples d'analyse proposés par mon Directeur de Thése
# 
# •	La taille des ménages, il faut aussi regarder l’évolution des ménages.
# •	Combien de personnes dans les ménages. Plus de ménage
# •	Analyse l’âge
# •	Proportion des ménages avec enfants ou sans enfants
# •	La taille des ménages par sexe du chef de ménage
# •	Faire des groupes par sexe
# •	Par statut matrimonial
# •	Par déduction
# •	Différence entre les sexes de chef de ménages
# •	Taux de fertilité, taux de mortalités
```


```{r eval=FALSE, include=FALSE}
# Replace 'NA' with 0 (for the nbr deces )

## For RGPH 2013
menage_2013 <- menage_2013 %>%
  
  #replace all NA values with zero
  mutate_if(is.numeric , replace_na, replace = 0)

## For RGPH 2002
menage_2002 <- menage_2002 %>%
  
  #replace all NA values with zero
  mutate_if(is.numeric , replace_na, replace = 0)



```



```{r eval=FALSE, include=FALSE}
# Ranking variables

## For RGPH 2013
menage_2013 <- menage_2013 %>%
 select(id_dr, ID_MENAGE, RGPH, everything())

## For RGPH 2002
menage_2002 <- menage_2002 %>%
 select(id_dr, ID_MENAGE, RGPH, everything())
```






*Calculation of indices on household-related component variables (Age of household heads, by gender between the two censuses)*

```{r}
# Select the CM (Chef de ménage) and compute the average age

# Select relevant columns (B04, B06, B08, RGPH) from the "individus_2013" dataset
individus_2013 %>% 
  select(B04, B06, B08, RGPH) %>%

  # Filter to select only records where "B04" (Chef de ménage) equals 1
  # This is done to identify the household head
  filter(B04== 1) %>% # Household head (B04 == 1) "Chef de ménage"

  # Rename columns for clarity
  rename(
    sexe_cm = B06,        # Renamed B06 to "sexe_cm"
    lien_cm = B04,        # Renamed B04 to "lien_cm"
    age_cm = B08,         # Renamed B08 to "age_cm"
  ) %>% 

  # Combine this filtered data with a similar operation on the "individus_2002" dataset
  bind_rows(
    individus_2002 %>% 
      select(QB03, QB04, QB06, RGPH) %>%
    
    # Filter to select only records where "QB03" (Chef de ménage) equals "Chef de menage"
    filter(QB03 == 1 ) %>% # Household head (QB03 == 1) "Chef de menage"
      
    # Rename columns for consistency
    rename(
      sexe_cm = QB04,        # Renamed QB04 to "sexe_cm"
      lien_cm = QB03,        # Renamed QB03 to "lien_cm"
      age_cm = QB06,         # Renamed QB06 to "age_cm"
    ) 
    
  ) %>% 

  # Remove the "lien_cm" (link to household head) column
  select(-lien_cm) %>% 
  mutate( sexe_cm = ifelse(sexe_cm ==2, 0, 1)) %>% 
  
  # Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    label = list(sexe_cm ~ "Sexe du chef de ménage (Masculin = 1, Féminin = 0)",
                 age_cm ~ "Age du chef de ménage (en année)"),
    statistic = list(all_continuous() ~ "{mean}" 
                     #all_categorical() ~ "{p}"
                     ),
    digits = everything() ~ c(0,0),
    missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Caractéristique du Chef de ménage entre les deux recensements") %>%
  
  # Add difference statistics to the table
  add_difference() 
```



# Multivariate calculation for poverty indices (MPI)

## Education dimension

### Cleaning of the "education" variable

```{r eval=FALSE, include=FALSE}

# Transform values in the "QB15" column of the "individus_2002" dataset
# The transformation categorizes educational levels into descriptive labels

individus_2002 <- individus_2002 %>% 
  
  mutate(QB15 = case_when(
    
    QB15 == 0 ~ "aucune",      # If QB15 is 0, set it to "aucune"
    
    QB15 == 1 ~ "prescolaire", # If QB15 is 1, set it to "prescolaire"
    
    QB15 %in% seq(2,7,1) ~ "elementaire", # If QB15 is 2 to 7, set it to "elementaire"
    
    QB15 %in% seq(8,11,1) ~ "moyen", # If QB15 is 8 to 11, set it to "moyen"
    
    QB15 %in% seq(12,14,1) ~ "secondaire", # If QB15 is 12 to 14, set it to "secondaire"
    
    QB15 %in% seq(15,19,1) ~ "superieur", # If QB15 is 15 to 19, set it to "superieur"
    
    TRUE ~ as.character(QB15) # If none, keep the old value
    ) %>% 
  structure(label = Hmisc::label(individus_2002$QB15))) # Apply labels to the QB15 column


# Transform values in the "B32" column of the "individus_2013" dataset
# The transformation categorizes educational levels from codes into descriptive labels

individus_2013 <- individus_2013 %>% 
  
  mutate(B32 = case_when(
    
    B32 %in% seq(1,3,1)  ~ "prescolaire", # If B32 is in pre-school levels, set it to "prescolaire" for c("Petite section","Moyenne section", "Grnde section")
    
    B32 %in% seq(4,9,1) ~ "elementaire", # If B32 corresponds to elementary levels, set it to "elementaire" for c("Cours d'initiation (CI)", "Cours préparatoire (CP)", "Cours élémentaire 1ère année (CE1)", "Cours élémentaire 2ème année (CE2)", "Cours moyen 1ère année (CM1)", "Cours moyen 2ème année (CM2)"
    
    B32 %in% seq(10,13,1) ~ "moyen", # If B32 corresponds to middle school levels, set it to "moyen" for c("Sixième (6ième)", "Cinquième (5ième)", "Quatrième (4ième)", "Troisième (3ième)")
    
    B32 %in% seq(14,16,1) ~ "secondaire", # If B32 corresponds to high school levels, set it to "secondaire" for c("Seconde", "Première", "Terminale")
    
    B32 %in% seq(17,24,1) ~ "superieur", # If B32 corresponds to higher education levels, set it to "superieur" c("1ère année", "2ème année", "Licence/3ième année", "Maitrise/4ième année", "DEA/5ième année", "6ième année", "7ième année", "8ième année et +")
    TRUE ~ as.character(B32) %>% 
  structure(label = Hmisc::label(individus_2013$B32)))) # Apply labels to the B32 column
```


### Education indicators computation

*Years of schooling indicator*

```{r}
edu_var = c("Primaire","Moyen","Secondaire","Superieur")
#educ_var = c("Aucun","Primaire","Moyen","Secondaire","Superieur")
# For the 2002 individus data  

menage_2002 <- 
  menage_2002 %>% 
  
# Calculate a binary variable 'school_attendance' for each household.
# If the sum of education levels is greater than 0, it means at least one individual in the household is attending school (0), 
# otherwise, it means no one is attending school (1).
  mutate(years_schooling_dim = rowSums(select(., .dots=all_of(edu_var)),na.rm = T),
    years_schooling_dim = ifelse(years_schooling_dim > 0, 0, 1) %>% structure(label = "Education: years of schooling"))
  
# For the 2013 individus data  

menage_2013 <- 
  menage_2013 %>% 
  
# Calculate a binary variable 'school_attendance' for each household.
# If the sum of education levels is greater than 0, it means at least one individual in the household is attending school (0), 
# otherwise, it means no one is attending school (1).
  mutate(years_schooling_dim = rowSums(select(., .dots=all_of(edu_var)),na.rm = T),
    years_schooling_dim = ifelse(years_schooling_dim > 0, 0, 1) %>% structure(label = "Education: years of schooling"))
```

*School Attendance indicator*

```{r}

# Processing for the year 2002
menage_2002 <- 
  menage_2002 %>% 
  left_join(
    # Joining the data with summarized statistics per household
    individus_2002 %>%
    group_by(ID_MENAGE) %>%
    summarise(
      # Calculating the number of children (lien_cm=="children" and age < 13) with no school attendance 
      
      school_attendance = sum(educ == 0 & age5 > 13 & QB03 == 3, na.rm = TRUE),
      
      #The total number of (lien_cm=="children") in the household
      children_hh = sum(QB03 == 3,na.rm = TRUE)
    ) %>%
    ungroup() %>%
      
    # Computing a dimension for school attendance in the household
    mutate(
      school_attendance_dim = ifelse(school_attendance == children_hh, 1, 0)%>% structure(label = "Education: school attendance")
    ) %>%
      
    # Selecting relevant columns for the join
    select(ID_MENAGE, school_attendance_dim),
    by = c("ID_MENAGE")
  )

# Processing for the year 2013
menage_2013 <- 
  menage_2013 %>% 
  left_join(
    # Joining the data with summarized statistics per household
    individus_2013 %>%
    group_by(ID_MENAGE) %>%
    summarise(
      
      # Calculating the number of children (lien_cm=="children" and age < 13) with no school attendance  (NA for B32)
      school_attendance = sum(is.na(B32) & B08 > 13 & B04 == 3, na.rm = TRUE),
      
      # Computing a dimension for school attendance in the household
      children_hh = sum(B04 == 3,na.rm = TRUE)
    ) %>%
    ungroup() %>%
      
    # Computing a dimension for school attendance in the household
    mutate(
      school_attendance_dim = ifelse(school_attendance == children_hh, 1, 0) %>% structure(label = "Education: school attendance")
    ) %>%
      
    # Selecting relevant columns for the join
    select(ID_MENAGE, school_attendance_dim),
    by = c("ID_MENAGE")
  )
```


## Standard of Living dimension

### Cleaning the "Standard of Living" variables

```{r}

# Renaming variables so that they can have the same name in the two
# datasets

# Renaming variables so that they can have the same name in the two datasets

## For the RGPH 2002
menage_2002 <- menage_2002 %>% 
  
  # Create and rename variables in the "menage_2002" dataset
  mutate(
       # Create a new variable "nb_indiv" with labels from the "menage_2002" dataset
    #nb_indiv = Nb_indiv %>% structure(label=Hmisc::label(menage_2002$Nb_indiv)),
    
    # Create a new variable "id_dr" with labels from the "menage_2002" dataset
    id_dr = DR %>% structure(label=Hmisc::label(menage_2002$DR)),
    
    # Create a new variable "type_log" as a factor with labels from the "menage_2002" dataset
    type_log = E01 %>% structure(label=Hmisc::label(menage_2002$E01)),
    
    # Create a new variable "piece_occup" as a factor with labels from the "menage_2002" dataset
    piece_occup = E02 %>% structure(label=Hmisc::label(menage_2002$E02)),
    
    # Create a new variable "stat_occupation" as a factor with labels from the "menage_2002" dataset
    stat_occupation = E03 %>% structure(label=Hmisc::label(menage_2002$E03)),
    
    # Create a new variable "type_aisance" as a factor with labels from the "menage_2002" dataset
    type_aisance = E04 %>% structure(label=Hmisc::label(menage_2002$E04)),
    
    # Create a new variable "approv_eau" as a factor with labels from the "menage_2002" dataset
    approv_eau = E05 %>% structure(label=Hmisc::label(menage_2002$E05)),
    
    # Create a new variable "eclairage" as a factor with labels from the "menage_2002" dataset
    eclairage = E06 %>% structure(label=Hmisc::label(menage_2002$E06)),
    
    # Create a new variable "combustible_cuisine" as a factor with labels from the "menage_2002" dataset
    combustible_cuisine = E07 %>% structure(label=Hmisc::label(menage_2002$E07)),
    
    # Create new variables "nature_mur," "nature_toit," and "nature_sol" as factors with labels from the "menage_2002" dataset
    nature_mur = E08 %>% structure(label=Hmisc::label(menage_2002$E08)),
    nature_toit = E09 %>% structure(label=Hmisc::label(menage_2002$E09)),
    nature_sol = E10 %>% structure(label=Hmisc::label(menage_2002$E10)),
    
    # Create new variables "radio," "televiseur," "video," and "refri_conge" as factors with labels from the "menage_2002" dataset
    radio = E11B %>% structure(label=Hmisc::label(menage_2002$E11B)),
    televiseur = E11C %>% structure(label=Hmisc::label(menage_2002$E11C)),
    video = E11D %>% structure(label=Hmisc::label(menage_2002$E11D)),
    refri_conge = E11E %>% structure(label=Hmisc::label(menage_2002$E11E)),
    
    # Create new variables "telephone," "rechaud_cuis," "foyer_ameli," and "climatisateur" as factors with labels from the "menage_2002" dataset
    telephone = E11F %>% structure(label=Hmisc::label(menage_2002$E11F)),
    rechaud_cuis = E11G %>% structure(label=Hmisc::label(menage_2002$E11G)),
    foyer_ameli = E11H %>% structure(label=Hmisc::label(menage_2002$E11H)),
    climatisateur = E11I %>% structure(label=Hmisc::label(menage_2002$E11I)),
    
    # Create new variables "machine_coudre," "voiture," "mobylette," "bicyclette," "caleche_charette," and "pirogue" as factors with labels from the "menage_2002" dataset
    machine_coudre = E11J %>% structure(label=Hmisc::label(menage_2002$E11J)),
    voiture = E12B %>% structure(label=Hmisc::label(menage_2002$E12B)),
    mobylette = E12C %>% structure(label=Hmisc::label(menage_2002$E12C)),
    bicyclette = E12E %>% structure(label=Hmisc::label(menage_2002$E12E)),
    caleche_charette = E12F %>% structure(label=Hmisc::label(menage_2002$E12F)),
    pirogue = E12G %>% structure(label=Hmisc::label(menage_2002$E12G)),
    
    
# Create a new variable "houe_charrue" as a factor with labels from the "menage_2002" dataset
houe_charrue = E13B %>% structure(label=Hmisc::label(menage_2002$E13B)),

# Create a new variable "caleche_char" as a factor with labels from the "menage_2002" dataset
caleche_char = E13C %>% structure(label=Hmisc::label(menage_2002$E13C)),

# Create a new variable "animaux_traite" as a factor with labels from the "menage_2002" dataset
animaux_traite = E13D %>% structure(label=Hmisc::label(menage_2002$E13D)),

# Create a new variable "tracteur" as a factor with labels from the "menage_2002" dataset
tracteur = E13E %>% structure(label=Hmisc::label(menage_2002$E13E)),

# Create a new variable "voiture_cam" as a factor with labels from the "menage_2002" dataset
voiture_cam = E13F %>% structure(label=Hmisc::label(menage_2002$E13F)),

# Create a new variable "mobylette_bicycl" as a factor with labels from the "menage_2002" dataset
mobylette_bicycl = E13G %>% structure(label=Hmisc::label(menage_2002$E13G)),

# Create a new variable "refriger_congelat" as a factor with labels from the "menage_2002" dataset
refriger_congelat = E13I %>% structure(label=Hmisc::label(menage_2002$E13I)),

# Create a new variable "machine_coudre" as a factor with labels from the "menage_2002" dataset
machine_coudre = E13J %>% structure(label=Hmisc::label(menage_2002$E13J)),

# Create a new variable "materiel_musique" as a factor with labels from the "menage_2002" dataset
materiel_musique = E13K %>% structure(label=Hmisc::label(menage_2002$E13K)),

# Create a new variable "chaise_bache" as a factor with labels from the "menage_2002" dataset
chaise_bache = E13L %>% structure(label=Hmisc::label(menage_2002$E13L)),

# Create a new variable "telephone_tele" as a factor with labels from the "menage_2002" dataset
telephone_tele = E13M %>% structure(label=Hmisc::label(menage_2002$E13M)),

# Create a new variable "photocopieuse" as a factor with labels from the "menage_2002" dataset
photocopieuse = E13N %>% structure(label=Hmisc::label(menage_2002$E13N)),

# Create a new variable "ordinateur" as a factor with labels from the "menage_2002" dataset
ordinateur = E13O %>% structure(label=Hmisc::label(menage_2002$E13O)),

# Create a new variable "moulin" as a factor with labels from the "menage_2002" dataset
moulin = E13P %>% structure(label=Hmisc::label(menage_2002$E13P)),

# Create a new variable "appareil_photo" as a factor with labels from the "menage_2002" dataset
appareil_photo = E13Q %>% structure(label=Hmisc::label(menage_2002$E13Q)),

# Create a new variable "terrain_bat" as a factor with labels from the "menage_2002" dataset
terrain_bat = E13R %>% structure(label=Hmisc::label(menage_2002$E13R)),

# Create a new variable "evacut_ordures" as a factor with labels from the "menage_2002" dataset
evacut_ordures = E14 %>% structure(label=Hmisc::label(menage_2002$E14)),

# Create a new variable "evacut_eaux_usees" as a factor with labels from the "menage_2002" dataset
evacut_eaux_usees = E15 %>% structure(label=Hmisc::label(menage_2002$E15)),

# Create a new variable "repas_saute" as a factor with labels from the "menage_2002" dataset
repas_saute = E16 %>% structure(label=Hmisc::label(menage_2002$E16)),

# Create a new variable "soins_medicaux" as a factor with labels from the "menage_2002" dataset
soins_medicaux = E17 %>% structure(label=Hmisc::label(menage_2002$E17))
  )


# Mutate and rename variables in the "menage_2013" dataset
menage_2013 <- menage_2013 %>% 
  
  mutate(
# Create a new variable "id_dr" as a factor with labels from the "menage_2013" dataset
id_dr = IDDR %>% structure(label=Hmisc::label(menage_2013$IDDR)),

# Create a new variable "nb_indiv" as a factor with labels from the "menage_2013" dataset
#nb_indiv = Nbre_indiv %>% structure(label=Hmisc::label(menage_2013$Nbre_indiv)),

# Create a new variable "type_log" as a factor with labels from the "menage_2013" dataset
type_log = E01 %>% structure(label=Hmisc::label(menage_2013$E01)),

# Create a new variable "piece_occup" as a factor with labels from the "menage_2013" dataset
piece_occup = E02 %>% structure(label=Hmisc::label(menage_2013$E02)),

# Create a new variable "stat_occupation" as a factor with labels from the "menage_2013" dataset
stat_occupation = E03 %>% structure(label=Hmisc::label(menage_2013$E03)),

# Create a new variable "type_aisance" as a factor with labels from the "menage_2013" dataset
type_aisance = E08 %>% structure(label=Hmisc::label(menage_2013$E08)),

# Create a new variable "approv_eau" as a factor with labels from the "menage_2013" dataset
approv_eau = E09 %>% structure(label=Hmisc::label(menage_2013$E09)),

# Create a new variable "eclairage" as a factor with labels from the "menage_2013" dataset
eclairage = E11 %>% structure(label=Hmisc::label(menage_2013$E11)),

# Create a new variable "combustible_cuisine" as a factor with labels from the "menage_2013" dataset
combustible_cuisine = E12 %>% structure(label=Hmisc::label(menage_2013$E12)),

# Create a new variable "nature_mur" as a factor with labels from the "menage_2013" dataset
nature_mur = E05 %>% structure(label=Hmisc::label(menage_2013$E05)),

# Create a new variable "nature_toit" as a factor with labels from the "menage_2013" dataset
nature_toit = E06 %>% structure(label=Hmisc::label(menage_2013$E06)),

# Create a new variable "nature_sol" as a factor with labels from the "menage_2013" dataset
nature_sol = E07 %>% structure(label=Hmisc::label(menage_2013$E07)),

# Create a new variable "radio" as a factor with labels from the "menage_2013" dataset
radio = E13_1 %>% structure(label=Hmisc::label(menage_2013$E13_1)),

# Create a new variable "televiseur" as a factor with labels from the "menage_2013" dataset
televiseur = E13_2 %>% structure(label=Hmisc::label(menage_2013$E13_2)),

# Create a new variable "video" as a factor with labels from the "menage_2013" dataset
video = E13_3 %>% structure(label=Hmisc::label(menage_2013$E13_3)),

# Create a new variable "refri_conge" as a factor with labels from the "menage_2013" dataset
refri_conge = E13_4 %>% structure(label=Hmisc::label(menage_2013$E13_4)),

# Create a new variable "telephone" as a factor with labels from the "menage_2013" dataset
telephone = E13_6 %>% structure(label=Hmisc::label(menage_2013$E13_6)),

# Create a new variable "rechaud_cuis" as a factor with labels from the "menage_2013" dataset
rechaud_cuis = E13_18 %>% structure(label=Hmisc::label(menage_2013$E13_18)),

# Create a new variable "foyer_ameli" as a factor with labels from the "menage_2013" dataset
foyer_ameli = E13_7 %>% structure(label=Hmisc::label(menage_2013$E13_7)),

# Create a new variable "climatisateur" as a factor with labels from the "menage_2013" dataset
climatisateur = E13_8 %>% structure(label=Hmisc::label(menage_2013$E13_8)),

# Create a new variable "machine_coudre" as a factor with labels from the "menage_2013" dataset
machine_coudre = E13_9 %>% structure(label=Hmisc::label(menage_2013$E13_9)),
    # Create a new variable "voiture" as a factor with labels from the "menage_2013" dataset
voiture = E15_5 %>% structure(label=Hmisc::label(menage_2013$E15_5)),

# Create a new variable "mobylette" as a factor with labels from the "menage_2013" dataset
mobylette = E15_6 %>% structure(label=Hmisc::label(menage_2013$E15_6)),

# Create a new variable "bicyclette" as a factor with labels from the "menage_2013" dataset
bicyclette = E14_3 %>% structure(label=Hmisc::label(menage_2013$E14_3)),

# Create a new variable "caleche_charette" as a factor with labels from the "menage_2013" dataset
caleche_charette = E14_4 %>% structure(label=Hmisc::label(menage_2013$E14_4)),

# Create a new variable "pirogue" as a factor with labels from the "menage_2013" dataset
pirogue = E15_7 %>% structure(label=Hmisc::label(menage_2013$E15_7)),

# Create a new variable "houe_charrue" as a factor with labels from the "menage_2013" dataset
houe_charrue = E15_1 %>% structure(label=Hmisc::label(menage_2013$E15_1)),

# Create a new variable "caleche_char" as a factor with labels from the "menage_2013" dataset
caleche_char = E15_2 %>% structure(label=Hmisc::label(menage_2013$E15_2)),

# Create a new variable "animaux_traite" as a factor with labels from the "menage_2013" dataset
animaux_traite = E15_3 %>% structure(label=Hmisc::label(menage_2013$E15_3)),

# Create a new variable "tracteur" as a factor with labels from the "menage_2013" dataset
tracteur = E15_4 %>% structure(label=Hmisc::label(menage_2013$E15_4)),

# Create a new variable "voiture_cam" as a factor with labels from the "menage_2013" dataset
voiture_cam = E15_5 %>% structure(label=Hmisc::label(menage_2013$E15_5)),

# Create a new variable "mobylette_bicycl" as a factor with labels from the "menage_2013" dataset
mobylette_bicycl = E15_6 %>% structure(label=Hmisc::label(menage_2013$E15_6)),

# Create a new variable "refriger_congelat" as a factor with labels from the "menage_2013" dataset
refriger_congelat = E15_8 %>% structure(label=Hmisc::label(menage_2013$E15_8)),

# Create a new variable "machine_coudre" as a factor with labels from the "menage_2013" dataset
machine_coudre = E15_9 %>% structure(label=Hmisc::label(menage_2013$E15_9)),

# Create a new variable "materiel_musique" as a factor with labels from the "menage_2013" dataset
materiel_musique = E15_10 %>% structure(label=Hmisc::label(menage_2013$E15_10)),

# Create a new variable "chaise_bache" as a factor with labels from the "menage_2013" dataset
chaise_bache = E15_11 %>% structure(label=Hmisc::label(menage_2013$E15_11)),

# Create a new variable "telephone_tele" as a factor with labels from the "menage_2013" dataset
telephone_tele = E15_12 %>% structure(label=Hmisc::label(menage_2013$E15_12)),

# Create a new variable "photocopieuse" as a factor with labels from the "menage_2013" dataset
photocopieuse = E15_13 %>% structure(label=Hmisc::label(menage_2013$E15_13)),

# Create a new variable "ordinateur" as a factor with labels from the "menage_2013" dataset
ordinateur = E15_14 %>% structure(label=Hmisc::label(menage_2013$E15_14)),

# Create a new variable "moulin" as a factor with labels from the "menage_2013" dataset
moulin = E15_15 %>% structure(label=Hmisc::label(menage_2013$E15_15)),

# Create a new variable "appareil_photo" as a factor with labels from the "menage_2013" dataset
appareil_photo = E15_16 %>% structure(label=Hmisc::label(menage_2013$E15_16)),

# Create a new variable "terrain_bat" as a factor with labels from the "menage_2013" dataset
terrain_bat = E15_17 %>% structure(label=Hmisc::label(menage_2013$E15_17)),

# Create a new variable "evacut_ordures" as a factor with labels from the "menage_2013" dataset
evacut_ordures = E16 %>% structure(label=Hmisc::label(menage_2013$E16)),

# Create a new variable "evacut_eaux_usees" as a factor with labels from the "menage_2013" dataset
evacut_eaux_usees = E17 %>% structure(label=Hmisc::label(menage_2013$E17)),

# Create a new variable "repas_saute" as a factor with labels from the "menage_2013" dataset
repas_saute = E18 %>% structure(label=Hmisc::label(menage_2013$E18)),

# Create a new variable "soins_medicaux" as a factor with labels from the "menage_2013" dataset
soins_medicaux = E20 %>% structure(label=Hmisc::label(menage_2013$E20))
    )



```


```{r}
# Creating dummy variables for categorical columns in the 'menage_2002' dataframe
menage_2002 <- menage_2002 %>% 
  fastDummies::dummy_columns(select_columns = c("repas_saute", "evacut_eaux_usees", "nature_sol", "nature_toit", "nature_mur", "combustible_cuisine", "eclairage", "approv_eau", "type_aisance", "stat_occupation", "type_log")) %>% 
  mutate(
    
    # Répas sauté
    repas_saute_1 = as.numeric(repas_saute_1) %>% structure(label=paste0(Hmisc::label(menage_2002$repas_saute), ": Oui")),
         
         repas_saute_2 = as.numeric(repas_saute_2) %>% structure(label=paste0(Hmisc::label(menage_2002$repas_saute), ": Non")),
         
    # Evacuation des eaux usées
         evacut_eaux_usees_1 = as.numeric(evacut_eaux_usees_1) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Réseau d'égout")), 
                  
         evacut_eaux_usees_2 = as.numeric(evacut_eaux_usees_2) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Canal fermé")),
                  evacut_eaux_usees_3 = as.numeric(evacut_eaux_usees_3) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Canal ciel ouvert")),
         
                  evacut_eaux_usees_4 = as.numeric(evacut_eaux_usees_4) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Grille bouche avaloir")), 
                  
         evacut_eaux_usees_5 = as.numeric(evacut_eaux_usees_5) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Dans la mer/fleuve")),
         
                  evacut_eaux_usees_6 = as.numeric(evacut_eaux_usees_6) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Cours d'eau")),
         
          evacut_eaux_usees_7 = as.numeric(evacut_eaux_usees_7) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Trou creusé")),
         
                  evacut_eaux_usees_8 = as.numeric(evacut_eaux_usees_8) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Dans la rue/nature")), 
                  
         evacut_eaux_usees_9 = as.numeric(evacut_eaux_usees_9) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Autre")),
         
    
    # Nature du sol
         nature_sol_1 = as.numeric(nature_sol_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Carrelage")),
         
         nature_sol_2 = as.numeric(nature_sol_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Ciment")),
         
         nature_sol_3 = as.numeric(nature_sol_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Argile/banco")),
         
         nature_sol_4 = as.numeric(nature_sol_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Sable")),
         
         nature_sol_5 = as.numeric(nature_sol_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Autre")),
         
        nature_toit_1 = as.numeric(nature_toit_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Beton")),
        
    
    # Nature du toit
                nature_toit_2 = as.numeric(nature_toit_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Tuile/ardoise")),
        
                nature_toit_3 = as.numeric(nature_toit_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Zinc")),
        
                nature_toit_4 = as.numeric(nature_toit_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Chaume")),
        
                nature_toit_5 = as.numeric(nature_toit_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Autre")),
        
        
    # Nature du mur 
        nature_mur_1 = as.numeric(nature_mur_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Ciment")),
         
                nature_mur_2 = as.numeric(nature_mur_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Bois")),
        
                nature_mur_3 = as.numeric(nature_mur_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Banco")),
        
                nature_mur_4 = as.numeric(nature_mur_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Paille/tige")),
        
                nature_mur_5 = as.numeric(nature_mur_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Autre")),
        
        
    # Combustible de cuisine
                  combustible_cuisine_1 = as.numeric(combustible_cuisine_1) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Bois")),
                           combustible_cuisine_2 = as.numeric(combustible_cuisine_2) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Charbon")),
                          combustible_cuisine_3 = as.numeric(combustible_cuisine_3) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Gaz")),
                          combustible_cuisine_4 = as.numeric(combustible_cuisine_4) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Electricite")),
                          combustible_cuisine_5 = as.numeric(combustible_cuisine_5) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Autre")),
        
    # éclairage
                  eclairage_1 = as.numeric(eclairage_1) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Electricite")),
        
                          eclairage_2 = as.numeric(eclairage_2) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": olaire")),
                          eclairage_3 = as.numeric(eclairage_3) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Grpe.electrogene")),
                          eclairage_4 = as.numeric(eclairage_4) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Gaz")),
        
                          eclairage_5 = as.numeric(eclairage_5) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Lampe a petrole")),
                          eclairage_6 = as.numeric(eclairage_6) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Lampe tempete")),
                          eclairage_7 = as.numeric(eclairage_7) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Bougie")),
                          eclairage_8 = as.numeric(eclairage_8) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Bois")),
        
                          eclairage_9 = as.numeric(eclairage_9) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Autre")),
         
    
    # Approvisionnement en eau
                  approv_eau_1 = as.numeric(approv_eau_1) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Puits interieur")),
        
                          approv_eau_2 = as.numeric(approv_eau_2) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Puits exterieur")),
        
                          approv_eau_3 = as.numeric(approv_eau_3) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Robinet inter.")),
        
                          approv_eau_4 = as.numeric(approv_eau_4) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Robinet public")),
        
                          approv_eau_5 = as.numeric(approv_eau_5) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Forage")),
        
                          approv_eau_6 = as.numeric(approv_eau_6) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Vendeurs d'eau")),
        
                          approv_eau_7 = as.numeric(approv_eau_7) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Source, cours d'eau")),
        
                          approv_eau_8 = as.numeric(approv_eau_8) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Autre")),
    
    # Type d'aisance
         
                  type_aisance_1 = as.numeric(type_aisance_1) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Chasse egouts")),
        
                          type_aisance_2 = as.numeric(type_aisance_2) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Chasse fosse")),
        
                          type_aisance_3 = as.numeric(type_aisance_3) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Latrines")),
        
                          type_aisance_4 = as.numeric(type_aisance_4) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Edicule public")),
        
                          type_aisance_5 = as.numeric(type_aisance_5) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Dans la nature")),
        
                          type_aisance_6 = as.numeric(type_aisance_6) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Autre")),
        
    # Statut d'occupation
                  stat_occupation_1 = as.numeric(stat_occupation_1) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Proprietaire")),
                          stat_occupation_2 = as.numeric(stat_occupation_2) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Co_proprietaire")),
                          stat_occupation_3 = as.numeric(stat_occupation_3) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Locataire")),
                          stat_occupation_4 = as.numeric(stat_occupation_4) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Loge' employeur")),
                          stat_occupation_5 = as.numeric(stat_occupation_5) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Loge' famille")),
                          stat_occupation_6 = as.numeric(stat_occupation_6) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Autre")),
         
    # Type de logement
                  type_log_1 = as.numeric(type_log_1) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Case")),
         
                          type_log_2 = as.numeric(type_log_2) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Baraque")),
        
                          type_log_3 = as.numeric(type_log_3) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Maison basse")),
        
                          type_log_4 = as.numeric(type_log_4) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Maison à étage")),
        
                          type_log_5 = as.numeric(type_log_5) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Appartement")),
        
                          type_log_6 = as.numeric(type_log_6) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Autre"))
         )


```


```{r}
# Creating additional dummy variables for remaining categorical columns in the 'menage_2013' dataframe
menage_2013 <- menage_2013 %>% 
  fastDummies::dummy_columns(select_columns = c("evacut_eaux_usees", "nature_sol", "nature_toit", "nature_mur", "combustible_cuisine", "eclairage", "approv_eau", "type_aisance", "stat_occupation", "type_log"))%>% 
  mutate(
    
    # Evacuation des eaux usées
         evacut_eaux_usees_1 = as.numeric(evacut_eaux_usees_1) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Réseau d'égout")), 
                  
         evacut_eaux_usees_2 = as.numeric(evacut_eaux_usees_2) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Canal fermé")),
                  evacut_eaux_usees_3 = as.numeric(evacut_eaux_usees_3) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Canal à ciel ouvert")),
         
                  evacut_eaux_usees_4 = as.numeric(evacut_eaux_usees_4) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Grille, bouche avaloir")), 
                  
         evacut_eaux_usees_5 = as.numeric(evacut_eaux_usees_5) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Fosse/puisard")),
         
                  evacut_eaux_usees_6 = as.numeric(evacut_eaux_usees_6) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Dans la mer, fleuve")),
         
          evacut_eaux_usees_7 = as.numeric(evacut_eaux_usees_7) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Cours d'eau")),
         
                  evacut_eaux_usees_8 = as.numeric(evacut_eaux_usees_8) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Trou creusé")), 
                  
         evacut_eaux_usees_9 = as.numeric(evacut_eaux_usees_9) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Dans la rue/nature")),
         
             evacut_eaux_usees_98 = as.numeric(evacut_eaux_usees_98) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Autre")),
         
    # Nature du sol
         nature_sol_1 = as.numeric(nature_sol_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Carrelage")),
         
         nature_sol_2 = as.numeric(nature_sol_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Ciment")),
         
         nature_sol_3 = as.numeric(nature_sol_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Argile/banco")),
         
         nature_sol_4 = as.numeric(nature_sol_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Sable")),
         
         nature_sol_5 = as.numeric(nature_sol_5) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Tapis")),
         
             nature_sol_6 = as.numeric(nature_sol_6) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Moquette")),
    
             nature_sol_7 = as.numeric(nature_sol_7) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Bois ciré")),
    
             nature_sol_8 = as.numeric(nature_sol_8) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Autre")),
    
    # Nature du toit
        nature_toit_1 = as.numeric(nature_toit_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Béton")),
        
                nature_toit_2 = as.numeric(nature_toit_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Tuile/ardoise")),
        
                nature_toit_3 = as.numeric(nature_toit_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Zinc")),
        
                nature_toit_4 = as.numeric(nature_toit_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Chaume/paille")),
        
                nature_toit_0 = as.numeric(nature_toit_0) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Autre")),
        
        
    # Nature du mur 
        nature_mur_1 = as.numeric(nature_mur_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment")),
         
                nature_mur_2 = as.numeric(nature_mur_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment avec carreaux")),
        
                nature_mur_3 = as.numeric(nature_mur_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment et marbre")),
        
                nature_mur_4 = as.numeric(nature_mur_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment avec bois")),
        
                nature_mur_5 = as.numeric(nature_mur_5) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Bois")),
        
                nature_mur_6 = as.numeric( nature_mur_6) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Banco")),
        
                nature_mur_7 = as.numeric(nature_mur_7) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Banco+ enduit ciment")),
        
                # nature_mur_8 = nature_mur_8) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Paille/Tige")),
        
                nature_mur_9 = as.numeric(nature_mur_9) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Autre")),
        
    # Combustible de cuisine
                  combustible_cuisine_1 = as.numeric(combustible_cuisine_1) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Bois")),
                  combustible_cuisine_2 = as.numeric(combustible_cuisine_2) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Charbon")),
                  combustible_cuisine_3 = as.numeric(combustible_cuisine_3) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Gaz")),
                          combustible_cuisine_4 = as.numeric(combustible_cuisine_4) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Electricite")),
                          combustible_cuisine_5 = as.numeric(combustible_cuisine_5) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Bouse de vache")),
                          combustible_cuisine_8 = as.numeric(combustible_cuisine_8) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Autre")),
        
    # éclairage
                  eclairage_1 = as.numeric(eclairage_1) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Electricite")),
                          eclairage_2 = as.numeric(eclairage_2) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": olaire")),
                          eclairage_3 = as.numeric(eclairage_3) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Grpe.electrogene")),
                          eclairage_4 = as.numeric(eclairage_4) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Gaz")),
        
                          eclairage_5 = as.numeric(eclairage_5) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe a petrole")),
                          eclairage_6 = as.numeric(eclairage_6) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe tempête")),
                          eclairage_7 = as.numeric(eclairage_7) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Bougie")),
                          eclairage_8 = as.numeric(eclairage_8) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Bois")),
        
                          eclairage_9 = as.numeric(eclairage_9) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe rechargeable")),
                              eclairage_98 = as.numeric(eclairage_98) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Autre")),
    
    # Approvisionnement en eau
                  approv_eau_1 = as.numeric(approv_eau_1) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet dans logement")),
        
                          approv_eau_2 = as.numeric(approv_eau_2) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet dans cour/parcelle")),
        
                          approv_eau_3 = as.numeric(approv_eau_3) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet public/borne fontaine")),
        
                          approv_eau_4 = as.numeric(approv_eau_4) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits à pompe ou forage")),
        
                          approv_eau_5 = as.numeric(approv_eau_5) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits protégés")),
        
                          approv_eau_6 = as.numeric(approv_eau_6) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits non protégés")),
        
                          approv_eau_7 = as.numeric(approv_eau_7) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Vendeurs d’eau")),
        
                          approv_eau_8 = as.numeric(approv_eau_8) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Camion citerne")),
    
                          approv_eau_9 = as.numeric(approv_eau_9) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Eau de pluie")),
        
                          approv_eau_10 = as.numeric(approv_eau_10) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Source, cour d'eau")),
        
                          approv_eau_98 = as.numeric(approv_eau_98) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Autre")),
    
    
    # Type d'aisance
                  type_aisance_11 = as.numeric(type_aisance_11) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Chasse raccordée à l’égout")),
        
                          type_aisance_12 = as.numeric(type_aisance_12) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Chasse avec fosse")),
        
                          type_aisance_21 = as.numeric(type_aisance_21) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Couverte")),
        
                          type_aisance_22 = as.numeric(type_aisance_22) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Non couverte")),
        
                          type_aisance_23 = as.numeric(type_aisance_23) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Ventilée améliorée")),
        
                          type_aisance_31 = as.numeric(type_aisance_31) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Edicule public")),
    
                              type_aisance_41 = as.numeric(type_aisance_41) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Dans la nature")),
        
                          type_aisance_88 = as.numeric(type_aisance_88) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Autre")),
        
    # Statut d'occupation
                  stat_occupation_1 = as.numeric(stat_occupation_1) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Proprietaire")),
                          stat_occupation_2 = as.numeric(stat_occupation_2) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Co proprietaire")),
                          stat_occupation_3 = as.numeric(stat_occupation_3) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Locataire")),
                          stat_occupation_4 = as.numeric(stat_occupation_4) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Co locataire")),
                          stat_occupation_5 = as.numeric(stat_occupation_5) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Location vente")),
                          stat_occupation_6 = as.numeric(stat_occupation_6) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Logé par employeur")),
                          stat_occupation_7 = as.numeric(stat_occupation_7) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Logé par les parents/amis")),
                          stat_occupation_8 = as.numeric(stat_occupation_8) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Autre")),
         
    # Type de logement
                  type_log_1 = as.numeric(type_log_1) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Case")),
         
                          type_log_2 = as.numeric(type_log_2) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Baraque")),
        
                          type_log_3 = as.numeric(type_log_3) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Maison basse")),
        
                          type_log_4 = as.numeric(type_log_4) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Maison à étage")),
        
                          type_log_5 = as.numeric(type_log_5) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Immeuble à Appartement")),
        
                          type_log_0 = as.numeric(type_log_0) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Autre"))
         )



```


### Standard of living indicators computation

**Data source**: 

* *For the Sanitation*: [link](https://www.who.int/data/gho/data/indicators/indicator-details/GHO/population-using-improved-sanitation-facilities-(-))

* *For water*:
[link](https://data.unicef.org/topic/water-and-sanitation/drinking-water/)
 
```{r}

# For RGPH 2002
menage_2002 <- menage_2002 %>% 
  # Creating a dimension for cooking fuel, checking if it's wood or charcoal
  # Bois=1, Charbon = 2 and Autre=5
  mutate(cooking_fuel_dim = ifelse(combustible_cuisine == 1 | combustible_cuisine == 2| combustible_cuisine == 5, 1, 0) %>% structure(label = "Standard of living: cooking fuel"), 
         
  # Creating a dimension for sanitation, checking for specific types (public edifice, nature, other)
  # Edicule public==4, Dans la nature==4, Autre==6
         sanitation_dim = ifelse(type_aisance %in% seq(4, 6), 1, 0) %>% structure(label = "Standard of living: sanitation"),
         
  # Creating a dimension for drinking water, considering specific sources (indoor well, outdoor well, watercourse, other)
  # Puits interieur = 1, Puits exterieur = 2, cours d'eau = 7 et autre = 8
         drinking_water_dim = ifelse(approv_eau %in% c(1, 2, 7, 8), 1, 0) %>% structure(label = "Standard of living: drinking water"),
         
  # Creating a dimension for electricity, considering different sources (storm lamp, candle, wood, other)
  # Lampe tempête = 6, Bougie = 7, Bois=8, Autre= 9,
         electricity_dim = ifelse(eclairage %in% c(6, 7, 8, 9), 1, 0)%>% structure(label = "Standard of living: electricity"),
         
  # Creating a dimension for housing quality, checking specific types of construction materials for walls, floor, and roof
        # For nature_mur, 3 is "Banco", 4 is "Paille/tige" and 5 is 'autre".
         # For nature_sol, 3 is 'Argile/banco", 4 is Sable and 5 is "autre".
         # For nature_toit, 4 is "Chaume" and 5 is "autre"
         housing_dim = ifelse(nature_mur %in% c(3, 4, 5) & nature_sol %in% c(3, 4, 5) & nature_toit %in% c(4, 5), 1, 0)%>% structure(label = "Standard of living: housing"),
         
  # Creating a dimension for household assets, checking if various assets are absent in the household
  
         assets_dim = ifelse(radio == 0 & televiseur == 0 & telephone == 0 & animaux_traite == 0 & bicyclette == 0 & mobylette == 0 & mobylette_bicycl == 0 & 
                           refri_conge == 0 & refriger_congelat == 0 & voiture == 0 & voiture_cam == 0, 1, 0) %>% structure(label = "Standard of living: assets")
  )



# For RGPH 2013
menage_2013 <- menage_2013 %>% 
  # Creating a dimension for cooking fuel, checking if it's wood or charcoal
  # Bois=1, Charbon = 2, Bouse de vache=5 and Autre=8
  mutate(cooking_fuel_dim = ifelse(combustible_cuisine %in% c(1,2,5,8), 1, 0)%>% structure(label = "Standard of living: cooking fuel"), 
         
  # Creating a dimension for sanitation, checking for specific types (public edifice, nature, other)
  # Non couverte =22, Edicule public==31, Dans la nature==41, Autre==88
         sanitation_dim = ifelse(type_aisance %in% c(22, 31, 41, 88), 1, 0)%>% structure(label = "Standard of living: sanitation"),
         
  # Creating a dimension for drinking water, considering specific sources (indoor well, outdoor well, watercourse, other)
  # Puits non protégés = 6, Eau de pluie = 9, Source, cour d'eau = 10 and autre = 98
         drinking_water_dim = ifelse(approv_eau %in% c(6, 9, 10, 98), 1, 0)%>% structure(label = "Standard of living: drinking water"),
         
  # Creating a dimension for electricity, considering different sources (storm lamp, candle, wood, other)
  # Lampe à pétrole artisanale=5, Lampe tempête = 6, Bougie = 7, Bois=8, Lampe rechargeable= 9, Autre=98
         electricity_dim = ifelse(eclairage %in% c(5, 6, 7, 8, 9, 98), 1, 0)%>% structure(label = "Standard of living: electricity"),
         
  # Creating a dimension for housing quality, checking specific types of construction materials for walls, floor, and roof
        # For nature_mur, 5 is Bois, 6 is "Banco", 7 is "Banco+ enduit ciment",  8 is 'Paille/Tige" and 9 is "Autre".
         # For nature_sol, 3 is 'Argile/banco", 4 is Sable and 7 is "Bois ciré" and 8 is "autre".
         # For nature_toit, 0 is "Autre" and 4 is "Chaume/paille"
         housing_dim = ifelse(nature_mur %in% c(5, 6, 7, 8, 9) & nature_sol %in% c(3, 4, 7, 8) & nature_toit %in% c(0, 4), 1, 0)%>% structure(label = "Standard of living: housing"),
         
  # Creating a dimension for household assets, checking if various assets are absent in the household
  
         assets_dim = ifelse(radio == 0 & televiseur == 0 & telephone == 0 & telephone_tele == 0 & animaux_traite == 0 & bicyclette == 0 & mobylette == 0 & mobylette_bicycl == 0 & 
                           refri_conge == 0 & refriger_congelat == 0 & voiture == 0 & voiture_cam == 0, 1, 0)%>% structure(label = "Standard of living: assets")
  )



```




*Additionnal cleaning*

```{r}

# For the RGPH 2013 dataset, recode the "soins_medicaux" variable
# This code modifies the "soins_medicaux" variable, replacing "2" with "0" if present

menage_2013 <- menage_2013 %>%
  mutate(soins_medicaux = as.numeric(soins_medicaux),
            soins_medicaux = 
                dplyr::recode(soins_medicaux,"2" = 0),
            soins_medicaux = soins_medicaux %>% structure(label=Hmisc::label(menage_2002$E17))) 

# For the RGPH 2002 dataset, recode the "soins_medicaux" variable
# This code modifies the "soins_medicaux" variable, replacing "2" with "0", "3" with NA, and "9" with NA
menage_2002 <- menage_2002 %>%
  mutate(soins_medicaux = as.numeric(soins_medicaux),
           soins_medicaux = dplyr::recode(soins_medicaux, "2" = 0, "3" = NULL, "9" = NULL), 
         soins_medicaux = soins_medicaux %>% structure(label=Hmisc::label(menage_2002$E17)))

```

## Health dimension

### Cleaning of the "Health" variables and indicators computation

```{r wrangling deces}

# For the RGPH 2002 dataset, determine if there was a death in the household in the last twelve months
# This code calculates whether there was a death in the household based on specific columns
deces_2002 <- deces_2002 %>% 
  mutate(
    est_decedee = ifelse(is.na(C03) & is.na(C04) & is.na(C05) & is.na(GROUPE_AGE_DECES), 0, 1) %>% structure(label = "y-a-t-il eu décès dans le ménage au cours des douze derniers mois?")
  ) %>% 
  group_by(ID_MENAGE) %>% 
  summarise(
    
    # Count the number of death in the household
    deces = sum(est_decedee, na.rm = TRUE),
    
    # Count the number of death in the household below the age of 5
    deces_under_five = sum(est_decedee==1 & C04 <=5,
                                       na.rm = TRUE)) %>% 
  ungroup() %>% 
  # Create a variable for an under 18 death in the household
  mutate(deces_under_five_dim = as.factor(ifelse(deces_under_five > 0, 1, ifelse(is.na(deces_under_five), 0, 0)) %>%
                                            structure(label = "Y a-t-il eu un décés d'un enfant de moins de cinq ans dans le ménage ?")),
         deces_under_five = deces_under_five %>%
                                            structure(label = "Nombre de décés de moins de cinq ans")) 

# For the RGPH 2013 dataset, calculate the number of deaths in each household
# This code groups the data by household and calculates the number of deaths in each household
deces_2013 <- deces_2013 %>% 
  group_by(ID_MENAGE) %>% 
  summarise(deces = sum(C01, na.rm = TRUE)
    
    # Count the number of death in the household below the age of 18
  ) %>% 
  ungroup() %>% 
  # Create the variable for the number of death in the household
  mutate(deces = as.numeric(deces) %>%
           structure(label = "Nombre de décés dans le ménage")) 

```



```{r}
# For the RGPH 2013 dataset, select specific columns and merge with "deces_2013" data
# This code selects the "id_dr," "ID_MENAGE," and columns 77 to the last column in "menage_2013"
# It then performs a left join with "deces_2013" based on the "ID_MENAGE" key

# Selecting specific columns from the 'menage_2013' dataset
menage_2013 <- menage_2013 %>% 
  select(id_dr, ID_MENAGE, everything()) %>%
  
# Performing a left join with the 'deces_2013' dataset based on 'ID_MENAGE'
  left_join(deces_2013, by = "ID_MENAGE") %>%
  
# Adding a new column 'RGPH' with the value "2013" and setting a label
  mutate(RGPH = "2013" %>% structure(label = "RGPH"),
  
# Adding a label to the 'deces' column
         deces = deces %>% structure(label = "Nombre de décès dans le ménage au cours des 12 derniers mois"))

# For the RGPH 2002 dataset, select specific columns and merge with "deces_2013" data
# This code selects the "id_dr," "ID_MENAGE," and columns 71 to the last column in "menage_2002"
# It then performs a left join with "deces_2013" based on the "ID_MENAGE" key

# Selecting specific columns from the 'menage_2002' dataset
menage_2002 <- menage_2002 %>% 
  select(id_dr, ID_MENAGE, num_men, everything()) %>%
  
# Performing a left join with the 'deces_2002' dataset based on 'ID_MENAGE'
  left_join(deces_2002, by = "ID_MENAGE") %>%
  
# Adding a new column 'RGPH' with the value "2002" and setting a label
  mutate(RGPH = "2002" %>% structure(label = "RGPH"),
  
# Adding a new column 'ID_MENAGE_corrige' by checking if 'num_men' is missing and using 'ID_MENAGE' if so, then setting a label
         ID_MENAGE_corrige = ifelse(is.na(num_men), ID_MENAGE, num_men) %>% structure(label = "ID MENAGE")) %>%
  
# Selecting and removing specific columns
  select(-c(ID_MENAGE, num_men)) %>%
  
# Renaming the 'ID_MENAGE_corrige' column to 'ID_MENAGE'
  rename(ID_MENAGE = ID_MENAGE_corrige) %>%
  
# Adding a new column 'deces' by checking if it's missing and setting it to 0, with a label
  mutate(deces = ifelse(is.na(deces), 0, deces) %>% structure(label = "Nombre de décès dans le ménage au cours des 12 derniers mois"),
         
         # Replace Na by 0
         
         deces_under_five = ifelse(is.na(deces_under_five), 0, deces_under_five),
         
         deces_under_five_dim = ifelse(is.na(deces_under_five_dim), 0, deces_under_five_dim) %>% structure(label = "Health: child mortality"))

```





```{r}
# Filtering the 'menage_2013' dataset to exclude specific values in the 'id_dr' column
menage_2013 <- menage_2013 %>% 
  filter(!(id_dr %in% c("014301140124", "014301140125")))



```


# Exporting the datasets


```{r}
## For RGPH 2013
individus_2002 <- individus_2002 %>%
 select(id_dr,ID_MENAGE, RGPH, everything())# 

## For RGPH 2002
individus_2013 <- individus_2013 %>%
 select(id_dr,ID_MENAGE, RGPH, everything())# 
# Saving in SPSS format
haven::write_sav(individus_2002,paste0(here::here(), "/output/output_data/individus_2002.sav"))
# Saving in SPSS format
haven::write_sav(individus_2013,paste0(here::here(), "/output/output_data/individus_2013.sav"))
```


```{r, results='hide'}
# Exporting the "menage_2013" and "menage_2002" datasets

# For the "menage_2002"
# Saving in SPSS format
haven::write_sav(menage_2002,paste0(here::here(), "/output/output_data/menage_2002.sav"))

# Saving in R format
saveRDS(menage_2002,file=paste0(here::here(), "/output/output_data/menage_2002.rds"))

# For the "menage_2013"
# Saving in SPSS format
haven::write_sav(menage_2013,paste0(here::here(), "/output/output_data/menage_2013.sav"))

# Saving in R format
saveRDS(menage_2013,file=paste0(here::here(), "/output/output_data/menage_2013.rds"))


# For the "sp_rgph_2002"
# Saving in the sf format
sf::st_write(obj=sp_rgph_2002, paste0(here::here(), "/output/output_data/EAs_2002_new.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) # we change the id_dr

# For the "sp_rgph_2013"
sf::st_write(obj=sp_rgph_2013, paste0(here::here(), "/output/output_data/EAs_2013.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) # we change the id_dr


```

```{r}
## Removing all objects
rm(list=ls())
```
