==> script/02_Exploring-Senegal-Census-data.Rmd <==
---
title: "Exploring Senegal Census data with visualization"
author: "IBRAHIM KASSOUM Habibou & HEMA Aboubacar"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, error=FALSE)
```


## Project Sénégal

### Preambule: 

```{r}
rm(list=ls())
```


```{r Package needed, results = "hide"}


## Importing library
### List of required packages
required_packages <- c("tidyverse","janitor" ,"readr","dplyr","haven","sf", "flextable","sp", "factoextra", "FactoMineR","gtsummary", "sjPlot", "fastDummies")

# Check if packages are installed
missing_packages <- setdiff(required_packages, installed.packages()[,"Package"])

### Install missing packages
if (length(missing_packages) > 0) {
  install.packages(missing_packages)
}

### Load all packages
lapply(required_packages, library, character.only = TRUE)

```

### Import data

```{r}

# Read shapefile data for 2002 and 2013
sp_rgph_2002 <- sf::read_sf(paste0(here::here(),"/output/output_data/EAs_2002_new.shp"))

sp_rgph_2013 <- sf::read_sf(paste0(here::here(),"/data/SHAPEFILE/EAs_2013.shp"))

# Display the first 10 rows of the 2002 and 2013 shapefiles
head(sp_rgph_2002, 10L)
head(sp_rgph_2013, 10L)
```


```{r}

# Read and check for duplicated rows in the "menage_2002" dataset
menage_2002 <- haven::read_sav(paste0(here::here(),"/output/output_data/menage_2002.sav"))
menage_2002 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "menage_2013" dataset
menage_2013 <- haven::read_sav(paste0(here::here(),"/output/output_data/menage_2013.sav"))
menage_2013 %>% janitor::get_dupes() 
```


```{r}
par(mfrow=c(1,2))
plot(sp_rgph_2002$geometry, main="Sénégal 2002")
plot(sp_rgph_2013$geometry, main="Sénégal 2013")
```

```{r}

indicateur_menage_2002 <- menage_2002 %>% 
  group_by(RGPH, id_dr, sexe_cm) %>% 
  summarise(n=n(),
            age_moyen=mean(age_cm, na.rm = TRUE))


indicateur_menage_2013 <- menage_2013 %>% 
  group_by(RGPH, id_dr, sexe_cm) %>% 
  summarise(n=n(),
            age_moyen=mean(age_cm, na.rm = TRUE))
```


```{r}

indicateur_menage_2002 <- sp_rgph_2002 %>% 
  left_join(indicateur_menage_2002, by="id_dr") %>% 
  select(names(indicateur_menage_2002), geometry) %>% 
  filter(!is.na(sexe_cm)) %>% 
  mutate(
    sexe_cm=as.numeric(sexe_cm),
    sexe_cm = recode(sexe_cm, 
                         "1"="Homme",
                         "2"="Femme")) %>% 
  filter(!is.na(sexe_cm)) 
# %>% 
#   mutate(sexe_cm= recode(sexe_cm, 
#                          "1"="Homme",
#                          "2"="Femme"))

# indicateur_menage_2002$sexe_cm <- car::recode(indicateur_menage_2002$sexe_cm, "1 = Homme; 2=Femme")


# indicateur_menage_2013 <- sp_rgph_2013 %>% 
#   left_join(indicateur_menage_2013, by="id_dr") 

g_2002 <- ggplot(indicateur_menage_2002, aes(fill = age_moyen)) + 
          geom_sf() + 
          scale_fill_viridis_c() + 
          theme_void() + 
          facet_wrap(~sexe_cm, ncol = 2) +
          labs(fill = "",
               title = "",
               subtitle = "",
               caption = "2002 Senegal Census")



indicateur_menage_2013 <- sp_rgph_2013 %>% 
  rename(id_dr=COD_DR2012) %>% 
  left_join(indicateur_menage_2013, by="id_dr") %>% 
  select(names(indicateur_menage_2013), geometry) %>% 
  filter(!is.na(sexe_cm)) %>% 
  mutate(
    sexe_cm=as.numeric(sexe_cm),
    sexe_cm= recode(sexe_cm, 
                         "1"="Homme",
                         "2"="Femme")) %>% 
  filter(!is.na(sexe_cm)) 
# %>%
#   mutate(sexe_cm= recode(sexe_cm,
#                          "1"="Homme",
#                          "2"="Femme"))
# indicateur_menage_2013$sexe_cm <- dplyr::recode(indicateur_menage_2013$sexe_cm, "1"="Homme","2"="Femme")

# indicateur_menage_2013 <- sp_rgph_2013 %>% 
#   left_join(indicateur_menage_2013, by="id_dr") 

g_2013 <- ggplot(indicateur_menage_2013, aes(fill = age_moyen)) + 
          geom_sf() + 
          scale_fill_viridis_c() + 
          theme_void() + 
          facet_wrap(~sexe_cm, ncol = 2) +
          labs(fill = "",
               title = "",
               subtitle = "",
               caption = "2013 Senegal Census")


gridExtra::grid.arrange(g_2002,g_2013, nrow=2, top="Age moyen du CM par DR")


```

```{r}
ggplot(indicateur_menage_2013, aes(fill = age_moyen)) + 
          geom_sf() + 
          scale_fill_viridis_c() + 
          theme_void() + 
          facet_wrap(~sexe_cm, ncol = 2) +
          labs(fill = "",
               title = "",
               subtitle = "",
               caption = "2013 Senegal Census")
```


```{r eval=FALSE, include=FALSE}
menage_2002 <- sp_rgph_2002 %>% 
  left_join(menage_2002, by="id_dr")

menage_2013 <- sp_rgph_2013 %>% 
  left_join(menage_2013, by="id_dr")
```
