==> script/02_Tabulation-Senegal-Census-data (by DR).Rmd <==
---
title: "Tabulation of Covariates and Outcomes for the Senegal census (by DR)"
author: "Habibou Ibrahim"
date: '`r format (Sys.Date(), "%d %B %Y")`'
output: 
  officedown::rdocx_document:
    mapstyles:
      Normal: ['First Paragraph']
---

```{r setup, include = TRUE, warning=FALSE, echo = FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
# library(officedown)
# library(officer)

# fp <- fp_par(
#   text.align = "center", 
#   padding.bottom = 20, padding.top = 120, 
#   border.bottom = fp_border())

#ft <- fp_text(shading.color='#EFEFEF', bold = F)
```

# Table of content

<!---BLOCK_TOC--->

\newpage


```{r, include = TRUE, warning=FALSE, echo = FALSE}
# Clear the current workspace by removing all objects.
# This is useful to start with a clean slate and avoid potential conflicts or unexpected behavior.
rm(list=ls())
```

## Importing library
```{r Package needed, results = "hide", include = TRUE, warning=FALSE, echo = FALSE}


## Importing library
### List of required packages
required_packages <- c("tidyverse","janitor" ,"readr","dplyr","haven","sf",
                       "flextable","sp", "factoextra", "FactoMineR","gtsummary", "sjPlot")

# Check if packages are installed
missing_packages <- setdiff(required_packages, installed.packages()[,"Package"])

### Install missing packages
if (length(missing_packages) > 0) {
  install.packages(missing_packages)
}

### Load all packages
lapply(required_packages, library, character.only = TRUE)

```

```{r, include = TRUE, warning=FALSE, echo = FALSE}
# We set the theme to compact to ensure table fit the landscape page
# This theme will apply to all the other tables in this Rmd
set_gtsummary_theme(theme_gtsummary_compact())
```


## Importing the datasets an checking for duplicated row

### Importing the shapefile data

```{r, results='hide', include = TRUE, warning=FALSE, echo = FALSE}
# Read shapefile data for 2002 and 2013
sp_rgph_2002 <- sf::read_sf(paste0(here::here(),"/output/output_data/EAs_2002_new.shp"))
# Mutating the 'sp_rgph_2002' dataset to create a new 'id_dr' column based on conditions

sp_rgph_2013 <- sf::read_sf(paste0(here::here(),"/output/output_data/EAs_2013.shp"))

# Display the first 10 rows of the 2002 and 2013 shapefiles
head(sp_rgph_2002, 10L)
head(sp_rgph_2013, 10L)
```

### Importing the census datasets 

**For the 2002 census data**

```{r, results='hide', include = TRUE, warning=FALSE, echo = FALSE}

# Read and check for duplicated rows in the "menage_2002" dataset
menage_2002 <- haven::read_sav(paste0(here::here(),"/output/output_data/menage_2002.sav"))
menage_2002 %>% janitor::get_dupes()
menage_2002 %>% janitor::get_dupes(ID_MENAGE)

```


**For the 2013 census dataset**

```{r, include = FALSE, warning=FALSE, echo = FALSE}

# Read and check for duplicated rows in the "menage_2013" dataset
menage_2013 <- haven::read_sav(paste0(here::here(),"/output/output_data/menage_2013.sav"))
menage_2013 %>% janitor::get_dupes() 
menage_2013 %>% janitor::get_dupes(ID_MENAGE)
```

**Grouping by IDDR**

```{r, results='hide', include = TRUE, warning=FALSE, echo = FALSE}

# For the RGPH 2002
dr_2002 <- menage_2002 %>% 
    group_by(id_dr) %>% 
    summarise_if(is.numeric, sum, na.rm = TRUE) %>% 
    ungroup()%>% 
    mutate(RGPH="2002")


# For the RGPH 2013
dr_2013 <- menage_2013 %>% 
    group_by(id_dr) %>% 
    summarise_if(is.numeric, sum, na.rm = TRUE)%>% 
    ungroup() %>% 
    mutate(RGPH="2013")
```



# Study Variables

## Socio-demographic variables

### Sex/gender

```{r, results='hold', include = TRUE, warning=FALSE, echo = FALSE}

gender_var = c("homme","femme","hh_size")
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,homme,femme,hh_size) %>% 
  
  plyr::rbind.fill(
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
        # Selecting relevant variables
      select(RGPH, homme,femme,hh_size)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    type = c(gender_var~"continuous"),
    label = list(
      homme ~ "Hommes",
      femme ~ "Femmes",
      hh_size ~ "Taille du ménage"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Composition des ménages") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 
```




### Age
```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For babies under one year old, write “0” years. For people aged 98 and over, the instruction was to write 98 for age.

age_var = c("nbr_babies","nbr_under_14","nbr_young_people","nbr_adult", "nbr_elder_adult", "nbr_senior")

dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,nbr_babies,nbr_under_14,nbr_young_people,nbr_adult, nbr_elder_adult, nbr_senior) %>% 
  
  plyr::rbind.fill(
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
       # Selecting relevant variables
      select(RGPH,nbr_babies,nbr_under_14,nbr_young_people,nbr_adult, nbr_elder_adult, nbr_senior)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column

    type = c(age_var~"continuous"),
        
    label = list(
      nbr_babies ~ "Nombre d'enfant agée de moins d'un an",
      nbr_under_14 ~ "Nombre d'enfant agée de moins de 14 ans",
      nbr_young_people ~ "Nombre d'individus agée entre 15 et 29 ans",
      nbr_adult ~ "Nombre d'individus agées entre 30 et 44 ans",
      nbr_elder_adult ~ "Nombre d'individus agées entre 45 et 59 ans",
      nbr_senior ~ "Nombre d'individus agée de plus de 60 ans"
      ),
    
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Composition des ménages par age") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```



### Marital Status

```{r, include = TRUE, warning=FALSE, echo = FALSE}

mat_status_var = c("Monogame","Polygame","Celibataire","Veuf","Divorce")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,Monogame,Polygame,Celibataire,Veuf,Divorce) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,Monogame,Polygame,Celibataire,Veuf,Divorce)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(
      Monogame ~ "Nombre de monogame dans le ménage",
      Polygame ~ "Nombre de polygame dans le ménage",
      Celibataire ~ "Nombre de celibataire dans le ménage",
      Veuf ~ "Nombre de veuf/veuve dans le ménage",
      Divorce ~ "Nombre de divorcé dans le ménage"
      ),
    
    type = c(mat_status_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Situation matrimoniale") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```

### Fertility


```{r, include = TRUE, warning=FALSE, echo = FALSE}
#Changer les labels
fertility_var = c("fertility")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,fertility) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,fertility)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column

    type = c(fertility_var~"continuous"),
    label = list(
      fertility ~ "Fertilité"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Table de fertilité") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```

### Educational level

```{r, include = TRUE, warning=FALSE, echo = FALSE}

educ_var = c("Aucun","Primaire","Moyen","Secondaire","Superieur")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,Aucun,Primaire,Moyen,Secondaire,Superieur) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,Aucun,Primaire,Moyen,Secondaire,Superieur)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    label = list(Aucun ~ "Aucun",
      Primaire ~ "Primaire",
      Moyen ~ "Moyen",
      Secondaire ~ "Sécondaire",
      Superieur ~ "Supérieur"),
    type = c(educ_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Niveau d'instruction") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```


### Literacy level

```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For the RGPH 2002 dataset
literacy_var = c("Literacy_French","Literacy_Arabic","Literacy_Wolof","Literacy_Pulaar","Literacy_Sereer","Literacy_Mandingo","Literacy_Diola","Literacy_Soninke")



#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,Literacy_French,Literacy_Arabic,Literacy_Wolof,Literacy_Pulaar,Literacy_Sereer,Literacy_Mandingo,Literacy_Diola,Literacy_Soninke) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,Literacy_French,Literacy_Arabic,Literacy_Wolof,Literacy_Pulaar,Literacy_Sereer,Literacy_Mandingo,Literacy_Diola,Literacy_Soninke)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(Literacy_French ~ "Français",
      Literacy_Arabic ~ "Arabe",
      Literacy_Wolof ~ "Wolof",
      Literacy_Pulaar ~ "Pulaar",
      Literacy_Sereer ~ "Sereer",
      Literacy_Mandingo ~ "Mandingo",
      Literacy_Diola ~ "Diola",
      Literacy_Soninke ~ "Sononké"
      ),
    
    type = c(literacy_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Alphabétisation dans le ménage") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```
### Migration variable

```{r, include = TRUE, warning=FALSE, echo = FALSE}

mig_var = c("resident_pres","resident_abs","resident_vis")


#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,resident_pres,resident_abs,resident_vis) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,resident_pres,resident_abs,resident_vis)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(resident_pres ~ "Présent",
      resident_abs ~ "Absent",
      resident_vis ~ "Visiteur"
      ),
    
    type = c(mig_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Résidence dans le ménage") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 



```

### Maternal death variable

```{r, include = TRUE, warning=FALSE, echo = FALSE}

deces_mat_var = c("deces_mat_oui","deces_mat_non","deces_mat_nsp")

# Pas collecter and 2013
#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,deces_mat_oui,deces_mat_non,deces_mat_nsp) %>% 
  

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
   # by = RGPH,  # Group by "RGPH" column

    type = c(deces_mat_var~"continuous"),
    
    label = list(deces_mat_oui ~ "Oui",
      deces_mat_non ~ "Non",
      deces_mat_nsp ~ "NSP"
      ),
    
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Déces maternelle dans le ménage") %>%
  #add_overall() %>%
  add_n()
  # Add difference statistics to the table
  #add_difference() %>% 


```
## Socio-economic variables


### Employed variable

```{r, include = TRUE, warning=FALSE, echo = FALSE}

occup_var=c("occup_independant","occup_employeur","occup_salarie","occup_apprenti","occup_aideFamilial","occup_autre")   


#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,occup_independant,occup_employeur,occup_salarie,occup_apprenti,occup_aideFamilial,occup_autre) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,occup_independant,occup_employeur,occup_salarie,occup_apprenti,occup_aideFamilial,occup_autre)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(occup_independant ~ "Indépendant",
      occup_employeur ~ "Employeur",
      occup_salarie ~ "Salarié",
      occup_apprenti ~ "Apprenti",
      occup_aideFamilial ~ "Aide familial",
      occup_autre ~ "Autre"
      ),
    
    type = c(occup_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Occupation des membres du ménage") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 



```

### Occupational status variable

```{r , include = TRUE, warning=FALSE, echo = FALSE}

sit_occup_var=c("sit_occup_occup","sit_occup_chomeur","sit_occup_foyer","sit_occup_etudiant","sit_occup_rentier","sit_occup_retraite","sit_occup_autre")  


#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,sit_occup_occup,sit_occup_chomeur,sit_occup_foyer,sit_occup_etudiant,sit_occup_rentier,sit_occup_retraite,sit_occup_autre) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,sit_occup_occup,sit_occup_chomeur,sit_occup_foyer,sit_occup_etudiant,sit_occup_rentier,sit_occup_retraite,sit_occup_autre)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column

    type = c(sit_occup_var~"continuous"),
    
    label = list(sit_occup_occup ~ "Occupé",
      sit_occup_chomeur ~ "Chomeur",
      sit_occup_foyer ~ "Foyer",
      sit_occup_etudiant ~ "Étudiant",
      sit_occup_rentier ~ "Rentier",
      sit_occup_retraite ~ "Retraité",
      sit_occup_autre ~ "Autre"
      ),
    
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Situation dans l'occupation principale des membres du ménage") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```

### Income variable


```{r, include = TRUE, warning=FALSE, echo = FALSE}

income_var = c("I_repas_saute","I_soins_medicaux")

# Il faut revoir les labels dans cette region

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,I_repas_saute,I_soins_medicaux) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,I_repas_saute,I_soins_medicaux)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(
      I_repas_saute ~ "Repas sauté",
      I_soins_medicaux ~ "Soins médicaux"
      ),
    
    type = c(income_var~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Sauté") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 


```

## Household variables

### Capital goods variable

```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For the RGPH 2002 dataset
#OK
# 1. Radio, 2. Television, 3. Video/VCD/DVD, 4. Refrigerator/freezer, 5. Telephone, 6. Improved fireplace, 7. Air conditioner, 8. Sewing machine, 9. Water heater, 10. Stove

#For the RGPH 2013 dataset



```

### Access to basic equipment

#### For the RGPH 2002 dataset 
```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For the RGPH 2002 dataset
#OK
# 1. Main mode of household waste disposal, 2. Main mode of waste water disposal, 3. Main mode of lighting, 4. Main fuel for cooking
#For the RGPH 2013 dataset
# 
# equipment_var_cat <- c(paste0("repas_saute_", seq(1,2,1)), paste0("evacut_eaux_usees_",seq(1,9,1)),paste0("nature_sol_",seq(1,5,1)),paste0("nature_toit_",seq(1,5,1)),paste0("nature_mur_",seq(1,5,1)),paste0("combustible_cuisine_",seq(1,5,1)),paste0("eclairage_",seq(1,9,1)),paste0("approv_eau_",seq(1,8,1)),paste0("type_aisance_",seq(1,6,1)),paste0("stat_occupation_",seq(1,6,1)),paste0("type_log_",seq(1,6,1)))

# 
# equipment_var_dicho <- c("soins_medicaux","terrain_bat","appareil_photo","moulin","ordinateur","photocopieuse","telephone_tele","chaise_bache","materiel_musique","machine_coudre","refriger_congelat","mobylette_bicycl","voiture_cam","tracteur","animaux_traite","caleche_char","houe_charrue","pirogue","caleche_charette","bicyclette","mobylette","voiture","machine_coudre","climatisateur","foyer_ameli","rechaud_cuis","telephone","refri_conge","video","televiseur","radio")

equipment_var_num <- c("piece_occup","soins_medicaux","terrain_bat","appareil_photo","moulin","ordinateur","photocopieuse","telephone_tele","chaise_bache","materiel_musique","machine_coudre","refriger_congelat","mobylette_bicycl","voiture_cam","tracteur","animaux_traite","caleche_char","houe_charrue","pirogue","caleche_charette","bicyclette","mobylette","voiture","machine_coudre","climatisateur","foyer_ameli","rechaud_cuis","telephone","refri_conge","video","televiseur","radio", paste0("repas_saute_", seq(1,2,1)), paste0("evacut_eaux_usees_",seq(1,9,1)),paste0("nature_sol_",seq(1,5,1)),paste0("nature_toit_",seq(1,5,1)),paste0("nature_mur_",seq(1,5,1)),paste0("combustible_cuisine_",seq(1,5,1)),paste0("eclairage_",seq(1,9,1)),paste0("approv_eau_",seq(1,8,1)),paste0("type_aisance_",seq(1,6,1)),paste0("stat_occupation_",seq(1,6,1)),paste0("type_log_",seq(1,6,1)))

#,""

  # select(soins_medicaux,repas_saute,evacut_eaux_usees,terrain_bat,appareil_photo,moulin,ordinateur,photocopieuse,telephone_tele,chaise_bache,materiel_musique,machine_coudre,refriger_congelat,mobylette_bicycl,voiture_cam,tracteur,animaux_traite,caleche_char,houe_charrue,pirogue,caleche_charette,bicyclette,mobylette,voiture,machine_coudre,climatisateur,foyer_ameli,rechaud_cuis,telephone,refri_conge,video,televiseur,radio,nature_sol,nature_toit,nature_mur,combustible_cuisine,eclairage,approv_eau,type_aisance,stat_occupation,piece_occup,type_log)

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(soins_medicaux,terrain_bat,appareil_photo,moulin,ordinateur,photocopieuse,telephone_tele,chaise_bache,materiel_musique,machine_coudre,refriger_congelat,mobylette_bicycl,voiture_cam,tracteur,animaux_traite,caleche_char,houe_charrue,pirogue,caleche_charette,bicyclette,mobylette,voiture,machine_coudre,climatisateur,foyer_ameli,rechaud_cuis,telephone,refri_conge,video,televiseur,radio,piece_occup,starts_with("repas_saute_"),starts_with("evacut_eaux_usees_"),starts_with("nature_sol_"),starts_with("nature_toit_"),starts_with("nature_mur_"),starts_with("combustible_cuisine_"),starts_with("eclairage_"),starts_with("approv_eau_"),starts_with("type_aisance"),starts_with("stat_occupation_"),starts_with("type_log_")) %>% 
  
  mutate(

    soins_medicaux = labelled::remove_labels(soins_medicaux) %>% structure(label=Hmisc::label(menage_2002$soins_medicaux)),
    
    # repas_saute = labelled::remove_labels(repas_saute) %>% structure(label=Hmisc::label(menage_2002$repas_saute)),
    
    terrain_bat = labelled::remove_labels(terrain_bat) %>% structure(label=Hmisc::label(menage_2002$terrain_bat)),
    
    appareil_photo = labelled::remove_labels(appareil_photo) %>% structure(label=Hmisc::label(menage_2002$appareil_photo)),
    
    moulin = labelled::remove_labels(moulin) %>% structure(label=Hmisc::label(menage_2002$moulin)),
    
    ordinateur = labelled::remove_labels(ordinateur) %>% structure(label=Hmisc::label(menage_2002$ordinateur)),
    
    photocopieuse = labelled::remove_labels(photocopieuse) %>% structure(label=Hmisc::label(menage_2002$photocopieuse)),
    
    telephone_tele = labelled::remove_labels(telephone_tele) %>% structure(label=Hmisc::label(menage_2002$telephone_tele)),
    
    chaise_bache = labelled::remove_labels(chaise_bache) %>% structure(label=Hmisc::label(menage_2002$chaise_bache)),
    
    materiel_musique = labelled::remove_labels(materiel_musique) %>% structure(label=Hmisc::label(menage_2002$materiel_musique)),
    
    machine_coudre = labelled::remove_labels(machine_coudre) %>% structure(label=Hmisc::label(menage_2002$machine_coudre)),
    
    refriger_congelat = labelled::remove_labels(refriger_congelat) %>% structure(label=Hmisc::label(menage_2002$refriger_congelat)),

    mobylette_bicycl = labelled::remove_labels(mobylette_bicycl) %>% structure(label=Hmisc::label(menage_2002$mobylette_bicycl)),
    
    voiture_cam = labelled::remove_labels(voiture_cam) %>% structure(label=Hmisc::label(menage_2002$voiture_cam)),
    
    tracteur = labelled::remove_labels(tracteur) %>% structure(label=Hmisc::label(menage_2002$tracteur)),
    
    animaux_traite = labelled::remove_labels(animaux_traite) %>% structure(label=Hmisc::label(menage_2002$animaux_traite)),
    
    caleche_char = labelled::remove_labels(caleche_char) %>% structure(label=Hmisc::label(menage_2002$caleche_char)),
    
    houe_charrue = labelled::remove_labels(houe_charrue) %>% structure(label=Hmisc::label(menage_2002$houe_charrue)),
    
    pirogue = labelled::remove_labels(pirogue) %>% structure(label=Hmisc::label(menage_2002$pirogue)),
    
    caleche_charette = labelled::remove_labels(caleche_charette) %>% structure(label=Hmisc::label(menage_2002$caleche_charette)),

    bicyclette = labelled::remove_labels(bicyclette) %>% structure(label=Hmisc::label(menage_2002$bicyclette)),
    
    mobylette = labelled::remove_labels(mobylette) %>% structure(label=Hmisc::label(menage_2002$mobylette)),
    
    voiture = labelled::remove_labels(voiture) %>% structure(label=Hmisc::label(menage_2002$voiture)),
    
    machine_coudre = labelled::remove_labels(machine_coudre) %>% structure(label=Hmisc::label(menage_2002$machine_coudre)),

    climatisateur = labelled::remove_labels(climatisateur) %>% structure(label=Hmisc::label(menage_2002$climatisateur)),
    
    foyer_ameli = labelled::remove_labels(foyer_ameli) %>% structure(label=Hmisc::label(menage_2002$foyer_ameli)),
    
    rechaud_cuis = labelled::remove_labels(rechaud_cuis) %>% structure(label=Hmisc::label(menage_2002$rechaud_cuis)),
    
    telephone = labelled::remove_labels(telephone) %>% structure(label=Hmisc::label(menage_2002$telephone)),

    refri_conge = labelled::remove_labels(refri_conge) %>% structure(label=Hmisc::label(menage_2002$refri_conge)),
    
    video = labelled::remove_labels(video) %>% structure(label=Hmisc::label(menage_2002$video)),
    
    televiseur = labelled::remove_labels(televiseur) %>% structure(label=Hmisc::label(menage_2002$televiseur)),
    
    radio = labelled::remove_labels(radio) %>% structure(label=Hmisc::label(menage_2002$radio)),
    
    piece_occup = labelled::remove_labels(piece_occup) %>% structure(label=Hmisc::label(menage_2002$piece_occup)),
    
    # Répas sauté
    repas_saute_1 = as.numeric(repas_saute_1) %>% structure(label=paste0(Hmisc::label(menage_2002$repas_saute), ": Oui")),
         
         repas_saute_2 = as.numeric(repas_saute_2) %>% structure(label=paste0(Hmisc::label(menage_2002$repas_saute), ": Non")),
         
    # Evacuation des eaux usées
         evacut_eaux_usees_1 = as.numeric(evacut_eaux_usees_1) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Réseau d'égout")), 
                  
         evacut_eaux_usees_2 = as.numeric(evacut_eaux_usees_2) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Canal fermé")),
                  evacut_eaux_usees_3 = as.numeric(evacut_eaux_usees_3) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Canal ciel ouvert")),
         
                  evacut_eaux_usees_4 = as.numeric(evacut_eaux_usees_4) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Grille bouche avaloir")), 
                  
         evacut_eaux_usees_5 = as.numeric(evacut_eaux_usees_5) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Dans la mer/fleuve")),
         
                  evacut_eaux_usees_6 = as.numeric(evacut_eaux_usees_6) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Cours d'eau")),
         
          evacut_eaux_usees_7 = as.numeric(evacut_eaux_usees_7) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Trou creusé")),
         
                  evacut_eaux_usees_8 = as.numeric(evacut_eaux_usees_8) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Dans la rue/nature")), 
                  
         evacut_eaux_usees_9 = as.numeric(evacut_eaux_usees_9) %>% structure(label=paste0(Hmisc::label(menage_2002$evacut_eaux_usees), ":  Autre")),
         
    
    # Nature du sol
         nature_sol_1 = as.numeric(nature_sol_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Carrelage")),
         
         nature_sol_2 = as.numeric(nature_sol_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Ciment")),
         
         nature_sol_3 = as.numeric(nature_sol_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Argile/banco")),
         
         nature_sol_4 = as.numeric(nature_sol_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Sable")),
         
         nature_sol_5 = as.numeric(nature_sol_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_sol), ": Autre")),
         
        nature_toit_1 = as.numeric(nature_toit_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Beton")),
        
    
    # Nature du toit
                nature_toit_2 = as.numeric(nature_toit_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Tuile/ardoise")),
        
                nature_toit_3 = as.numeric(nature_toit_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Zinc")),
        
                nature_toit_4 = as.numeric(nature_toit_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Chaume")),
        
                nature_toit_5 = as.numeric(nature_toit_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_toit), ": Autre")),
        
        
    # Nature du mur 
        nature_mur_1 = as.numeric(nature_mur_1) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Ciment")),
         
                nature_mur_2 = as.numeric(nature_mur_2) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Bois")),
        
                nature_mur_3 = as.numeric(nature_mur_3) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Banco")),
        
                nature_mur_4 = as.numeric(nature_mur_4) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Paille/tige")),
        
                nature_mur_5 = as.numeric(nature_mur_5) %>% structure(label=paste0(Hmisc::label(menage_2002$nature_mur), ": Autre")),
        
        
    # Combustible de cuisine
                  combustible_cuisine_1 = as.numeric(combustible_cuisine_1) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Bois")),
                           combustible_cuisine_2 = as.numeric(combustible_cuisine_2) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Charbon")),
                          combustible_cuisine_3 = as.numeric(combustible_cuisine_3) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Gaz")),
                          combustible_cuisine_4 = as.numeric(combustible_cuisine_4) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Electricite")),
                          combustible_cuisine_5 = as.numeric(combustible_cuisine_5) %>% structure(label=paste0(Hmisc::label(menage_2002$combustible_cuisine), ": Autre")),
        
    # éclairage
                  eclairage_1 = as.numeric(eclairage_1) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Electricite")),
        
                          eclairage_2 = as.numeric(eclairage_2) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": olaire")),
                          eclairage_3 = as.numeric(eclairage_3) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Grpe.electrogene")),
                          eclairage_4 = as.numeric(eclairage_4) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Gaz")),
        
                          eclairage_5 = as.numeric(eclairage_5) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Lampe a petrole")),
                          eclairage_6 = as.numeric(eclairage_6) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Lampe tempete")),
                          eclairage_7 = as.numeric(eclairage_7) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Bougie")),
                          eclairage_8 = as.numeric(eclairage_8) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Bois")),
        
                          eclairage_9 = as.numeric(eclairage_9) %>% structure(label=paste0(Hmisc::label(menage_2002$eclairage), ": Autre")),
         
    
    # Approvisionnement en eau
                  approv_eau_1 = as.numeric(approv_eau_1) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Puits interieur")),
        
                          approv_eau_2 = as.numeric(approv_eau_2) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Puits exterieur")),
        
                          approv_eau_3 = as.numeric(approv_eau_3) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Robinet inter.")),
        
                          approv_eau_4 = as.numeric(approv_eau_4) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Robinet public")),
        
                          approv_eau_5 = as.numeric(approv_eau_5) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Forage")),
        
                          approv_eau_6 = as.numeric(approv_eau_6) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Vendeurs d'eau")),
        
                          approv_eau_7 = as.numeric(approv_eau_7) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Source, cours d'eau")),
        
                          approv_eau_8 = as.numeric(approv_eau_8) %>% structure(label=paste0(Hmisc::label(menage_2002$approv_eau), ": Autre")),
    
    # Type d'aisance
         
                  type_aisance_1 = as.numeric(type_aisance_1) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Chasse egouts")),
        
                          type_aisance_2 = as.numeric(type_aisance_2) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Chasse fosse")),
        
                          type_aisance_3 = as.numeric(type_aisance_3) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Latrines")),
        
                          type_aisance_4 = as.numeric(type_aisance_4) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Edicule public")),
        
                          type_aisance_5 = as.numeric(type_aisance_5) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Dans la nature")),
        
                          type_aisance_6 = as.numeric(type_aisance_6) %>% structure(label=paste0(Hmisc::label(menage_2002$type_aisance), ": Autre")),
        
    # Statut d'occupation
                  stat_occupation_1 = as.numeric(stat_occupation_1) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Proprietaire")),
                          stat_occupation_2 = as.numeric(stat_occupation_2) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Co_proprietaire")),
                          stat_occupation_3 = as.numeric(stat_occupation_3) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Locataire")),
                          stat_occupation_4 = as.numeric(stat_occupation_4) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Loge' employeur")),
                          stat_occupation_5 = as.numeric(stat_occupation_5) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Loge' famille")),
                          stat_occupation_6 = as.numeric(stat_occupation_6) %>% structure(label=paste0(Hmisc::label(menage_2002$stat_occupation), ": Autre")),
         
    # Type de logement
                  type_log_1 = as.numeric(type_log_1) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Case")),
         
                          type_log_2 = as.numeric(type_log_2) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Baraque")),
        
                          type_log_3 = as.numeric(type_log_3) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Maison basse")),
        
                          type_log_4 = as.numeric(type_log_4) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Maison à étage")),
        
                          type_log_5 = as.numeric(type_log_5) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Appartement")),
        
                          type_log_6 = as.numeric(type_log_6) %>% structure(label=paste0(Hmisc::label(menage_2002$type_log), ": Autre"))

  ) %>%  
  
  labelled::unlabelled() %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(

    type = c(#equipment_var_cat~"continuous",
             equipment_var_num~"continuous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Access aux equipements for the RGPH 2002") %>%
  add_n()

```


#### For the RGPH 2013 dataset 
```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For the RGPH 2013 dataset
#OK
# 1. Main mode of household waste disposal, 2. Main mode of waste water disposal, 3. Main mode of lighting, 4. Main fuel for cooking
#For the RGPH 2013 dataset

# equipment_var_cat <- c("evacut_eaux_usees","terrain_bat","nature_sol","nature_toit","nature_mur","combustible_cuisine","eclairage","approv_eau","type_aisance","stat_occupation","type_log","terrain_bat")
# 
# equipment_var_cat <- c( paste0("evacut_eaux_usees_",seq(1,9,1)),"evacut_eaux_usees_98",paste0("nature_sol_",seq(1,8,1)),paste0("nature_toit_",seq(0,4,1)),paste0("nature_mur_",seq(1,7,1)),"nature_mur_9",paste0("combustible_cuisine_",seq(1,5,1)),"combustible_cuisine_8",paste0("eclairage_",seq(1,9,1)),"eclairage_98",paste0("approv_eau_",seq(1,10,1)),"approv_eau_98","type_aisance_11","type_aisance_12","type_aisance_21","type_aisance_22","type_aisance_23","type_aisance_31","type_aisance_41","type_aisance_88",paste0("stat_occupation_",seq(1,8,1)),paste0("type_log_",seq(0,5,1)))
# 



# 
# equipment_var_dicho <- c("soins_medicaux","appareil_photo","moulin","ordinateur","photocopieuse","telephone_tele","chaise_bache","materiel_musique","machine_coudre","refriger_congelat","mobylette_bicycl","voiture_cam","tracteur","animaux_traite","caleche_char","houe_charrue","pirogue","caleche_charette","bicyclette","mobylette","voiture","machine_coudre","climatisateur","foyer_ameli","rechaud_cuis","telephone","refri_conge","video","televiseur","radio","repas_saute","terrain_bat")

equipment_var_num <- c("repas_saute","piece_occup","soins_medicaux","appareil_photo","moulin","ordinateur","photocopieuse","telephone_tele","chaise_bache","materiel_musique","machine_coudre","refriger_congelat","mobylette_bicycl","voiture_cam","tracteur","animaux_traite","caleche_char","houe_charrue","pirogue","caleche_charette","bicyclette","mobylette","voiture","machine_coudre","climatisateur","foyer_ameli","rechaud_cuis","telephone","refri_conge","video","televiseur","radio","terrain_bat",paste0("evacut_eaux_usees_",seq(1,9,1)),"evacut_eaux_usees_98",paste0("nature_sol_",seq(1,8,1)),paste0("nature_toit_",seq(0,4,1)),paste0("nature_mur_",seq(1,7,1)),"nature_mur_9",paste0("combustible_cuisine_",seq(1,5,1)),"combustible_cuisine_8",paste0("eclairage_",seq(1,9,1)),"eclairage_98",paste0("approv_eau_",seq(1,10,1)),"approv_eau_98","type_aisance_11","type_aisance_12","type_aisance_21","type_aisance_22","type_aisance_23","type_aisance_31","type_aisance_41","type_aisance_88",paste0("stat_occupation_",seq(1,8,1)),paste0("type_log_",seq(0,5,1)))


#For the RGPH 2013 dataset 
dr_2013 %>% 
      
# Selecting relevant variables
select(soins_medicaux,repas_saute,terrain_bat,appareil_photo,moulin,ordinateur,photocopieuse,telephone_tele,chaise_bache,materiel_musique,machine_coudre,refriger_congelat,mobylette_bicycl,voiture_cam,tracteur,animaux_traite,caleche_char,houe_charrue,pirogue,caleche_charette,bicyclette,mobylette,voiture,machine_coudre,climatisateur,foyer_ameli,rechaud_cuis,telephone,refri_conge,video,televiseur,radio,piece_occup,starts_with("evacut_eaux_usees_"),starts_with("nature_sol_"),starts_with("nature_toit_"),starts_with("nature_mur_"),starts_with("combustible_cuisine_"),starts_with("eclairage_"),starts_with("approv_eau_"),starts_with("type_aisance_"),starts_with("stat_occupation_"),starts_with("type_log_")) %>% 
  
  mutate(

    soins_medicaux = labelled::remove_labels(soins_medicaux) %>% structure(label=Hmisc::label(menage_2013$soins_medicaux)),
    
     repas_saute = labelled::remove_labels(repas_saute) %>% structure(label=Hmisc::label(menage_2013$repas_saute)),
    
    terrain_bat = labelled::remove_labels(terrain_bat) %>% structure(label=Hmisc::label(menage_2013$terrain_bat)),
    
    appareil_photo = labelled::remove_labels(appareil_photo) %>% structure(label=Hmisc::label(menage_2013$appareil_photo)),
    
    moulin = labelled::remove_labels(moulin) %>% structure(label=Hmisc::label(menage_2013$moulin)),
    
    ordinateur = labelled::remove_labels(ordinateur) %>% structure(label=Hmisc::label(menage_2013$ordinateur)),
    
    photocopieuse = labelled::remove_labels(photocopieuse) %>% structure(label=Hmisc::label(menage_2013$photocopieuse)),
    
    telephone_tele = labelled::remove_labels(telephone_tele) %>% structure(label=Hmisc::label(menage_2013$telephone_tele)),
    
    chaise_bache = labelled::remove_labels(chaise_bache) %>% structure(label=Hmisc::label(menage_2013$chaise_bache)),
    
    materiel_musique = labelled::remove_labels(materiel_musique) %>% structure(label=Hmisc::label(menage_2013$materiel_musique)),
    
    machine_coudre = labelled::remove_labels(machine_coudre) %>% structure(label=Hmisc::label(menage_2013$machine_coudre)),
    
    refriger_congelat = labelled::remove_labels(refriger_congelat) %>% structure(label=Hmisc::label(menage_2013$refriger_congelat)),

    mobylette_bicycl = labelled::remove_labels(mobylette_bicycl) %>% structure(label=Hmisc::label(menage_2013$mobylette_bicycl)),
    
    voiture_cam = labelled::remove_labels(voiture_cam) %>% structure(label=Hmisc::label(menage_2013$voiture_cam)),
    
    tracteur = labelled::remove_labels(tracteur) %>% structure(label=Hmisc::label(menage_2013$tracteur)),
    
    animaux_traite = labelled::remove_labels(animaux_traite) %>% structure(label=Hmisc::label(menage_2013$animaux_traite)),
    
    caleche_char = labelled::remove_labels(caleche_char) %>% structure(label=Hmisc::label(menage_2013$caleche_char)),
    
    houe_charrue = labelled::remove_labels(houe_charrue) %>% structure(label=Hmisc::label(menage_2013$houe_charrue)),
    
    pirogue = labelled::remove_labels(pirogue) %>% structure(label=Hmisc::label(menage_2013$pirogue)),
    
    caleche_charette = labelled::remove_labels(caleche_charette) %>% structure(label=Hmisc::label(menage_2013$caleche_charette)),

    bicyclette = labelled::remove_labels(bicyclette) %>% structure(label=Hmisc::label(menage_2013$bicyclette)),
    
    mobylette = labelled::remove_labels(mobylette) %>% structure(label=Hmisc::label(menage_2013$mobylette)),
    
    voiture = labelled::remove_labels(voiture) %>% structure(label=Hmisc::label(menage_2013$voiture)),
    
    machine_coudre = labelled::remove_labels(machine_coudre) %>% structure(label=Hmisc::label(menage_2013$machine_coudre)),

    climatisateur = labelled::remove_labels(climatisateur) %>% structure(label=Hmisc::label(menage_2013$climatisateur)),
    
    foyer_ameli = labelled::remove_labels(foyer_ameli) %>% structure(label=Hmisc::label(menage_2013$foyer_ameli)),
    
    rechaud_cuis = labelled::remove_labels(rechaud_cuis) %>% structure(label=Hmisc::label(menage_2013$rechaud_cuis)),
    
    telephone = labelled::remove_labels(telephone) %>% structure(label=Hmisc::label(menage_2013$telephone)),

    refri_conge = labelled::remove_labels(refri_conge) %>% structure(label=Hmisc::label(menage_2013$refri_conge)),
    
    video = labelled::remove_labels(video) %>% structure(label=Hmisc::label(menage_2013$video)),
    
    televiseur = labelled::remove_labels(televiseur) %>% structure(label=Hmisc::label(menage_2013$televiseur)),
    
    radio = labelled::remove_labels(radio) %>% structure(label=Hmisc::label(menage_2013$radio)),
    
        piece_occup = labelled::remove_labels(piece_occup) %>% structure(label=Hmisc::label(menage_2002$piece_occup)),
    
    # Evacuation des eaux usées
         evacut_eaux_usees_1 = as.numeric(evacut_eaux_usees_1) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Réseau d'égout")), 
                  
         evacut_eaux_usees_2 = as.numeric(evacut_eaux_usees_2) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Canal fermé")),
                  evacut_eaux_usees_3 = as.numeric(evacut_eaux_usees_3) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Canal à ciel ouvert")),
         
                  evacut_eaux_usees_4 = as.numeric(evacut_eaux_usees_4) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Grille, bouche avaloir")), 
                  
         evacut_eaux_usees_5 = as.numeric(evacut_eaux_usees_5) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Fosse/puisard")),
         
                  evacut_eaux_usees_6 = as.numeric(evacut_eaux_usees_6) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Dans la mer, fleuve")),
         
          evacut_eaux_usees_7 = as.numeric(evacut_eaux_usees_7) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Cours d'eau")),
         
                  evacut_eaux_usees_8 = as.numeric(evacut_eaux_usees_8) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Trou creusé")), 
                  
         evacut_eaux_usees_9 = as.numeric(evacut_eaux_usees_9) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Dans la rue/nature")),
         
             evacut_eaux_usees_98 = as.numeric(evacut_eaux_usees_98) %>% structure(label=paste0(Hmisc::label(menage_2013$evacut_eaux_usees), ":  Autre")),
         
    # Nature du sol
         nature_sol_1 = as.numeric(nature_sol_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Carrelage")),
         
         nature_sol_2 = as.numeric(nature_sol_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Ciment")),
         
         nature_sol_3 = as.numeric(nature_sol_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Argile/banco")),
         
         nature_sol_4 = as.numeric(nature_sol_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Sable")),
         
         nature_sol_5 = as.numeric(nature_sol_5) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Tapis")),
         
             nature_sol_6 = as.numeric(nature_sol_6) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Moquette")),
    
             nature_sol_7 = as.numeric(nature_sol_7) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Bois ciré")),
    
             nature_sol_8 = as.numeric(nature_sol_8) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_sol), ": Autre")),
    
    # Nature du toit
        nature_toit_1 = as.numeric(nature_toit_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Béton")),
        
                nature_toit_2 = as.numeric(nature_toit_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Tuile/ardoise")),
        
                nature_toit_3 = as.numeric(nature_toit_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Zinc")),
        
                nature_toit_4 = as.numeric(nature_toit_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Chaume/paille")),
        
                nature_toit_0 = as.numeric(nature_toit_0) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_toit), ": Autre")),
        
        
    # Nature du mur 
        nature_mur_1 = as.numeric(nature_mur_1) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment")),
         
                nature_mur_2 = as.numeric(nature_mur_2) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment avec carreaux")),
        
                nature_mur_3 = as.numeric(nature_mur_3) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment et marbre")),
        
                nature_mur_4 = as.numeric(nature_mur_4) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Ciment avec bois")),
        
                nature_mur_5 = as.numeric(nature_mur_5) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Bois")),
        
                nature_mur_6 = as.numeric( nature_mur_6) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Banco")),
        
                nature_mur_7 = as.numeric(nature_mur_7) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Banco+ enduit ciment")),
        
                # nature_mur_8 = nature_mur_8) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Paille/Tige")),
        
                nature_mur_9 = as.numeric(nature_mur_9) %>% structure(label=paste0(Hmisc::label(menage_2013$nature_mur), ": Autre")),
        
    # Combustible de cuisine
                  combustible_cuisine_1 = as.numeric(combustible_cuisine_1) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Bois")),
                  combustible_cuisine_2 = as.numeric(combustible_cuisine_2) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Charbon")),
                  combustible_cuisine_3 = as.numeric(combustible_cuisine_3) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Gaz")),
                          combustible_cuisine_4 = as.numeric(combustible_cuisine_4) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Electricite")),
                          combustible_cuisine_5 = as.numeric(combustible_cuisine_5) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Bouse de vache")),
                          combustible_cuisine_8 = as.numeric(combustible_cuisine_8) %>% structure(label=paste0(Hmisc::label(menage_2013$combustible_cuisine), ": Autre")),
        
    # éclairage
                  eclairage_1 = as.numeric(eclairage_1) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Electricite")),
                          eclairage_2 = as.numeric(eclairage_2) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": olaire")),
                          eclairage_3 = as.numeric(eclairage_3) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Grpe.electrogene")),
                          eclairage_4 = as.numeric(eclairage_4) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Gaz")),
        
                          eclairage_5 = as.numeric(eclairage_5) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe a petrole")),
                          eclairage_6 = as.numeric(eclairage_6) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe tempête")),
                          eclairage_7 = as.numeric(eclairage_7) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Bougie")),
                          eclairage_8 = as.numeric(eclairage_8) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Bois")),
        
                          eclairage_9 = as.numeric(eclairage_9) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Lampe rechargeable")),
                              eclairage_98 = as.numeric(eclairage_98) %>% structure(label=paste0(Hmisc::label(menage_2013$eclairage), ": Autre")),
    
    # Approvisionnement en eau
                  approv_eau_1 = as.numeric(approv_eau_1) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet dans logement")),
        
                          approv_eau_2 = as.numeric(approv_eau_2) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet dans cour/parcelle")),
        
                          approv_eau_3 = as.numeric(approv_eau_3) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Robinet public/borne fontaine")),
        
                          approv_eau_4 = as.numeric(approv_eau_4) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits à pompe ou forage")),
        
                          approv_eau_5 = as.numeric(approv_eau_5) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits protégés")),
        
                          approv_eau_6 = as.numeric(approv_eau_6) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Puits non protégés")),
        
                          approv_eau_7 = as.numeric(approv_eau_7) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Vendeurs d’eau")),
        
                          approv_eau_8 = as.numeric(approv_eau_8) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Camion citerne")),
    
                          approv_eau_9 = as.numeric(approv_eau_9) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Eau de pluie")),
        
                          approv_eau_10 = as.numeric(approv_eau_10) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Source, cour d'eau")),
        
                          approv_eau_98 = as.numeric(approv_eau_98) %>% structure(label=paste0(Hmisc::label(menage_2013$approv_eau), ": Autre")),
    
    
    # Type d'aisance
                  type_aisance_11 = as.numeric(type_aisance_11) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Chasse raccordée à l’égout")),
        
                          type_aisance_12 = as.numeric(type_aisance_12) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Chasse avec fosse")),
        
                          type_aisance_21 = as.numeric(type_aisance_21) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Couverte")),
        
                          type_aisance_22 = as.numeric(type_aisance_22) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Non couverte")),
        
                          type_aisance_23 = as.numeric(type_aisance_23) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Ventilée améliorée")),
        
                          type_aisance_31 = as.numeric(type_aisance_31) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Edicule public")),
    
                              type_aisance_41 = as.numeric(type_aisance_41) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Dans la nature")),
        
                          type_aisance_88 = as.numeric(type_aisance_88) %>% structure(label=paste0(Hmisc::label(menage_2013$type_aisance), ": Autre")),
        
    # Statut d'occupation
                  stat_occupation_1 = as.numeric(stat_occupation_1) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Proprietaire")),
                          stat_occupation_2 = as.numeric(stat_occupation_2) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Co proprietaire")),
                          stat_occupation_3 = as.numeric(stat_occupation_3) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Locataire")),
                          stat_occupation_4 = as.numeric(stat_occupation_4) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Co locataire")),
                          stat_occupation_5 = as.numeric(stat_occupation_5) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Location vente")),
                          stat_occupation_6 = as.numeric(stat_occupation_6) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Logé par employeur")),
                          stat_occupation_7 = as.numeric(stat_occupation_7) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Logé par les parents/amis")),
                          stat_occupation_8 = as.numeric(stat_occupation_8) %>% structure(label=paste0(Hmisc::label(menage_2013$stat_occupation), ": Autre")),
         
    # Type de logement
                  type_log_1 = as.numeric(type_log_1) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Case")),
         
                          type_log_2 = as.numeric(type_log_2) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Baraque")),
        
                          type_log_3 = as.numeric(type_log_3) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Maison basse")),
        
                          type_log_4 = as.numeric(type_log_4) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Maison à étage")),
        
                          type_log_5 = as.numeric(type_log_5) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Immeuble à Appartement")),
        
                          type_log_0 = as.numeric(type_log_0) %>% structure(label=paste0(Hmisc::label(menage_2013$type_log), ": Autre"))

  ) %>%
    labelled::unlabelled() %>% 
# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(

    
       type = c(#equipment_var_cat~"continuous",
                equipment_var_num~"continuous"#,
                 #equipment_var_dicho~"dichotomous"
                ),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2)
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Access aux equipements for RGPH 2013") %>%
  add_n()

```


### Household size

```{r, include = TRUE, warning=FALSE, echo = FALSE}
#For the RGPH 2002 dataset
#OK
#For the RGPH 2013 dataset
```

### Head of household

```{r, include = TRUE, warning=FALSE, echo = FALSE}
cm_var=c("niveau_instruction_cm","sexe_cm","age_cm")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,niveau_instruction_cm,sexe_cm,age_cm) %>%
  
  mutate(sexe_cm = as.factor(ifelse(sexe_cm ==2, 0, 1)) %>% 
           structure(label = Hmisc::label(dr_2002$sexe_cm))) %>% 

  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,niveau_instruction_cm,sexe_cm,age_cm) %>%
      
      mutate(sexe_cm = as.factor(ifelse(sexe_cm ==2, 0, 1)) %>% 
               structure(label = Hmisc::label(dr_2013$sexe_cm))))%>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    label = list(sexe_cm ~ "Sexe (Masculin = 1, Féminin = 0)",
      age_cm ~ "Age (en année)",
      niveau_instruction_cm ~ "Nbr d'années d'éducation"),
    type = c(c("age_cm","niveau_instruction_cm")~"continuous",
             c("sexe_cm")~"dichotomous"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Charactéristiques du CM") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```

### Household composition

```{r , include = TRUE, warning=FALSE, echo = FALSE}


liencm_var=c("liencm_cm","liencm_epous","liencm_enf","liencm_parent","liencm_frere","liencm_petit_fils","liencm_sans_lien")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,liencm_cm,liencm_epous,liencm_enf,liencm_parent,liencm_frere,liencm_petit_fils,liencm_sans_lien) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,liencm_cm,liencm_epous,liencm_enf,liencm_parent,liencm_frere,liencm_petit_fils,liencm_sans_lien)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(liencm_cm ~ "CM",
      liencm_epous ~ "Épouse",
      liencm_enf ~ "Enfant",
      liencm_parent ~ "Parent/Grand parent",
      liencm_frere ~ "Frére/soeur",
      liencm_petit_fils ~ "Pétit fils/fille",
      liencm_sans_lien ~ "Sans lien"),
    
    type = c(liencm_var~"continuous"),
    
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(0,0),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "Lien avec le CM") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 

```



# Multivariate calculation for poverty indices (MPI)



```{r, echo=FALSE, warning=FALSE, include=FALSE}

mpi_var=c("cooking_fuel_dim","sanitation_dim","drinking_water_dim","electricity_dim","housing_dim","assets_dim","school_attendance_dim","years_schooling_dim")

#For the RGPH 2002 dataset
dr_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH,cooking_fuel_dim,sanitation_dim,drinking_water_dim,electricity_dim,housing_dim,assets_dim,school_attendance_dim,years_schooling_dim,deces_under_five_dim) %>% 
  
  plyr::rbind.fill(
    
    #For the RGPH 2013 dataset 
    dr_2013 %>% 
      
      # Selecting relevant variables
      select(RGPH,cooking_fuel_dim,sanitation_dim,drinking_water_dim,electricity_dim,housing_dim,assets_dim,school_attendance_dim,years_schooling_dim
             #, deces_under_five_dim
             )) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    
    label = list(
      
      school_attendance_dim ~ "Education: School Attendance",
      deces_under_five_dim ~ "Health: Number of under five mortality"),
    
    type = c(mpi_var~"continuous"),
    
    statistic = list(
                     all_categorical() ~ "{n} ({p}%)",
                     all_continuous() ~ "{mean} ({sd})"
                     ),
    missing_text = "(Missing)",
    digits = everything() ~ c(2,2),
    #missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "MPI Dimension") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() 
```


## Standard of Living dimension

**Data source**: 

* *For the Sanitation*: [link](https://www.who.int/data/gho/data/indicators/indicator-details/GHO/population-using-improved-sanitation-facilities-(-))

* *For water*:
[link](https://data.unicef.org/topic/water-and-sanitation/drinking-water/)
 

*Additionnal cleaning*


## Health dimension



```{r, echo=FALSE, warning=FALSE, include=FALSE}
## Removing all objects
rm(list=ls())
```
