==> script/03_MPI Senegal Census data.Rmd <==
---
title: "Computing the MPI using the Senegal Census data and visualization"
author: "HEMA Aboubacar"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, error=FALSE)
```


# Preambule: 

This MPI computation is inspired from the method described in the [link](https://www.publichealthnotes.com/what-is-multidimensional-poverty-index-mpi/).
```{r}
rm(list=ls())
```
## Importing library

```{r Package needed, results = "hide"}


## Importing library
### List of required packages
required_packages <- c("tidyverse","janitor" ,"readr","dplyr","haven","sf", "flextable","sp", "factoextra", "FactoMineR","gtsummary", "sjPlot", "fastDummies","ggthemes")

# Check if packages are installed
missing_packages <- setdiff(required_packages, installed.packages()[,"Package"])

### Install missing packages
if (length(missing_packages) > 0) {
  install.packages(missing_packages)
}

### Load all packages
lapply(required_packages, library, character.only = TRUE)

```

## Importing dataset

```{r, results='hide'}

# Read and check for duplicated rows in the "menage_2002" dataset
menage_2002 <- base::readRDS(paste0(here::here(),"/output/output_data/menage_2002.rds"))
menage_2002 %>% janitor::get_dupes() 

# Read and check for duplicated rows in the "menage_2013" dataset
menage_2013 <- base::readRDS(paste0(here::here(),"/output/output_data/menage_2013.rds"))
menage_2013 %>% janitor::get_dupes() 

# Read shapefile data for 2002 and 2013
sp_rgph_2002 <- sf::read_sf(paste0(here::here(),"/output/output_data/EAs_2002_new.shp"))
sp_rgph_2013 <- sf::read_sf(paste0(here::here(),"/data/EAs_2013_2002_SHP/EAs_2013.shp"))


# Importing the crosswalk
#crosswalk_table_dr <- readxl::read_excel(paste0(here::here(),"/data/TABLE_CORRESPONDANCE_SPATIALE_DR.xls"))
```




```{r}
edu_var = c("Primaire","Moyen","Secondaire","Superieur")
#educ_var = c("Aucun","Primaire","Moyen","Secondaire","Superieur")
# For the 2002 individus data  

menage_2002 <- 
  menage_2002 %>% 
  
# Calculate a binary variable 'school_attendance' for each household.
# If the sum of education levels is greater than 0, it means at least one individual in the household is attending school (0), 
# otherwise, it means no one is attending school (1).
  mutate(years_schooling_dim = rowSums(select(., .dots=all_of(edu_var)),na.rm = T),
    years_schooling_dim = 
           as.factor(ifelse(years_schooling_dim > 0, 0, 1)) %>% structure(label = "Education: years of schooling"))
  
# For the 2013 individus data  

menage_2013 <- 
  menage_2013 %>% 
  
# Calculate a binary variable 'school_attendance' for each household.
# If the sum of education levels is greater than 0, it means at least one individual in the household is attending school (0), 
# otherwise, it means no one is attending school (1).
  mutate(years_schooling_dim = rowSums(select(., .dots=all_of(edu_var)),na.rm = T),
    years_schooling_dim = 
           as.factor(ifelse(years_schooling_dim > 0, 0, 1)) %>% structure(label = "Education: years of schooling"))
```


```{r}
# Convert to character the variables that are needed to compute the MPI

# For RGPH 2002
# Update the 'menage_2002' dataframe using the pipe operator '%>%'
menage_2002 <- menage_2002 %>%  
  # Add a new column 'MPI' to the dataframe
  mutate(
    deces_under_five_dim = as.numeric(as.character(deces_under_five_dim)),   # Convert 'deces_under_five_dim' to numeric
    deces_under_five_dim = ifelse(deces_under_five_dim>0,1,0),
    years_schooling_dim = as.numeric(as.character(years_schooling_dim)),     # Convert 'years_schooling_dim' to numeric
    school_attendance_dim = as.numeric(as.character(school_attendance_dim)), # Convert 'school_attendance_dim' to numeric
    cooking_fuel_dim = as.numeric(as.character(cooking_fuel_dim)),           # Convert 'cooking_fuel_dim' to numeric       
    sanitation_dim = as.numeric(as.character(sanitation_dim)),               # Convert 'sanitation_dim' to numeric   
    drinking_water_dim = as.numeric(as.character(drinking_water_dim)),       # Convert 'drinking_water_dim' to numeric
    electricity_dim = as.numeric(as.character(electricity_dim)),             # Convert 'electricity_dim' to numeric
    housing_dim = as.numeric(as.character(housing_dim)),                     # Convert 'housing_dim' to numeric
    assets_dim = as.numeric(as.character(assets_dim))                         # Convert 'assets_dim' to numeric
  )


# For RGPH 2013

# Update the 'menage_2013' dataframe using the pipe operator '%>%'
menage_2013 <- menage_2013 %>%  
  mutate(deces_under_five_dim=0,
    years_schooling_dim = as.numeric(as.character(years_schooling_dim)),     # Convert 'years_schooling_dim' to numeric
    school_attendance_dim = as.numeric(as.character(school_attendance_dim)), # Convert 'school_attendance_dim' to numeric
    cooking_fuel_dim = as.numeric(as.character(cooking_fuel_dim)),           # Convert 'cooking_fuel_dim' to numeric       
    sanitation_dim = as.numeric(as.character(sanitation_dim)),               # Convert 'sanitation_dim' to numeric   
    drinking_water_dim = as.numeric(as.character(drinking_water_dim)),       # Convert 'drinking_water_dim' to numeric
    electricity_dim = as.numeric(as.character(electricity_dim)),             # Convert 'electricity_dim' to numeric
    housing_dim = as.numeric(as.character(housing_dim)),                     # Convert 'housing_dim' to numeric
    assets_dim = as.numeric(as.character(assets_dim))                         # Convert 'assets_dim' to numeric
  )


```

```{r}
dimen = c("deces_under_five_dim","years_schooling_dim","school_attendance_dim","cooking_fuel_dim","sanitation_dim","drinking_water_dim","electricity_dim","housing_dim","assets_dim")

menage_2002 %>% 
  
  # Selecting relevant variables
  select(RGPH, deces_under_five_dim, years_schooling_dim,school_attendance_dim,cooking_fuel_dim,sanitation_dim,drinking_water_dim,electricity_dim,housing_dim,assets_dim) %>% 
  
  plyr::rbind.fill(
    
    # Selecting relevant variables
    menage_2013 %>% 
      select(RGPH, deces_under_five_dim, years_schooling_dim,school_attendance_dim,cooking_fuel_dim,sanitation_dim,drinking_water_dim,electricity_dim,housing_dim,assets_dim)) %>% 

# Generate a summary table using "tbl_summary" for the specified columns
  tbl_summary(
    by = RGPH,  # Group by "RGPH" column
    # label = list(years_schooling_dim,school_attendance_dim,cooking_fuel_dim,sanitation_dim,drinking_water_dim,electricity_dim,housing_dim,assets_dim),
    type = c(dimen~"categorical"),
    statistic = list(
                     all_categorical() ~ "{n} ({p}%,{n}/{N})",
                     all_continuous() ~ "{mean}"
                     ),
    digits = everything() ~ c(0,0),
    missing = "no"
  ) %>% 

  # Modify the table header to provide a descriptive label
  modify_header(label ~ "MPI - Dimensions") %>%
  add_overall() %>%
  add_n()%>% 
  # Add difference statistics to the table
  add_difference() %>% 
  bold_labels() 
```


```{r}
# For RGPH 2002

# Update the 'menage_2002' dataframe using the pipe operator '%>%'
menage_2002 <- menage_2002 %>% 
  
  # Add a new column 'MPI_index' to the dataframe
  mutate(
    
    # I use 1/3 as weight for the 'deces_under_five_dim' because the nutrition dimension is not available in the 2002 dataset. 
    

    deces_under_five_dim = deces_under_five_dim/3,  # Divide 'deces_under_five_dim' by 3
    years_schooling_dim = years_schooling_dim/6,     # Divide 'years_schooling_dim' by 6
      school_attendance_dim  = school_attendance_dim/6,   # Divide 'school_attendance_dim' by 6
         cooking_fuel_dim =cooking_fuel_dim/18,       # Divide 'cooking_fuel_dim' by 18
         sanitation_dim= sanitation_dim/18,       # Divide 'sanitation_dim' by 18
         drinking_water_dim =drinking_water_dim/18,    # Divide 'drinking_water_dim' by 18
          electricity_dim=electricity_dim/18,        # Divide 'electricity_dim' by 18
         housing_dim =housing_dim/18,           # Divide 'housing_dim' by 18
         assets_dim= assets_dim/18             # Divide 'assets_dim' by 18
          )


# For RGPH 2013

# Update the 'menage_2013' dataframe using the pipe operator '%>%'
menage_2013 <- menage_2013 %>% 
  
  # Add a new column 'MPI_index' to the dataframe
  # In the 2013 dataset, information for the health dimension are not available, so I will reallocate the weight uniformly across each dimension.
  
  mutate(
    deces_under_five_dim = 0, 
    years_schooling_dim = years_schooling_dim/4,     # Divide 'years_schooling_dim' by 4
      school_attendance_dim  = school_attendance_dim/4,   # Divide 'school_attendance_dim' by 4
         cooking_fuel_dim =cooking_fuel_dim/12,       # Divide 'cooking_fuel_dim' by 12
         sanitation_dim= sanitation_dim/12,       # Divide 'sanitation_dim' by 12
         drinking_water_dim =drinking_water_dim/12,    # Divide 'drinking_water_dim' by 12
          electricity_dim=electricity_dim/12,        # Divide 'electricity_dim' by 12
         housing_dim =housing_dim/12,           # Divide 'housing_dim' by 12
         assets_dim= assets_dim/12)             # Divide 'assets_dim' by 12


```


# Computing the MPI using the three dimension

```{r}

# For RGPH 2002

# Update the 'menage_2002' dataframe using the pipe operator '%>%'
menage_2002 <- menage_2002 %>% 
  
  # Add a new column 'MPI_index' to the dataframe
  mutate(
    
    # I use 1/3 as weight for the 'deces_under_five_dim' because the nutrition dimension is not available in the 2002 dataset. 
    

    MPI_index = rowSums(select(., .dots=all_of(dimen)),na.rm = T) %>% 
          structure(label = "Multidimensional Poverty Index (MPI)"))


# For RGPH 2013

# Update the 'menage_2013' dataframe using the pipe operator '%>%'
menage_2013 <- menage_2013 %>% 
  
  # Add a new column 'MPI_index' to the dataframe
  # In the 2013 dataset, information for the health dimension are not available, so I will reallocate the weight uniformly across each dimension.
  
  mutate(

    MPI_index = rowSums(select(., .dots=all_of(dimen)),na.rm = T) %>% 
          structure(label = "Multidimensional Poverty Index (MPI)"))



```

L'âge du décès n'est pas mentionné dans le recensement de 2013, nous ne pouvons donc pas calculer le nombre de décès pour les personnes de moins de 18 ans.

```{r}
# Normalising by the number of houses in the hh

menage_2002 <- menage_2002 %>% 
  mutate(indice_peoplement = hh_size/piece_occup)

menage_2013 <- menage_2013 %>% 
  mutate(indice_peoplement = hh_size/piece_occup)
```


```{r}
mat_status_var = c("Monogame","Polygame","Celibataire","Veuf","Divorce")
fertility_var = c("fertility")
educ_var = c("Aucun","Primaire","Moyen","Secondaire","Superieur")
educ_cm_var = c("Aucun_cm","Primaire_cm","Moyen_cm","Secondaire_cm","Superieur_cm")
literacy_var = c("Literacy_French","Literacy_Arabic","Literacy_Wolof","Literacy_Pulaar","Literacy_Mandingo","Literacy_Diola","Literacy_Soninke")#"Literacy_Sereer",
#dec_var =c("deces_mat_oui")
mig_var = c("resident_pres","resident_abs","resident_vis")
gender_var = c("homme","femme","hh_size")
income_var = c("I_repas_saute","I_soins_medicaux")
sitoccup_var = c("sit_occup_occup","sit_occup_chomeur","sit_occup_etudiant","sit_occup_foyer","sit_occup_retraite","sit_occup_rentier","sit_occup_autre")

comp_var=c("nbr_babies","nbr_under_14","nbr_young_people","nbr_adult","nbr_elder_adult","nbr_senior")

occup_var = c("sit_occup_occup","sit_occup_chomeur","sit_occup_etudiant","sit_occup_foyer","sit_occup_retraite","sit_occup_rentier","sit_occup_autre" )

study_var = c(gender_var,
              fertility_var,
              educ_var,
              educ_cm_var,
              literacy_var,
              mig_var,
              income_var,
              sitoccup_var,
              comp_var,
              occup_var)
```

```{r}

# The RGPH 2002
# Grouping menage_2002 data by id_dr and summarizing relevant information
plt_df_2002 <- menage_2002 %>%
  group_by(id_dr) %>%
  summarise_at(study_var, sum, na.rm = TRUE) %>%
  left_join(
    menage_2002 %>%
  group_by(id_dr) %>%
  summarise(menage = n(),  # Counting the number of households in each id_dr group
            MPI_mean = mean(MPI_index, na.rm = TRUE),
            median_cm_age = median(age_cm, na.rm =TRUE),
            mean_cm_age = mean(age_cm, na.rm =TRUE),
            mean_ind_peo = mean(indice_peoplement, na.rm=TRUE),
            median_ind_peo = median(indice_peoplement, na.rm=TRUE),
            cm_homme = sum(sexe_cm == 1, na.rm = TRUE),
            cm_femme = sum(sexe_cm == 2, na.rm = TRUE),
            pct_cm_homme = cm_homme/(cm_homme + cm_femme),
            pct_cm_femme = cm_femme/(cm_homme + cm_femme)
            )
  
  )%>%
  # Ungrouping the data to avoid unintended side effects in subsequent operations
  ungroup() %>%

  # Calculating the proportion of poor households and adding the results as a new column
  mutate(#prop_poor = poor / menage,

         # The information on the survey
         RGPH = "2002") %>%


  # Joining with the sp_rgph_2002 data frame based on the id_dr column
  left_join(sp_rgph_2002, by = "id_dr")

# Average accross the household size
plt_df_2002 <- plt_df_2002 %>% 
  mutate(across(setdiff(study_var,educ_cm_var), ~ . / hh_size))

# Average across the number of hh in the DR
plt_df_2002 <- plt_df_2002 %>% 
  mutate(across(educ_cm_var, ~ . / menage))

# plt_df_2002 <- menage_2002 %>% 
#   group_by(id_dr) %>% 
#   summarise_at(study_var, mean, na.rm = TRUE) %>% 
#   left_join(
#      plt_df_2002
#   )
  # group_by(id_dr) %>%
  # summarise_at(menage = n(),  # Counting the number of households in each id_dr group
  #           MPI_mean = mean(MPI_index, na.rm = TRUE),
  #           median_cm_age = median(age_cm, na.rm =TRUE),
  #           mean_cm_age = mean(age_cm, na.rm =TRUE),
  #           cm_homme = sum(sexe_cm == 1, na.rm = TRUE),
  #           cm_femme = sum(sexe_cm == 2, na.rm = TRUE),
  #           pct_cm_homme = cm_homme/(cm_homme + cm_femme),
  #           pct_cm_femme = cm_femme/(cm_homme + cm_femme)),
  # 
  #           nbr_babies_mean = sum(nbr_babies, na.rm=TRUE),
  # )%>%  
  # Ungrouping the data to avoid unintended side effects in subsequent operations
  # ungroup() %>%
  
  # Calculating the proportion of poor households and adding the results as a new column
  # mutate(#prop_poor = poor / menage,
  #                 
  #        # The information on the survey
  #        RGPH = "2002") %>%
  # 
  # 
  # # Joining with the sp_rgph_2002 data frame based on the id_dr column
  # left_join(sp_rgph_2002, by = "id_dr")






# The RGPH 2013
# Grouping menage_2013 data by id_dr and summarizing relevant information
plt_df_2013 <- menage_2013 %>% 
  group_by(id_dr) %>% 
  summarise_at(study_var, sum, na.rm = TRUE) %>% 
  left_join(
    menage_2013 %>% 
      group_by(id_dr) %>%
      summarise(menage = n(),  # Counting the number of households in each id_dr group
                MPI_mean = mean(MPI_index, na.rm = TRUE),
                median_cm_age = median(age_cm, na.rm =TRUE),
                mean_cm_age = mean(age_cm, na.rm =TRUE),
                mean_ind_peo = mean(indice_peoplement, na.rm=TRUE),
                median_ind_peo = median(indice_peoplement, na.rm=TRUE),
                cm_homme = sum(sexe_cm == 1, na.rm = TRUE),
                cm_femme = sum(sexe_cm == 2, na.rm = TRUE),
                pct_cm_homme = cm_homme/(cm_homme + cm_femme),
                pct_cm_femme = cm_femme/(cm_homme + cm_femme))
  )%>%  
  # Ungrouping the data to avoid unintended side effects in subsequent operations
  ungroup() %>%
  
  # Calculating the proportion of poor households and adding the results as a new column
  mutate(#prop_poor = poor / menage,
    
    # The information on the survey
    RGPH = "2013") %>%
  
  
  # Joining with the sp_rgph_2013 data frame based on the id_dr column
  left_join(
    sp_rgph_2013 %>%
  rename(id_dr=COD_DR2012), by = "id_dr")


# Average accross the household size
plt_df_2013 <- plt_df_2013 %>% 
  mutate(across(setdiff(study_var,educ_cm_var), ~ . / hh_size))

# Average across the number of hh in the DR
plt_df_2013 <- plt_df_2013 %>% 
  mutate(across(educ_cm_var, ~ . / menage))

# plt_df_2013 <- menage_2013 %>% 
#   group_by(id_dr) %>% 
#   summarise_at(study_var, mean, na.rm = TRUE) %>% 
#   left_join(
#      plt_df_2013
#   )


```


```{r}
plt_df_2013 = plt_df_2013 %>% 
  dplyr::rename(REGION=REG,
                NOM_ARRDT=CAV,
                NOM_CA=CCRCA
                )

other_var = c(
              "MPI_mean",
              "menage",
              "median_cm_age",
              "mean_cm_age",
              "cm_homme",
              "cm_femme",
              "pct_cm_homme",
              "pct_cm_femme",
              "mean_ind_peo",
              "median_ind_peo",
              "geometry"
              )
```

```{r}
plt_df_2002 = plt_df_2002 %>% 
  dplyr::select(#"OBJECTID",
              "id_dr",
              "REGION",
              "DEPT",
              "NOM_ARRDT",
              "NOM_CA",
              "RGPH",
              study_var,
              other_var)


plt_df_2013 = plt_df_2013 %>% 
  dplyr::select(#"OBJECTID",
              "id_dr",
              "REGION",
              "DEPT",
              "NOM_ARRDT",
              "NOM_CA",
              "RGPH",
              study_var,
              other_var)
```


```{r}
# Adding other variables to the variables


```




```{r}
df_2002=plt_df_2002 %>%
  # Convert the resulting data frame to a spatial data frame using st_as_sf()
  st_as_sf()
df_2013=plt_df_2013 %>%
  # Convert the resulting data frame to a spatial data frame using st_as_sf()
  st_as_sf()



# lookup <- c(ID = "id_dr")
# df_2002=plt_df_2002 %>% dplyr::rename(all_of(lookup))
# df_2013=plt_df_2013 %>% dplyr::rename(all_of(lookup))

# Saving the merged dataset in the sf format (for the household data)
sf::st_write(obj = df_2002, paste0(here::here(), "/output/output_data/MPI_data_dr_2002.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE)

# Saving the merged dataset in the sf format (for the household data)
sf::st_write(obj = df_2013, paste0(here::here(), "/output/output_data/MPI_data_dr_2013.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE)


```


```{r}

# Binding the datasets and creating an sf object plot
# Combine data from plt_2002 and plt_2013 using rbind.fill from plyr package
MPI_data_dr <- plt_df_2002 %>%
  plyr::rbind.fill(plt_df_2013) %>% 



  # # Mutate columns and add labels using structure() for better readability
  # mutate(
  #   id_dr = id_dr %>% structure(label = "DR identifier"),
  #   MPI_mean = MPI_mean %>% structure(label = "Multidimensional Poverty Index (MPI)"),
  #   #prop_poor = prop_poor %>% structure(label = "Proportion of poor in the DR"),
  #   #poor = poor %>% structure(label = "Number of poor households in the DR"),
  #   menage = menage %>% structure(label = "Number of households in the DR"),
  #   hh_size = hh_size %>% structure(label = "Number of household members"),
  #   RGPH = RGPH %>% structure(label = "RGPH year"),
  #   Shape_Area = Shape_Area %>% structure(label = "Shape Area"),
  #   Shape_Leng = Shape_Leng %>% structure(label = "Shape Length"),
  #   REGION = REGION %>% structure(label = "Region"), 
  #   DEPT = DEPT %>% structure(label = "Department"),
  #   NOM_ARRDT = NOM_ARRDT %>% structure(label = "Name district"), 
  #   NOM_CA = NOM_CA %>% structure(label = "Name CA")
  # ) %>%


  # Convert the resulting data frame to a spatial data frame using st_as_sf()
  st_as_sf()

# Creating a ggplot object, mapping the fill color to the MPI
MPI_data_dr %>% 
    
      ggplot(aes(fill = MPI_mean)) + 
      
      # Adding spatial features to the plot
      geom_sf() +
      
      # Customizing the color scale for the fill
      scale_fill_continuous(type = "gradient", low = "white", high = "#1E1E1E", 
                             name = 'Proportion', breaks = 0.15 * 0:4, 
                             labels = c("0", "15", "30", "45", "60")) +
       
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
            legend.background = element_blank(),
            axis.ticks = element_blank(),
            axis.text = element_blank(),
            legend.position = c(.92, .03), 
            text = element_text(size = 8), 
            panel.grid = element_line(color = "white", size = 0.8))
      
      # Adding plot titles, caption and saving the plot
      #  + labs(
      #   title = "Indice Multidimensionnel de Pauvreté au Sénégal par année et district (en %)",
      #   caption = "Données : RGPH, Indicateur: Indice Multidimensionnel de Pauvreté (IMP).")
    ggsave(paste0(here::here(),"/output/output_img/MPI_rgph.pdf"), width = 8, height = 5)

ggsave(paste0(here::here(),"/output/output_img/MPI_rgph.png"), width = 8, height = 5)
```




<!-- # Exporting the datasets -->

<!-- ```{r eval=FALSE, include=FALSE, results='hide'} -->

<!-- # Selecting the relevant variables -->
<!-- MPI_data_dr <- MPI_data_dr %>%  -->
<!--   select(id_dr, REGION, DEPT, RGPH, NOM_ARRDT, NOM_CA, MPI_mean, menage, hh_size, Shape_Leng, Shape_Area, geometry) -->

<!-- # Saving the merged dataset in the sf format (for the household data) -->
<!-- sf::st_write(obj = MPI_data_dr, paste0(here::here(), "/output/output_data/MPI_data_dr.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) -->
<!-- ``` -->



<!-- ```{r eval=FALSE, include=FALSE} -->

<!-- # Selecting specific columns from the 'menage_2002' data frame -->
<!-- MPI_data_household <-  -->
<!--     menage_2002 %>%  -->
<!--     select(id_dr, ID_MENAGE, MPI_index, MPI_poor,  hh_size,  RGPH) %>%  -->

<!--   # Binding with the menage_2013 data frame -->

<!--     plyr::rbind.fill( -->
<!--       # Selecting specific columns from the 'menage_2013' data frame -->
<!--       menage_2013 <- -->
<!--         menage_2013 %>%  -->
<!--         select(id_dr, ID_MENAGE, MPI_index, MPI_poor,  hh_size, RGPH) -->
<!--     ) -->
<!-- ``` -->


<!-- ```{r eval=FALSE, include=FALSE} -->
<!-- # Exporting the MPI_data_dr dataset to include the MPI indicator -->

<!-- # Saving in SPSS format -->
<!-- haven::write_sav(MPI_data_household,paste0(here::here(), "/output/output_data/MPI_data_household.sav")) -->

<!-- # Saving in R format -->
<!-- saveRDS(MPI_data_household,file=paste0(here::here(), "/output/output_data/MPI_data_household.rds")) -->

<!-- ``` -->


https://ophi.org.uk/multidimensional-poverty-index/#:~:text=The%20global%20Multidimensional%20Poverty%20Index,that%20a%20person%20faces%20simultaneously.

https://ophi.org.uk/ophi-mpi-methodological-note-49/ 

https://www.ophi.org.uk/wp-content/uploads/OPHI_MPI_MN_49_2020.pdf

http://hdr.undp.org/sites/default/files/mpi_trainingmaterial_mcc_mk_clean_june_2015.pdf

https://www.ophi.org.uk/wp-content/uploads/MPI-Primer.pdf

https://www.ophi.org.uk/wp-content/uploads/OPHI-wp38_with_note.pdf

https://www.sciencedirect.com/science/article/abs/pii/S0305750X14000278#:~:text=The%20potential%20dimensions%20that%20a,culture%20(Alkire%2C%202008).

https://www.ukessays.com/essays/economics/strengths-and-weaknesses-of-multidimensional-indicators-of-poverty-economics-essay.php

https://mpra.ub.uni-muenchen.de/108501/1/MPRA_paper_108501.pdf
