==> script/05_Modelling Senegal Census data.Rmd <==
---
title: "Modeling Guediawaye Census data "
author: "HEMA Aboubacar"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: show
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, error=FALSE)
```



```{r}
rm(list=ls())
```

## Importing library

```{r Package needed, results = "hide"}


## Importing library
### List of required packages
required_packages <- c("tidyverse","janitor" ,"readr","dplyr","haven","sf", "flextable","sp", "factoextra", "FactoMineR","gtsummary", "sjPlot", "fastDummies","ggthemes","spdep","patchwork","corrr")

# Check if packages are installed
missing_packages <- setdiff(required_packages, installed.packages()[,"Package"])

### Install missing packages
if (length(missing_packages) > 0) {
  install.packages(missing_packages)
}

### Load all packages
lapply(required_packages, library, character.only = TRUE)

```


```{r}
# Read shapefile data for 2013 and 2013


MPI_data_dr_2013 <- sf::read_sf(paste0(here::here(),"/output/output_data/MPI_data_dr_2013.shp"))

MPI_data_dr_2002 <- sf::read_sf(paste0(here::here(),"/output/output_data/MPI_data_dr_2002.shp"))

MPI_data_dr <- MPI_data_dr_2013 %>%
  plyr::rbind.fill(MPI_data_dr_2002) %>% 
  st_as_sf()
```



# Regression modeling with Guediawaye Census data

```{r}

outcome = "MPI_men"
# vars_to_remove = c("id_dr","REGION","DEPT","NOM_ARR","NOM_CA","RGPH","homme",    "femme","hh_size","Aucun","Superir","Ltrcy_P","Ltrcy_S","Ltrcy_D","rsdnt_p","rsdnt_b","I_rps_s","I_sns_m","MPI_men","mdn_cm_","mn_cm_g","cm_homm","cm_femm","pct_cm_h","geometry")



labels_var = c("Id dr","Region","Dept.","Nom Arr.","Nom CA","RGPH","Proportion d'homme","proportion de femme","Taille moyenne du ménage","Fertilité moyenne","Niveau d'instruction: Aucun","Niveau d'instruction: Primaire","Niveau d'instruction: Moyen","Niveau d'instruction: Secondaire", "Niveau d'instruction: Supérieur",
               
"Niveau d'instruction du CM: Aucun","Niveau d'instruction du CM: Primaire","Niveau d'instruction du CM: Moyen","Niveau d'instruction du CM: Secondaire", "Niveau d'instruction du CM: Supérieur","Proportion de French","Proportion de Arabic ","Proportion de Wolof","Proportion de Pulaar","Proportion de Sereer","Proportion de Mandingo","Proportion de Diola","Proportion de Soninke","Proportion de résident permanent","Proportion de Résident Absent","Proportion de repas sauté","Proportion de soins médicaux",
               
"Proportion d'individus qui sont occupés dans leur occupation principale ","Proportion d'individus qui sont au chômage ou à la recherche d'un premier emploi dans leur occupation principale","Proportion d'individus qui sont occupés au foyer dans leur occupation principale","Proportion d'individus qui sont étudiants/élèves dans leur occupation principale","Proportion d'individus qui sont rentiers dans leur occupation principale","Proportion d'individus qui sont retraités ou personnes du 3ème âge dans leur occupation principale","Proportion d'individus qui ont d'autres situations dans leur occupation principale","Proportion d'enfant agée de moins d'un an ","Proportion d'enfant agée de moins de 14 ans","Proportion d'individus agée entre 15 et 29 ans","Proportion d'individus agées entre 30 et 44 ans","Proportion d'individus agées entre 45 et 59 ans","Proportion d'individus agée de plus de 60 ans",
"Proportion de ménage","Age médian du CM","Age moyen du CM","Nombre de CM de sexe masculin","Nombre de CM de sexe feminin","Proportion de CM de sexe masculin","Proportion de CM de sexe féminin",

"Moyenne de l'indice de peoplement", "Médiane de l'indice de peoplement",

"geometry")

# label_to_remove = c("Id dr","Region","Dept.","Nom Arr.","Nom CA","RGPH","Proportion d'homme","Proportion de femme","Taille ménage","Niveau d'instruction: Aucun", "Niveau d'instruction: Supérieur","Proportion de Pulaar ","Proportion de Sereer ","Proportion de Diola ","Proportion de résident permanent","Proportion de Résident Absent","Proportion de repas sauté","Proportion de soins médicaux","MPI","Age médian du CM dans le DR","Age moyen du CM dans le DR","Proportion de CM de sexe masculin","Proportion de CM de sexe feminin","Proportion de CM de sexe masculin","geometry")

# remove some variables because of the multicolinearity problem
vars_to_remove = c("id_dr","REGION","DEPT","NOM_ARR","NOM_CA","RGPH","homme","hh_size","Primair","Prmr_cm","Ltrcy_F","rsdnt_p","st_ccp_cc","st_ccp_rn","nbr_bbs","nbr_snr","MPI_men","cm_homm","cm_femm","pct_cm_h","geometry")


label_to_remove = c("Id dr","Region","Dept.","Nom Arr.","Nom CA","RGPH","Proportion d'homme","Taille ménage","Taille ménage","Niveau d'instruction: Aucun","Niveau d'instruction du CM: Primaire","Proportion de French","Proportion de résident permanent","Proportion d'individus qui sont occupés dans leur occupation principale","Proportion d'individus qui sont retraités ou personnes du 3ème âge dans leur occupation principale","Proportion d'enfant agée de moins d'un an","Proportion d'individus agée de plus de 60 ans","Nombre de CM de sexe masculin","Nombre de CM de sexe féminin","Proportion de CM de sexe masculin","geometry")


predictors_2002 = setdiff(names(MPI_data_dr_2002),vars_to_remove)
predictors_2002 = append(predictors_2002,"pop_density")
label_predictors_2002 = append(setdiff(labels_var, label_to_remove), "densité de la population")
  
predictors_2013 =  setdiff(names(MPI_data_dr_2013),vars_to_remove)
predictors_2013 =  append(predictors_2013,"pop_density")
label_predictors_2013 = append(setdiff(labels_var, label_to_remove), "densité de la population")

formula_2002 = paste0("log(",outcome,") ~ ",paste(predictors_2002[-length(predictors_2002)]," + ", collapse = ""),predictors_2002[length(predictors_2002)])

formula_2013 = paste0("log(",outcome,") ~ ",paste(predictors_2013[-length(predictors_2013)]," + ", collapse = ""),predictors_2013[length(predictors_2013)])


# formula_2002 <- paste0(formula_2002, " + pop_density")
# formula_2013 <- paste0(formula_2013, " + pop_density")

## Remove the household size from the formula

formula_2002 <- gsub(' + hh_size ','',formula_2002, fixed = TRUE)
formula_2013 <- gsub(' + hh_size ','',formula_2013, fixed = TRUE)

```

## Inspecting the outcome variable (MPI) with visualization

```{r}

mhv_map <- ggplot(MPI_data_dr_2013, aes(fill = MPI_men)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  theme_void() + 
  labs(fill = "MPI ")

mhv_histogram <- ggplot(MPI_data_dr_2013, aes(x = MPI_men)) + 
  geom_histogram(alpha = 0.5, fill = "navy", color = "navy",
                 bins = 100) + 
  theme_minimal() + 
  scale_x_continuous(labels = scales::label_number_si(accuracy = 0.1)) + 
  labs(x = "MPI")

#grid.arrange(mhv_map, mhv_histogram, nrow = 1)
# mhv_map + mhv_histogram + labs(title = "MPI value charts for Guediawaye Census 2013")
gridExtra::grid.arrange(mhv_map, mhv_histogram, nrow = 1,
                        top = "MPI value charts for Guediawaye Census 2013")

ggsave(paste0(here::here(),"/output/output_model/img/1_hv_histogram.png"), width = 8, height = 5)
```


```{r}
MPI_data_dr %>% 
  mutate(RGPH=ifelse(RGPH=="2013","Census 2013", "Census 2002")) %>% 
    
      ggplot(aes(fill = MPI_men)) + 
      
      geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  theme_void() + 
  labs(fill = "MPI")+
       
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )

ggsave(paste0(here::here(),"/output/output_model/img/2_mpi_rgph.png"), width = 8, height = 5)
```
```{r}
MPI_data_dr %>% 
  mutate(RGPH=ifelse(RGPH=="2013","Census 2013", "Census 2002")) %>%
  ggplot(aes(x = MPI_men)) + 
  geom_histogram(alpha = 0.5, fill = "navy", color = "navy",
                 bins = 100) + 
  theme_minimal() + 
  scale_x_continuous(labels = scales::label_number_si(accuracy = 0.1)) +
      xlab("MPI")+
      facet_wrap(~RGPH) +
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))

ggsave(paste0(here::here(),"/output/output_model/img/3_mpi_rgph_hist.png"), width = 8, height = 5)
```


```{r}

mhv_map_log <- ggplot(MPI_data_dr_2013, aes(fill = log(MPI_men))) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  theme_void() + 
  labs(fill = "MPI\nvalue (log)")

mhv_histogram_log <- ggplot(MPI_data_dr_2013, aes(x = log(MPI_men))) + 
  geom_histogram(alpha = 0.5, fill = "navy", color = "navy",
                 bins = 100) + 
  theme_minimal() + 
  scale_x_continuous() + 
  labs(x = "MPI (log)")

gridExtra::grid.arrange(mhv_map_log, mhv_histogram_log, nrow = 1,
                        top = "Logged MPI value charts for Guediawaye Census 2013")

ggsave(paste0(here::here(),"/output/output_model/img/4_hv_histogram_log.png"), width = 8, height = 5)

```


```{r}
MPI_data_dr %>% 
    
      ggplot(aes(fill = log(MPI_men))) + 
      
      geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  theme_void() + 
  labs(fill = "MPI\nvalue (log)")+
       
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )
ggsave(paste0(here::here(),"/output/output_model/img/5_log_mpi_rgph.png"), width = 8, height = 5)
```


## A first regression model

```{r}
library(sf)
library(units)

# predictors = c("menage","Primair","Moyen","Secondr","Ltrcy_F","pct_cm_f","hh_size","Ltrcy_A","Ltrcy_W","Ltrcy_M","fertlty")

# outcome = "MPI_men"
# MPI_data_dr_2013_for_model<- MPI_data_dr_2013 %>%
#   dplyr::select(MPI_men,predictors) %>% 
#   mutate(pop_density = as.numeric(set_units(hh_size / st_area(.), "1/km2"))) %>% 
#   dplyr::select(-hh_size)

outcome = "MPI_men"
MPI_data_dr_2013_for_model<- MPI_data_dr_2013 %>%
  mutate(pop_density = as.numeric(set_units(hh_size / st_area(.), "1/km2"))) %>% 
  dplyr::select(MPI_men,predictors_2013)

##
# MPI_data_dr_2002_for_model<- MPI_data_dr_2002 %>%
#   dplyr::select(MPI_men,predictors) %>% 
#   mutate(pop_density = as.numeric(set_units(hh_size / st_area(.), "1/km2"))) %>% 
#   dplyr::select(-hh_size)

MPI_data_dr_2002_for_model<- MPI_data_dr_2002 %>%
  mutate(pop_density = as.numeric(set_units(hh_size / st_area(.), "1/km2"))) %>% 
  dplyr::select(MPI_men,predictors_2002) 

# Appending pop density as a predictor for 
```

```{r}

# formula <- "log(MPI_men) ~ menage  + Primair + Moyen + Secondr + Ltrcy_F + pct_cm_f + pop_density + Ltrcy_A + Ltrcy_W + Ltrcy_M + fertlty"


### For 2002
model_2002 <- lm(formula = formula_2002, data = MPI_data_dr_2002_for_model)

### For 2013

model_2013 <- lm(formula = formula_2013, data = MPI_data_dr_2013_for_model)



stargazer::stargazer(model_2002, model_2013, type="text",

covariate.labels = label_predictors_2002, out = (paste0(here::here(),"/output/output_model/1_reg_lineaire.txt")))
```


```{r}

##unlist(noquote(paste0(predictors_2002, " ~ " , shQuote(label_predictors_2002), collapse = " , ")))

tbl_2002 <- gtsummary::tbl_regression(model_2002,
                          include = everything(),
                          intercept = TRUE, show_single_row = everything(),
                          label = list(femme ~ "- Proportion de femme" , fertlty ~ "- Fertilité" , Aucun ~ "- Aucun" , Moyen ~ "- Moyen" , Secondr ~ "- Secondaire" , Superir ~ "- Supérieur" , Aucn_cm ~ "- Aucun" , Moyn_cm ~ "- Moyen" , Scndr_c ~ "- Secondaire" , Sprr_cm ~ "- Supérieur" , Ltrcy_A ~ "- Arabic" , Ltrcy_W ~ "- Wolof" , Ltrcy_P ~ "- Pulaar" , Ltrcy_M ~ "- Sereer" , Ltrcy_D ~ "- Mandingo" , Ltrcy_S ~ "- Diola" , rsdnt_b ~ "- Proportion de résident permanent", rsdnt_v ~ "- Proportion de Résident Absent", I_rps_s ~ "- Proportion de repas sauté" , I_sns_m ~ "- Proportion de soins médicaux" , st_ccp_ch ~ "- Proportion d'individus qui sont au chômage ou à la recherche d'un premier emploi" , st_ccp_td ~ "- Proportion d'individus qui sont occupés au foyer" , st_ccp_f ~ "- Proportion d'individus qui sont étudiants/élèves" , st_ccp_rt ~ "- Proportion d'individus qui sont rentiers" ,  st_ccp_tr ~ "- Proportion d'individus qui ont d'autres situations" , nbr__14 ~ "- moins de 14 ans" , nbr_yn_ ~ "- entre 15 et 29 ans" , nbr_dlt ~ "- entre 30 et 44 ans" , nbr_ld_ ~ "- entre 45 et 59 ans" ,  menage ~ "- Proportion de ménage" , mdn_cm_ ~ "- Age médian du CM" , mn_cm_g ~ "- Age moyen du CM" , pct_cm_f ~ "- Proportion de CM de sexe féminin" , mn_nd_p ~ "- Moyenne de l'indice de peoplement" , mdn_nd_ ~ "- Médiane de l'indice de peoplement" , pop_density ~ "- Densité de la population")


                          )%>%
  modify_header(update = list(
  label ~ "**Covariates**",
  estimate ~ '**Estimate**',
  std.error ~ '**Std. Error**',
  statistic ~ '**z value**', 
  p.value = "**Pr(>|z|)**")) %>% 
  add_significance_stars(hide_ci = TRUE, hide_p = FALSE) %>% 
  add_glance_table(include = c(r.squared, adj.r.squared, nobs)) %>%
  modify_column_unhide(columns = c(statistic, std.error)) 

```

```{r}
tbl_2013 <- gtsummary::tbl_regression(model_2013,
                          include = everything(),
                          intercept = TRUE, show_single_row = everything(),
                          label = list(femme ~ "- Proportion de femme" , fertlty ~ "- Fertilité" , Aucun ~ "- Aucun" , Moyen ~ "- Moyen" , Secondr ~ "- Secondaire" , Superir ~ "- Supérieur" , Aucn_cm ~ "- Aucun" , Moyn_cm ~ "- Moyen" , Scndr_c ~ "- Secondaire" , Sprr_cm ~ "- Supérieur" , Ltrcy_A ~ "- Arabic" , Ltrcy_W ~ "- Wolof" , Ltrcy_P ~ "- Pulaar" , Ltrcy_M ~ "- Sereer" , Ltrcy_D ~ "- Mandingo" , Ltrcy_S ~ "- Diola" , rsdnt_b ~ "- Proportion de résident permanent", rsdnt_v ~ "- Proportion de Résident Absent", I_rps_s ~ "- Proportion de repas sauté" , I_sns_m ~ "- Proportion de soins médicaux" , st_ccp_ch ~ "- Proportion d'individus qui sont au chômage ou à la recherche d'un premier emploi" , st_ccp_td ~ "- Proportion d'individus qui sont occupés au foyer" , st_ccp_f ~ "- Proportion d'individus qui sont étudiants/élèves" , st_ccp_rt ~ "- Proportion d'individus qui sont rentiers" ,  st_ccp_tr ~ "- Proportion d'individus qui ont d'autres situations" , nbr__14 ~ "- moins de 14 ans" , nbr_yn_ ~ "- entre 15 et 29 ans" , nbr_dlt ~ "- entre 30 et 44 ans" , nbr_ld_ ~ "- entre 45 et 59 ans" ,  menage ~ "- Proportion de ménage" , mdn_cm_ ~ "- Age médian du CM" , mn_cm_g ~ "- Age moyen du CM" , pct_cm_f ~ "- Proportion de CM de sexe féminin" , mn_nd_p ~ "- Moyenne de l'indice de peoplement" , mdn_nd_ ~ "- Médiane de l'indice de peoplement" , pop_density ~ "- Densité de la population")


                          )%>%
  modify_header(update = list(
  label ~ "**Covariates**",
  estimate ~ '**Estimate**',
  std.error ~ '**Std. Error**',
  statistic ~ '**z value**', 
  p.value = "**Pr(>|z|)**")) %>% 
  add_significance_stars(hide_ci = TRUE, hide_p = FALSE) %>% 
  add_glance_table(include = c(r.squared, adj.r.squared, nobs)) %>%
  modify_column_unhide(columns = c(statistic, std.error))

```



```{r}
tbl_mrg <-
  tbl_merge(
    tbls = list(tbl_2002, tbl_2013),
    tab_spanner = c("**Census 2002**", "**Census 2013**")
  ) %>% 
  as_gt() %>%
  gt::tab_row_group(
    group = "Occupation: (réf. occupés)",
    rows = 22:26) %>%
  gt::tab_row_group(
    group = "Groupe d'age: (réf. 60 ans et plus)",
    rows = 27:30) %>%
  gt::tab_row_group(
    group = "Caractéristiques de la population:",
    rows = 35:37)    %>%
    gt::tab_row_group(
    group = "Caractéristiques du ménage: ",
    rows = c(2, 3, 20, 21, 31,34,32,33)) %>%
  gt::tab_row_group(
    group = "Niveau d'instruction: (ref. Primaire)",
    rows = 4:7) %>%
  gt::tab_row_group(
    group = "Niveau d'instruction du CM: (ref. Primaire)",
    rows = 8:11) %>%
  gt::tab_row_group(
    group = "Langues parlées: (ref. Français)",
    rows = 12:17) %>%
  gt::tab_row_group(
    group = "Migration: (ref. Visiteur)",
    rows = c(18,19))

tbl_mrg
```

```{r}
## Removing the unwanted variables from the model

# remove some variables because of the multicolinearity problem
vars_to_remove_add = c("Aucun","Moyen","Secondr","Superir","mn_cm_g")


label_to_remove_add = c("Niveau d'instruction: Aucun","Niveau d'instruction: Moyen","Niveau d'instruction: Secondaire", "Niveau d'instruction: Supérieur","Age moyen du CM")


predictors_2002 = setdiff(predictors_2002,vars_to_remove_add)
label_predictors_2002 = setdiff(label_predictors_2002, label_to_remove_add)
  
predictors_2013 =  setdiff(predictors_2013,c(vars_to_remove_add,"Aucn_cm"))
label_predictors_2013 = setdiff(label_predictors_2013, c(label_to_remove_add,"Niveau d'instruction du CM: Aucun"))

formula_2002 = paste0("log(",outcome,") ~ ",paste(predictors_2002[-length(predictors_2002)]," + ", collapse = ""),predictors_2002[length(predictors_2002)])

formula_2013 = paste0("log(",outcome,") ~ ",paste(predictors_2013[-length(predictors_2013)]," + ", collapse = ""),predictors_2013[length(predictors_2013)])


# Formula 
# formula_2002 <- gsub(' + hh_size ','',formula_2002, fixed = TRUE)
# formula_2013 <- gsub(' + hh_size ','',formula_2013, fixed = TRUE)


# Data selection

outcome = "MPI_men"


MPI_data_dr_2002_for_model<- MPI_data_dr_2002_for_model %>%
  dplyr::select(MPI_men,predictors_2002) 

MPI_data_dr_2013_for_model<- MPI_data_dr_2013_for_model %>%
  dplyr::select(MPI_men,predictors_2013)



```




```{r}
# Rerun the regression model

### For 2002
model_2002 <- lm(formula = formula_2002, data = MPI_data_dr_2002_for_model)

### For 2013

model_2013 <- lm(formula = formula_2013, data = MPI_data_dr_2013_for_model)



stargazer::stargazer(model_2002, model_2013, type="text",

covariate.labels = label_predictors_2002, out = (paste0(here::here(),"/output/output_model/2_reg_lineaire.txt")))
```


```{r}

##unlist(noquote(paste0(predictors_2002, " ~ " , shQuote(label_predictors_2002), collapse = " , ")))

tbl_2002 <- gtsummary::tbl_regression(model_2002,
                          include = everything(),
                          intercept = TRUE, show_single_row = everything(),
                          label = list(femme ~ "- Proportion de femme" , fertlty ~ "- Fertilité" , Aucn_cm ~ "- Aucun" , Moyn_cm ~ "- Moyen" , Scndr_c ~ "- Secondaire" , Sprr_cm ~ "- Supérieur" , Ltrcy_A ~ "- Arabic" , Ltrcy_W ~ "- Wolof" , Ltrcy_P ~ "- Pulaar" , Ltrcy_M ~ "- Sereer" , Ltrcy_D ~ "- Mandingo" , Ltrcy_S ~ "- Diola" , rsdnt_b ~ "- Proportion de résident permanent", rsdnt_v ~ "- Proportion de Résident Absent", I_rps_s ~ "- Proportion de repas sauté" , I_sns_m ~ "- Proportion de soins médicaux" , st_ccp_ch ~ "- Proportion d'individus qui sont au chômage ou à la recherche d'un premier emploi" , st_ccp_td ~ "- Proportion d'individus qui sont occupés au foyer" , st_ccp_f ~ "- Proportion d'individus qui sont étudiants/élèves" , st_ccp_rt ~ "- Proportion d'individus qui sont rentiers" ,  st_ccp_tr ~ "- Proportion d'individus qui ont d'autres situations" , nbr__14 ~ "- moins de 14 ans" , nbr_yn_ ~ "- entre 15 et 29 ans" , nbr_dlt ~ "- entre 30 et 44 ans" , nbr_ld_ ~ "- entre 45 et 59 ans" ,  menage ~ "- Proportion de ménage" , mdn_cm_ ~ "- Age médian du CM" , pct_cm_f ~ "- Proportion de CM de sexe féminin" , mn_nd_p ~ "- Moyenne de l'indice de peoplement" , mdn_nd_ ~ "- Médiane de l'indice de peoplement" , pop_density ~ "- Densité de la population")


                          )%>%
  modify_header(update = list(
  label ~ "**Covariates**",
  estimate ~ '**Estimate**',
  std.error ~ '**Std. Error**',
  statistic ~ '**z value**', 
  p.value = "**Pr(>|z|)**")) %>% 
  add_significance_stars(hide_ci = TRUE, hide_p = FALSE) %>% 
  add_glance_table(include = c(r.squared, adj.r.squared, nobs)) %>%
  modify_column_unhide(columns = c(statistic, std.error)) 


```

```{r}
tbl_2013 <- gtsummary::tbl_regression(model_2013,
                          include = everything(),
                          intercept = TRUE, show_single_row = everything(),
                          label = list(femme ~ "- Proportion de femme" , fertlty ~ "- Fertilité" , Moyn_cm ~ "- Moyen" , Scndr_c ~ "- Secondaire" , Sprr_cm ~ "- Supérieur" , Ltrcy_A ~ "- Arabic" , Ltrcy_W ~ "- Wolof" , Ltrcy_P ~ "- Pulaar" , Ltrcy_M ~ "- Sereer" , Ltrcy_D ~ "- Mandingo" , Ltrcy_S ~ "- Diola" , rsdnt_b ~ "- Proportion de résident permanent", rsdnt_v ~ "- Proportion de Résident Absent", I_rps_s ~ "- Proportion de repas sauté" , I_sns_m ~ "- Proportion de soins médicaux" , st_ccp_ch ~ "- Proportion d'individus qui sont au chômage ou à la recherche d'un premier emploi" , st_ccp_td ~ "- Proportion d'individus qui sont occupés au foyer" , st_ccp_f ~ "- Proportion d'individus qui sont étudiants/élèves" , st_ccp_rt ~ "- Proportion d'individus qui sont rentiers" ,  st_ccp_tr ~ "- Proportion d'individus qui ont d'autres situations" , nbr__14 ~ "- moins de 14 ans" , nbr_yn_ ~ "- entre 15 et 29 ans" , nbr_dlt ~ "- entre 30 et 44 ans" , nbr_ld_ ~ "- entre 45 et 59 ans" ,  menage ~ "- Proportion de ménage" , mdn_cm_ ~ "- Age médian du CM" , pct_cm_f ~ "- Proportion de CM de sexe féminin" , mn_nd_p ~ "- Moyenne de l'indice de peoplement" , mdn_nd_ ~ "- Médiane de l'indice de peoplement" , pop_density ~ "- Densité de la population")


                          )%>%
  modify_header(update = list(
  label ~ "**Covariates**",
  estimate ~ '**Estimate**',
  std.error ~ '**Std. Error**',
  statistic ~ '**z value**', 
  p.value = "**Pr(>|z|)**")) %>% 
  add_significance_stars(hide_ci = TRUE, hide_p = FALSE) %>% 
  add_glance_table(include = c(r.squared, adj.r.squared, nobs)) %>%
  modify_column_unhide(columns = c(statistic, std.error))

```



```{r}
tbl_mrg <-
  tbl_merge(
    tbls = list(tbl_2002, tbl_2013),
    tab_spanner = c("**Census 2002**", "**Census 2013**")
  ) %>% 
  as_gt() %>%
  gt::tab_row_group(
    group = "Occupation: (réf. occupés)",
    rows = 18:22) %>%
  gt::tab_row_group(
    group = "Groupe d'age: (réf. 60 ans et plus)",
    rows = 23:26) %>%
  gt::tab_row_group(
    group = "Caractéristiques de la population:",
    rows = 28:32)    %>%
    gt::tab_row_group(
    group = "Caractéristiques du ménage: ",
    rows = c(2, 3, 16, 17, 27)) %>%
  gt::tab_row_group(
    group = "Niveau d'instruction du CM: (ref. Primaire)",
    rows = 4:7) %>%
  gt::tab_row_group(
    group = "Langues parlées: (ref. Français)",
    rows = 8:13) %>%
  gt::tab_row_group(
    group = "Migration: (ref. Visiteur)",
    rows = 14:15)

tbl_mrg
```


```{r}


#library(corrr)

### 2002
dfw_estimates_2002 <- MPI_data_dr_2002_for_model%>%
  select(-MPI_men) %>%
  st_drop_geometry()

correlations <- correlate(dfw_estimates_2002, method = "pearson")
network_plot(correlations) + labs(title = "Network plot of correlations between model predictors for Guediawaye Census 2002")

ggsave(paste0(here::here(),"/output/output_model/img/6_corr_var_2002.png"), width = 8, height = 5)

### 2013
dfw_estimates_2013 <- MPI_data_dr_2013_for_model%>%
  select(-MPI_men) %>%
  st_drop_geometry()

correlations <- correlate(dfw_estimates_2013, method = "pearson")

network_plot(correlations) + labs(title = "Network plot of correlations between model predictors for Guediawaye Census 2013")


ggsave(paste0(here::here(),"/output/output_model/img/6_corr_var_2013.png"), width = 8, height = 5)

```

```{r}
library(car)

vif(model_2013)
vif(model_2002)
```

## Dimension reduction with principal components analysis


```{r}
pca_2013 <- prcomp(
  formula = ~., 
  data = dfw_estimates_2013, 
  scale. = TRUE, 
  center = TRUE
)

summary(pca_2013)
##
pca_2002 <- prcomp(
  formula = ~., 
  data = dfw_estimates_2002, 
  scale. = TRUE, 
  center = TRUE
)

summary(pca_2002)
```

```{r}
pca_2013_tibble <- pca_2013$rotation %>%
  as_tibble(rownames = "predictor")
pca_2013_tibble

###
pca_2002_tibble <- pca_2002$rotation %>%
  as_tibble(rownames = "predictor")
pca_2002_tibble
```



```{r}
# For 2002
pca_2002_graph <- pca_2002_tibble %>%
  select(predictor:PC5) %>%
  pivot_longer(PC1:PC5, names_to = "component", values_to = "value") %>%
  ggplot(aes(x = value, y = predictor)) + 
  geom_col(fill = "darkgreen", color = "darkgreen", alpha = 0.5) + 
  facet_wrap(~component, nrow = 1) + 
  labs(y = NULL, x = "Value", title = "Census 2002") + 
  theme_minimal()
```


```{r}
# For 2013
pca_2013_graph <- pca_2013_tibble %>%
                  select(predictor:PC5) %>%
                  pivot_longer(PC1:PC5, names_to = "component", values_to = "value") %>%
                  ggplot(aes(x = value, y = predictor)) + 
                  geom_col(fill = "darkgreen", color = "darkgreen", alpha = 0.5) + 
                  facet_wrap(~component, nrow = 1) + 
                  labs(y = NULL, x = "Value", title = "Census 2013") + 
                  theme_minimal()
```



```{r eval=TRUE, include=TRUE}
gridExtra::grid.arrange(pca_2002_graph,pca_2013_graph, nrow = 1, top="Loadings for first five principal components  for Guediawaye ")

ggsave(paste0(here::here(),"/output/output_model/img/7_hist_pca_rgph.png"), width = 8, height = 5)
```

```{r}
components_2013 <- predict(pca_2013, dfw_estimates_2013)

dfw_pca_2013 <- MPI_data_dr_2013_for_model%>%
  select(MPI_men) %>%
  cbind(components_2013) 

ggplot(dfw_pca_2013, aes(fill = PC1)) +
  geom_sf(color = NA) +
  labs(title = "Map of principal component 1 for Guediawaye Census 2013") +
  theme_void() +
  scale_fill_viridis_c(direction = -1)
```

```{r}
components_2002 <- predict(pca_2002, dfw_estimates_2002)

dfw_pca_2002 <- MPI_data_dr_2002_for_model%>%
  select(MPI_men) %>%
  cbind(components_2002) 

ggplot(dfw_pca_2002, aes(fill = PC1)) +
  geom_sf(color = NA) +
  labs(title = "Map of principal component 1 for Guediawaye Census 2002") +
  theme_void() +
  scale_fill_viridis_c(direction = -1)
```
```{r}
dfw_pca_2002$RGPH<-"Census 2002"
dfw_pca_2013$RGPH<-"Census 2013"
dfw_pca <- dfw_pca_2002 %>%
  plyr::rbind.fill(dfw_pca_2013) %>% 
  st_as_sf()

label_pca <- paste0("Dimension ",1:11)
predictors_pca <- paste0("PC",1:11)

purrr::map(predictors_pca, function(curr_var){ 

curr_lab <- label_pca[match(as.character(sym(curr_var)), predictors_pca)]

curr_plot <- dfw_pca %>% 
      ggplot(aes(fill = !!sym(curr_var)))  + 
      
      # Adding spatial features to the plot
      geom_sf(color = NA) +
       scale_fill_viridis_c(direction = -1)+
      labs(title = curr_lab, fill =  "values") +
      facet_wrap(~RGPH) +
      
# Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )

print(curr_plot)

ggsave(paste0(here::here(),"/output/output_model/img/8_values_",curr_lab,"_rgph.png"), width = 8, height = 5)
}, .progress = TRUE)
```





<!-- ```{r} -->
<!-- dfw_pca %>%  -->

<!--       ggplot(aes(fill = PC2))  +  -->

<!--       # Adding spatial features to the plot -->
<!--       geom_sf(color = NA) + -->
<!--        scale_fill_viridis_c(direction = -1)+ -->
<!--       facet_wrap(~RGPH) + -->

<!--       # Adjusting the plot theme -->
<!--       theme_map(base_size = 8) + -->
<!--       theme(panel.background = element_rect(), -->
<!--                 #legend.background = element_blank(), -->
<!--                 axis.ticks = element_blank(), -->
<!--                 axis.text = element_blank(), -->
<!--                 legend.position = "bottom",  -->
<!--                 text = element_text(size = 8),  -->
<!--                 #panel.grid = element_line(color = "white", size = 0.8), -->
<!--                 legend.box.background = element_rect(), -->
<!--                 legend.box.margin = margin(6, 6, 6, 6), -->
<!--                 plot.background = element_rect(fill = "white"))+ -->
<!--       ggspatial::annotation_scale( -->
<!--         location = "br", -->
<!--         bar_cols = c("grey60", "white"), -->
<!--         text_family = "ArcherPro Book" -->
<!--       ) + -->
<!--       ggspatial::annotation_north_arrow( -->
<!--         location = "tl", which_north = "true", -->
<!--         #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"), -->
<!--         #which_north = "grid", -->
<!--       height = unit(1, "cm"), -->
<!--       width = unit(1, "cm"), -->
<!--       ) -->

<!-- ggsave(paste0(here::here(),"/output/output_model/img/8_pc2_rgph.png"), width = 8, height = 5) -->
<!-- ``` -->

```{r}
# For the model output
# Saving in the "dfw_pca"
sf::st_write(obj=dfw_pca, paste0(here::here(), "/output/output_model/shp/dfw_pca.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) 

```


```{r}

### For 2002
pca_2002_formula <- paste0("log(MPI_men) ~ ", 
                      paste0('PC', 1:10, collapse = ' + '))

pca_2002_model <- lm(formula = pca_2002_formula, data = dfw_pca_2002)


### For 2013
pca_2013_formula <- paste0("log(MPI_men) ~ ", 
                      paste0('PC', 1:10, collapse = ' + '))

pca_2013_model <- lm(formula = pca_2013_formula, data = dfw_pca_2013)



stargazer::stargazer(pca_2002_model, pca_2013_model, type="text", out = (paste0(here::here(),"/output/output_model/2_reg_PCA.html")))
```

## Spatial regression



```{r}
MPI_data_dr_2002_for_model$residuals <- residuals(model_2002)

ggplot(MPI_data_dr_2002_for_model, aes(x = residuals)) + 
  geom_histogram(bins = 100, alpha = 0.5, color = "navy",
                 fill = "navy") + 
  labs(title = "Distribution of model residuals for Guediawaye Census 2002") +
  theme_minimal()
```

```{r}
MPI_data_dr_2013_for_model$residuals <- residuals(model_2013)

ggplot(MPI_data_dr_2013_for_model, aes(x = residuals)) + 
  geom_histogram(bins = 100, alpha = 0.5, color = "navy",
                 fill = "navy") + 
  labs(title = "Distribution of model residuals for Guediawaye Census 2013") +
  theme_minimal()
```


```{r}
MPI_data_dr_for_model <- MPI_data_dr_2002_for_model %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(MPI_data_dr_2013_for_model %>%
  dplyr::mutate(RGPH = "Census 2013"))

ggplot(MPI_data_dr_for_model, aes(x = residuals)) + 
      geom_histogram(bins = 100, alpha = 0.5, color = "navy",
                 fill = "navy") +
       
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
            legend.background = element_blank(),
            axis.ticks = element_blank(),
            axis.text = element_blank(),
            legend.position = "bottom", 
            text = element_text(size = 8), 
            panel.grid = element_line(color = "white", size = 0.8))

ggsave(paste0(here::here(),"/output/output_model/3_reg_lineaire_residuals.png"), width = 8, height = 5)
```


```{r}
library(spdep)

wts_2002 <- MPI_data_dr_2002_for_model %>%
  poly2nb() %>%
  nb2listw()

moran.test(MPI_data_dr_2002_for_model$residuals, wts_2002)
```


```{r}

wts_2013 <- MPI_data_dr_2013_for_model %>%
  poly2nb() %>%
  nb2listw()

moran.test(MPI_data_dr_2013_for_model$residuals, wts_2013)
```



```{r}
MPI_data_dr_2013_for_model$lagged_residuals <- lag.listw(wts_2013, MPI_data_dr_2013_for_model$residuals)

ggplot(MPI_data_dr_2013_for_model, aes(x = residuals, y = lagged_residuals)) + 
  theme_minimal() + 
  labs(title = "Moran scatterplot of residual spatial autocorrelation for Guediawaye Census 2013") +
  geom_point(alpha = 0.5) + 
  geom_smooth(method = "lm", color = "red")
```
```{r}
MPI_data_dr_2002_for_model$lagged_residuals <- lag.listw(wts_2002, MPI_data_dr_2002_for_model$residuals)

ggplot(MPI_data_dr_2002_for_model, aes(x = residuals, y = lagged_residuals)) + 
  theme_minimal() + 
  labs(title = "Moran scatterplot of residual spatial autocorrelation for Guediawaye Census 2002") +
  geom_point(alpha = 0.5) + 
  geom_smooth(method = "lm", color = "red")
```
```{r}
MPI_data_dr_for_model <- MPI_data_dr_2002_for_model %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(MPI_data_dr_2013_for_model %>%
  dplyr::mutate(RGPH = "Census 2013"))

ggplot(MPI_data_dr_for_model, aes(x = residuals, y = lagged_residuals)) + 
  geom_point(alpha = 0.5) + 
  geom_smooth(method = "lm", color = "red")+
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
            legend.background = element_blank(),
            axis.ticks = element_blank(),
            axis.text = element_blank(),
            legend.position = "bottom", 
            text = element_text(size = 8), 
            panel.grid = element_line(color = "white", size = 0.8))

ggsave(paste0(here::here(),"/output/output_model/img/9_residuals_lagresiduals.png"), width = 8, height = 5)
```



# Geographically weighted regression

## Choosing a bandwidth for GWR

```{r}
library(GWmodel)
library(sf)

dfw_data_sp_2002 <- MPI_data_dr_2002_for_model%>%
  as_Spatial()

bw <- bw.gwr(
  formula = formula_2002, 
  data = dfw_data_sp_2002, 
  kernel = "bisquare",
  adaptive = TRUE
)

dfw_data_sp_2013 <- MPI_data_dr_2013_for_model%>%
  as_Spatial()

bw <- bw.gwr(
  formula = formula_2013, 
  data = dfw_data_sp_2013, 
  kernel = "bisquare",
  adaptive = TRUE
)

###

```

```{r}

### 2002
gw_model_2002 <- gwr.basic(
  formula = formula_2002, 
  data = dfw_data_sp_2002, 
  bw = bw,
  kernel = "bisquare",
  adaptive = TRUE
)

### 2013
gw_model_2013 <- gwr.basic(
  formula = formula_2013, 
  data = dfw_data_sp_2013, 
  bw = bw,
  kernel = "bisquare",
  adaptive = TRUE
)


```

```{r}
names(gw_model_2013)

##
names(gw_model_2002)
```

```{r}
gw_model_2013_results <- gw_model_2013$SDF %>%
  st_as_sf() 

names(gw_model_2013_results)

###
gw_model_2002_results <- gw_model_2002$SDF %>%
  st_as_sf() 

names(gw_model_2002_results)
```


```{r}
ggplot(gw_model_2002_results, aes(fill = Local_R2)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local R-squared values from the GWR model for Guediawaye Census 2002",fill="R-squared") +
  theme_void()
```

```{r}
ggplot(gw_model_2013_results, aes(fill = Local_R2)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local R-squared values from the GWR model for Guediawaye Census 2013",fill="R-squared") +
  theme_void()
```


```{r}
gw_model_results <- gw_model_2002_results %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(gw_model_2013_results %>%
  dplyr::mutate(RGPH = "Census 2013")) %>% 
  st_as_sf()


gw_model_results %>% 
    
      ggplot(aes(fill = Local_R2)) + 
      
      # Adding spatial features to the plot
      geom_sf() +
      
      scale_fill_viridis_c(direction = -1) +
       
      facet_wrap(~RGPH) +
      labs(fill="R-squared")+
# Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )

ggsave(paste0(here::here(),"/output/output_model/img/10_rsquared_rgph.png"), width = 8, height = 5)
```


```{r}
# Saving the model output
# Saving in the "gw_model_results"
sf::st_write(obj=gw_model_results, paste0(here::here(), "/output/output_model/shp/gw_model_results.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) 

```


```{r}
ggplot(gw_model_2002_results, aes(fill = menage)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local parameter estimates for household members for Guediawaye Census 2002") +
  theme_void() + 
  labs(fill = "Local β for \nHH")
```

```{r}
ggplot(gw_model_2013_results, aes(fill = menage)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local parameter estimates for household members for Guediawaye Census 2013") +
  theme_void() + 
  labs(fill = "Local β for \nHH")
```


```{r}
gw_model_results <- gw_model_2002_results %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(gw_model_2013_results %>%
  dplyr::mutate(RGPH = "Census 2013")) %>% 
  st_as_sf()


gw_model_results %>% 
    
      ggplot(aes(fill = menage)) + 
      
      # Adding spatial features to the plot
      geom_sf() +
      
      scale_fill_viridis_c(direction = -1) +
  labs(fill = "Local β for \nHH")+
      facet_wrap(~RGPH) +
      
# Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )

ggsave(paste0(here::here(),"/output/output_model/img/11_local_beta_rgph.png"), width = 8, height = 5)
```

```{r}
ggplot(gw_model_2002_results, aes(fill = fertlty)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local parameter estimates for population fertility for Guediawaye Census 2002") +
  theme_void() + 
  labs(fill = "Local β for \n household's fertility")
```

```{r}
ggplot(gw_model_2013_results, aes(fill = fertlty)) + 
  geom_sf(color = NA) + 
  scale_fill_viridis_c(direction = -1) + 
  labs(title = "Local parameter estimates for population fertility for Guediawaye Census 2013") +
  theme_void() + 
  labs(fill = "Local β for \n household's fertility")
```


```{r}
gw_model_results <- gw_model_2002_results %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(gw_model_2013_results %>%
  dplyr::mutate(RGPH = "Census 2013")) %>% 
  st_as_sf()


purrr::map(predictors_2002, function(curr_var){ 

curr_lab <- label_predictors_2002[match(as.character(sym(curr_var)), predictors_2002)]

curr_plot <- gw_model_results %>% 
    
      ggplot(aes(fill = !!sym(curr_var))) + # Use !!sym() to handle variable names dynamically
      
      # Adding spatial features to the plot
      geom_sf() +
      
      scale_fill_viridis_c(direction = -1) +
      #scale_fill_continuous(labels = function(x) format(x, scientific = TRUE))+
  labs(
    title = curr_lab,
    fill =  "Local β") +
      facet_wrap(~RGPH) +
      
# Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )
  
print(curr_plot)

ggsave(paste0(here::here(),"/output/output_model/img/12_beta_",curr_lab,"_rgph.png"), width = 8, height = 5)
}, .progress = TRUE)
```



# Classification and clustering of Guediawaye census data

## Geodemographic classification

```{r}
set.seed(1994)

dfw_kmeans_2013 <- dfw_pca_2013 %>%
  st_drop_geometry() %>%
  select(PC1:PC8) %>%
  kmeans(centers = 6)

table(dfw_kmeans_2013$cluster)

##
dfw_kmeans_2002 <- dfw_pca_2002 %>%
  st_drop_geometry() %>%
  select(PC1:PC8) %>%
  kmeans(centers = 6)

table(dfw_kmeans_2002$cluster)
```



```{r}
dfw_clusters_2002 <- dfw_pca_2002 %>%
  mutate(cluster = as.character(dfw_kmeans_2002$cluster))

ggplot(dfw_clusters_2002, aes(fill = cluster)) + 
  geom_sf(size = 0.1) + 
  #scale_fill_brewer(palette = "Set1") + 
  scale_fill_brewer(palette = "Set3")+
  labs(title = "Map of geodemographic clusters for Guediawaye Census 2002") +
  theme_void() + 
  labs(fill = "Cluster ")
```


```{r}
dfw_clusters_2013 <- dfw_pca_2013 %>%
  mutate(cluster = as.character(dfw_kmeans_2013$cluster))

ggplot(dfw_clusters_2013, aes(fill = cluster)) + 
  geom_sf(size = 0.1) + 
  scale_fill_brewer(palette = "Set3")+
  labs(title = "Map of geodemographic clusters for Guediawaye Census 2013") +
  theme_void() + 
  labs(fill = "Cluster ")
```

```{r}
dfw_clusters <- dfw_clusters_2002 %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(dfw_clusters_2013 %>%
  dplyr::mutate(RGPH = "Census 2013")) %>% 
  st_as_sf()


dfw_clusters %>% 
    
      ggplot(aes(fill = cluster)) + 
      
  geom_sf(size = 0.1) + 
  scale_fill_brewer(palette = "Set3")+ 
      facet_wrap(~RGPH) +
      
# Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
                #legend.background = element_blank(),
                axis.ticks = element_blank(),
                axis.text = element_blank(),
                legend.position = "bottom", 
                text = element_text(size = 8), 
                #panel.grid = element_line(color = "white", size = 0.8),
                legend.box.background = element_rect(),
                legend.box.margin = margin(6, 6, 6, 6),
                plot.background = element_rect(fill = "white"))+
      ggspatial::annotation_scale(
        location = "br",
        bar_cols = c("grey60", "white"),
        text_family = "ArcherPro Book"
      ) +
      ggspatial::annotation_north_arrow(
        location = "tl", which_north = "true",
        #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
        #which_north = "grid",
      height = unit(1, "cm"),
      width = unit(1, "cm"),
      )

ggsave(paste0(here::here(),"/output/output_model/img/13_cluster_rgph.png"), width = 8, height = 5)
```

```{r}
# Saving the model output
# Saving in the "dfw_clusters"
sf::st_write(obj=dfw_clusters, paste0(here::here(), "/output/output_model/shp/dfw_clusters_pca.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) 

```



```{r}
library(plotly)

cluster_plot <- ggplot(dfw_clusters_2013, 
                       aes(x = PC1, y = PC2, color = cluster)) + 
  geom_point() + 
  scale_color_brewer(palette = "Set1") + 
  labs(title = "Interactive scatterplot of PC1 and PC2 colored by cluster for Guediawaye Census 2013") +
  theme_minimal()

ggplotly(cluster_plot) %>%
  layout(legend = list(orientation = "h", y = -0.15, 
                       x = 0.2, title = "Cluster"))
```


```{r}
cluster_plot <- ggplot(dfw_clusters_2002, 
                       aes(x = PC1, y = PC2, color = cluster)) + 
  geom_point() + 
  scale_color_brewer(palette = "Set1") + 
  labs(title = "Interactive scatterplot of PC1 and PC2 colored by cluster for Guediawaye Census 2002") +
  theme_minimal()

ggplotly(cluster_plot) %>%
  layout(legend = list(orientation = "h", y = -0.15, 
                       x = 0.2, title = "Cluster"))
```


## Spatial clustering & regionalization

```{r}
library(spdep)

input_vars <- dfw_pca_2013 %>%
  select(PC1:PC8) %>%
  st_drop_geometry() %>%
  as.data.frame() 

skater_nbrs <- poly2nb(dfw_pca_2013, queen = TRUE)
costs <- nbcosts(skater_nbrs, input_vars)
skater_weights <- nb2listw(skater_nbrs, costs, style = "B")
```

```{r}
mst <- mstree(skater_weights)

regions <- skater(
  mst[,1:2], 
  input_vars, 
  ncuts = 5,
  crit = 10
)
```


```{r}
dfw_clusters_2013$region <- as.character(regions$group)

ggplot(dfw_clusters_2013, aes(fill = region)) + 
  geom_sf(size = 0.1) + 
  labs(title = "Map of contiguous regions derived with the SKATER algorithm for Guediawaye Census 2013") +
  scale_fill_brewer(palette = "Set3")+
  theme_void()
```
```{r}
input_vars <- dfw_pca_2002 %>%
  select(PC1:PC8) %>%
  st_drop_geometry() %>%
  as.data.frame() 

skater_nbrs <- poly2nb(dfw_pca_2002, queen = TRUE)
costs <- nbcosts(skater_nbrs, input_vars)
skater_weights <- nb2listw(skater_nbrs, costs, style = "B")
```

```{r}
mst <- mstree(skater_weights)

regions <- skater(
  mst[,1:2], 
  input_vars, 
  ncuts = 5,
  crit = 10
)
```


```{r}
dfw_clusters_2002$region <- as.character(regions$group)

ggplot(dfw_clusters_2002, aes(fill = region)) + 
  geom_sf(size = 0.1) + 
  labs(title = "Map of contiguous regions derived with the SKATER algorithm for Guediawaye Census 2002") +
  scale_fill_brewer(palette = "Set3")+ 
  theme_void()
```


```{r}
dfw_clusters <- dfw_clusters_2002 %>%
  dplyr::mutate(RGPH = "Census 2002") %>% 
  plyr::rbind.fill(dfw_clusters_2013 %>%
  dplyr::mutate(RGPH = "Census 2013")) %>% 
  st_as_sf()


dfw_clusters %>% 
    
      ggplot(aes(fill = region)) + 
      
  geom_sf(size = 0.1) + 
  scale_fill_brewer(palette = "Set3")+
      facet_wrap(~RGPH) +
      
      # Adjusting the plot theme
      theme_map(base_size = 8) +
      theme(panel.background = element_rect(),
            #legend.background = element_blank(),
            axis.ticks = element_blank(),
            axis.text = element_blank(),
            legend.position = "bottom", 
            text = element_text(size = 8), 
            #panel.grid = element_line(color = "white", size = 0.8),
            legend.box.background = element_rect(),
            legend.box.margin = margin(6, 6, 6, 6),
            plot.background = element_rect(fill = "white"))+
  ggspatial::annotation_scale(
    location = "br",
    bar_cols = c("grey60", "white"),
    text_family = "ArcherPro Book"
  ) +
  ggspatial::annotation_north_arrow(
    location = "tl", which_north = "true",
    #pad_x = unit(0.4, "in"), pad_y = unit(0.4, "in"),
    #which_north = "grid",
  height = unit(1, "cm"),
  width = unit(1, "cm"),
  )

ggsave(paste0(here::here(),"/output/output_model/img/14_dfw_pca_rgph.png"), width = 8, height = 5)
```

```{r}
# Saving the model output
# Saving in the "dfw_clusters"
sf::st_write(obj=dfw_clusters, paste0(here::here(), "/output/output_model/shp/dfw_clusters.shp"),  driver = "ESRI Shapefile", delete_layer = TRUE) 

```
